### 2025-02-20 #7809 @editorialbot
[PRE REVIEW]: Low scaling BSE implementation in the exciting code
**Submitting author:** @maurerben (<a href="http://orcid.org/0000-0001-9152-7390">Benedikt Maurer</a>)
**Repository:** https://github.com/exciting/exciting
**Branch with paper.md** (empty if default branch): fastBSE/paper
**Version:** neon
