+++ src/src_xs/src_rttddft/rttddft_density.f90
@@ -10,44 +10,64 @@
 !> Module that manages what concerns charge density in RT-TDDFT calculations
 module rttddft_Density
   use asserts, only: assert
+  use precision, only: dp, i32
+  use modmpi, only: mpi_env_k
   use rttddft_timings, only: Print_Timings, Timing_RTTDDFT_density, timesec_RTTDDFT
+  use constants, only: zzero, real_zero
+  use mod_convergence, only : iscl
+  use mod_potential_and_density, only: rhomt, rhoir
+  use mod_eigenvalue_occupancy, only: occsv, nstfv
+  use rttddft_Wavefunction, only: wavefunction_set
+  use mod_rhoir, only: genrhoir
+  use mod_rhovalk, only: rhovalk
 
   implicit none
 
   private
 
-  public :: update_density
+  public :: update_density, save_and_frozen, frozen, groundstate
+
+  !> Enum with the density evaluation mode
+  !> There are 4 options: get active density, save density, ground density
+  enum, bind(C)
+    enumerator :: density_case
+    enumerator :: active_and_frozen, save_and_frozen, groundstate, frozen
+  end enum
 
 contains
-  !> In `update_density`, we obtain the charge density at time \(t\). 
+  !> In `update_density`, we obtain the charge density at time \(t\) 
+  !> and put it in global 'rho_mt" and 'rho_ir' arrays.
   !> It is calculated from the WFs, using the same scheme as in the GS
   !> calcultations (refer to `scf_cycle.f90` for the case
   !> `input%groundstate%useDensityMatrix` `.false.`)
-  subroutine update_density( first_kpt, wavefunctions, it, normalize, l_rad_step, printTimings, t_dens )
-    use modmpi, only: mpi_env_k
-    use precision, only: dp, i32
-    use modmain, only : iscl
-    use mod_potential_and_density, only: rhomt, rhoir
-
+  subroutine update_density( first_kpt, psi, it, normalize, l_rad_step, &
+      rhomt_frozen, rhoir_frozen, printTimings, t_dens, dens_case )
     !> The first k point
     integer(i32), intent(in) :: first_kpt
-    !> Wavefunctions in LAPW basis (nmatmax, nstfv, first_kpt : last_kpt)
-    complex(dp), intent(in) :: wavefunctions(:, :, first_kpt :)
+    !> Set of KS wavefunctions
+    class(wavefunction_set), intent(in) :: psi
     !> number of the current iteration (employed to give possible warnings)
-    integer, intent(in)             :: it
+    integer(i32), intent(in) :: it
     !> If `.true.`, normalize the charge density
-    logical, intent(in)             :: normalize
+    logical, intent(in) :: normalize
     !> radial step length
-    integer(i32), intent(in)        :: l_rad_step
+    integer(i32), intent(in) :: l_rad_step
+    !> Frozen part of the muffin-tin density (lmmaxvr, nrmtmax, natmtot)
+    real(dp), optional, intent(in) :: rhomt_frozen(:, :, :)
+    !> Frozen part of the IR density (ngrtot)
+    real(dp), optional, intent(in) :: rhoir_frozen(:)
     !> Object that packs information about printing of timings [[Print_Timings]]
     type(Print_Timings), optional, intent(in) :: printTimings
     !> Object that packs information about timings to update the electronic density
     type(Timing_RTTDDFT_density), optional, intent(out) :: t_dens
+    !> Enum telling which part of the wavefuntion should be used for density evaluation
[…]
