+++ tools/exciting_tools/excitingtools/utils/valid_attributes.py
@@ -650,6 +650,7 @@ realTimeTDDFT_attribute_types = {"TaylorOrder": (int, 1),
                                  "eeInteraction": (str, ["IPA", "aDFT"]),
                                  "endTime": (float, 1),
                                  "normalizeWF": (bool, 1),
+                                 "numberOfFrozenStates": (int, 1),
                                  "printAfterIterations": (int, 1),
                                  "printTimingDetailed": (bool, 1),
                                  "printTimingGeneral": (bool, 1),

[…]
