### 2025-02-23 fc921a3
Updated deviceACC for GPU merge
