+++ external/deviceacc/src/linalg/device/linalg_device_common.f90
@@ -21,26 +21,44 @@ module device_linalg_common_interface
 
     ! The iso_c_binding is included inside the pragmas because its positioning differs in both cases
 #if defined(NVIDIAGPU) || defined(AMDGPU)
-    use omp_lib
-    use iso_c_binding
-    use magma2
+    use iso_c_binding, only: c_ptr, c_int, c_size_t
+    use magma2, only: MagmaLeft, MagmaRight, MagmaLower, MagmaUpper, MagmaNoTrans, &
+                      MagmaTrans, MagmaConjTrans, magma_memset_async, &
+                      magma_get_zgetri_nb, magma_zgemm, magma_get_cgetri_nb, &
+                      magma_zdotu, magma_cdotu, magma_zdotc, magma_cdotc, &
+                      magma_cgerc, magma_caxpy, magmablas_cgeadd2, magma_cgetri_gpu, &
+                      magmablas_zgeadd2, magma_zgemm_batched, magma_zhemv, magma_chemv, &
+                      magma_chemm, magma_ccopyvector_async, magma_zgeru, magma_zcopyvector_async, &
+                      magma_zhemm, magma_cgetrf_gpu, magma_chemv, magma_zgerc, &
+                      magma_zgetrf_gpu, magma_cgeru, magma_cgemm_batched, magma_zgetri_gpu, &
+                      magma_zaxpy, magma_cgemm
 #endif
 #if defined(INTELGPU)
-    use omp_lib
-    use iso_c_binding
-    use onemkl_lapack_omp_offload_lp64
-    use onemkl_blas_omp_offload_lp64
+    use iso_c_binding, only: c_ptr, c_int, c_size_t, c_f_pointer
+    use onemkl_lapack_omp_offload_lp64, only: cgetrf, zgetrf, cgetri, zgetri
+    use onemkl_blas_omp_offload_lp64,   only: cgemm, zgemm, chemv, zhemv, chemm, zhemm, &
+                                              cgemm_batch_strided, zgemm_batch_strided, &
+                                              cdotu, zdotu, cdotc, zdotc, caxpy, zaxpy, &
+                                              cgerc, zgerc, cgeru, zgeru, mkl_comatadd_batch_strided, &
+                                              mkl_zomatadd_batch_strided, ccopy, zcopy
 #endif
     use iso_fortran_env,  only: i32=>int32, r32=>real32, r64=>real64
+    use m_memory_device,  only: bytes_single_complex, bytes_double_complex, &
+                                generate_batched_array, get_device_pointer
+ 
     use m_device_world_t, only: device_world_t
 
     implicit none
 
     private 
-    public :: cgemm_gpu, cdotc_gpu, cdotu_gpu, cgetrf_gpu, cgetri_gpu, get_cgetri_nb_gpu, &
-              zgemm_gpu, zdotc_gpu, zdotu_gpu, zgetrf_gpu, zgetri_gpu, get_zgetri_nb_gpu, &
-              caxpy_gpu, zaxpy_gpu
-              
+    public :: cgemm_gpu,  cdotc_gpu, cdotu_gpu, cgetrf_gpu, cgetri_gpu, get_cgetri_nb_gpu, &
+              zgemm_gpu,  zdotc_gpu, zdotu_gpu, zgetrf_gpu, zgetri_gpu, get_zgetri_nb_gpu, &
+              caxpy_gpu,  zaxpy_gpu, cgemm_batched_gpu, zgemm_batched_gpu, &
+              cgerc_gpu,  cgeru_gpu, zgerc_gpu, zgeru_gpu, &
+              chemv_gpu,  zhemv_gpu, chemm_gpu, zhemm_gpu, &
+              cgeadd_gpu, zgeadd_gpu, &
[…]
