### 2025-03-06 2b4c167
RTTDDFT: EH propagator update
