+++ src/src_xs/src_rttddft/rttddft_init.f90
@@ -15,7 +15,7 @@ module rttddft_init
   use mod_atoms, only: natmtot
   use mod_bands, only: evalfv, nomax, numin, ikcbm, ikvbm, ikvcm
   use mod_core_states, only: ncg
-  use mod_eigensystem, only: nmatmax
+  use mod_eigensystem, only: nmatmax, nmat
   use mod_eigenvalue_occupancy, only: occsv, nstfv, nstsv, efermi
   use mod_gkvector, only: ngk, ngkmax, gkc, tpgkc, sfacgk
   use mod_kpoint, only: vkl, nkpt
@@ -40,6 +40,7 @@ module rttddft_init
   use rttddft_potential, only: update_potential
   use rttddft_VectorPotential, only: Vector_Potential, Vector_Potential_Field
   use rttddft_Wavefunction, only: wavefunction_set
+  use propagators, only: create_propagator, propagator_type => propagator
 
   implicit none
   
@@ -49,12 +50,12 @@ module rttddft_init
 
 contains
 !> This subroutine initializes many global variables in a RT-TDDFT calculation.
-subroutine initialize_rttddft( rt_inp, propagator_needs_extrapolation, vec_pot, a_tot_t_minus_dt, molecular_dynamics, psi, &
-    overlap, ham_init, ham_time, ham_past, apwalm, pmat, pmatmt, rhomt_frozen, rhoir_frozen )
+subroutine initialize_rttddft( rt_inp, propagator, vec_pot, a_tot_t_minus_dt, molecular_dynamics, psi, &
+    overlap, ham_init, ham_time, ham_past, apwalm, pmat, pmatmt, rhomt_frozen, rhoir_frozen, occs_tol )
   !> Argument that encapsulates the input options of rttddft
   type(rttddft_input_keys), intent(in) :: rt_inp
-  !> If `.true.`, propagator needs to extrapolate \(H\)
-  logical, intent(in) :: propagator_needs_extrapolation
+  !> Argument that encapsulates the propagator
+  class(propagator_type), allocatable, intent(out) :: propagator
   !> type that encapsulates the vector potential
   type(Vector_Potential), intent(in) :: vec_pot
   !> \(\mathbf{A}_{tot}\) at time \( t-\Delta t\) 
@@ -81,10 +82,12 @@ subroutine initialize_rttddft( rt_inp, propagator_needs_extrapolation, vec_pot, 
   real(dp), allocatable, intent(out) :: rhomt_frozen(:, :, :)
   !> Frozen part of the IR density (to be allocated in `array_allocation` block)
   real(dp), allocatable, intent(out) :: rhoir_frozen(:)
+  !> Minimal value of occupation for the state to be 'occupied'
+  real(dp), intent(in) :: occs_tol
 
   real(dp), parameter :: epsilon_rgkmax = 1e-14_dp
-  integer(i32) :: ik, first_kpt, last_kpt
-  logical :: evolve_H0, my_rank_writes_to_output
+  integer(i32) :: ik, first_kpt, last_kpt, n_occupied, last_occupied
+  logical :: evolve_H0, my_rank_writes_to_output, success
   real(dp) :: voff(3)
   type(Vector_Potential_Field) :: a_aux
   complex(dp), allocatable :: psi_gnd_lapwlo(:, :, :)
@@ -122,10 +125,11 @@ subroutine initialize_rttddft( rt_inp, propagator_needs_extrapolation, vec_pot, 
   ! One-shot GS calculation
   ! Since an XS calculation with Hybrid functionals uses the GS parameters, a one shot GS calculation serves no purpose
   if (.not. hybrids_used()) call gndstateq( voff, RTDDFT_GND_sufix//filext )
-
[…]
