+++ .gitlab/jupyter-exciting-python3-gnu-smp.yml
@@ -1,17 +1,18 @@
-# Test the Jupytenutorials
+# Test the Jupyter tutorials
 
 include:
   - local: '/.gitlab/python-libs.yml'
 
 test-fedora-buster-gnu-smp-jupyter:
  image: gitdocker.physik.hu-berlin.de/sol/exciting:fedora-GCC14-hdf5-fftw-scalapack-mpich
+
  stage: test
  dependencies:
   - build:fedora-buster-gnu-smp
   - python:excitingtools-python3
  tags:
   - docker
- timeout: 6h  # Increased timeout from default 3h to 6h
+ timeout: 24h  # Increased timeout from default 3h to 24h
  # One might consider moving construction of venv to the build state and make it an artifact
  script:
   - cd tools/excitingjupyter
@@ -24,29 +25,27 @@ test-fedora-buster-gnu-smp-jupyter:
   - cd excitingjupyter
   - time jupyter nbconvert --execute 01_getting_started/01_tutorial_how_to_start_an_exciting_calculation.ipynb --to markdown
   - time python3 01_getting_started/assertions_how_to_start_an_exciting_calculation.py
-  #- time jupyter nbconvert --execute 01_getting_started/02_tutorial_simple_convergence_tests.ipynb --to markdown # commented out because of timeout
-  #- time jupyter nbconvert --execute 01_getting_started/03_tutorial_electronic_band_structure_and_density_of_states.ipynb --to markdown
-  #- time jupyter nbconvert --execute 01_getting_started/04_tutorial_understanding_the_exciting_species_files.ipynb --to markdown # commented out because only markdown
-  #- time jupyter nbconvert --execute 02_ground_state/01_tutorial_exchange_correlation_functionals.ipynb --to markdown # commented out because of timeout
-  #- time jupyter nbconvert --execute 02_ground_state/05_tutorial_hybrid_functional_calculations.ipynb --to markdown # commented out because of timeout
-  #- time jupyter nbconvert --execute 02_ground_state/04_tutorial_spin_orbit_coupling.ipynb --to markdown # commented out because of timeout
-  #- time jupyter nbconvert --execute 02_ground_state/01_Methods/03_tutorial_spin_polarized_calculations_for_bcc_Fe.ipynb --to markdown
-  #- time jupyter nbconvert --execute 02_ground_state/10_tutorial_low_dimensional_systems.ipynb --to markdown # commented out because of timeout
-  #- time jupyter nbconvert --execute 02_ground_state/02_Electronic_Properties/01_tutorial_wannier_functions_for_interpolation_in_reciprocal_space.ipynb --to markdown
-  #- time jupyter nbconvert --execute 02_ground_state/02_Electronic_Properties/02_tutorial_fermi_surface_visualization.ipynb --to markdown
-  #- time jupyter nbconvert --execute 02_ground_state/02_Electronic_Properties/03_tutorial_spin_texture_calculation.ipynb --to markdown
-  #- time jupyter nbconvert --execute 02_ground_state/tutorial_excitingtools.ipynb --to markdown
-  #- time jupyter nbconvert --execute 03_excited_states/01_GW/01_tutorial_electronic_band_structure_from_gw.ipynb --to markdown
-  #- time python3 03_excited_states/01_GW/assertions_01_tutorial_electronic_band_structure_from_gw.py
-  #- time jupyter nbconvert --execute 03_excited_states/02_BSE/01_tutorial_excited_states_from_bse.ipynb --to markdown
-  #- time python3 03_excited_states/02_BSE/assertions_01_tutorial_excited_states_from_BSE.py
[…]
