+++ src/src_xs/src_rttddft/rttddft_main.f90
@@ -18,7 +18,6 @@ module rttddft_main
   use MD_io, only: MD_out
   use mod_atoms, only: natmtot, natoms, nspecies, atposc, idxas
   use mod_charge_and_moment, only: chgval
-  use mod_eigenvalue_occupancy, only: occsv
   use mod_eigensystem, only: nmat
   use mod_misc, only: filext
   use mod_kpoint, only: nkpt, wkpt, kpt_latt => vkl
@@ -64,7 +63,7 @@ module rttddft_main
   
   integer(i32), parameter :: i_spin = 1
 
-  public  :: coordinate_rttddft_calculation
+  public :: coordinate_rttddft_calculation
 
 contains
 
@@ -105,9 +104,11 @@ contains
     real(dp), allocatable :: rhomt_frozen(:, :, :)
     ! Frozen part of the IR density (ngrtot)
     real(dp), allocatable :: rhoir_frozen(:)
+    ! Initial occupations array (nstates, first_kpt : last_kpt)
+    real(dp), allocatable :: occupations(:, :)
 
     integer(i32) :: it, first_kpt, last_kpt, first_step, last_step
-    integer(i32) :: i_print, timeStepMultiplier, first_active
+    integer(i32) :: i_print, timeStepMultiplier
     logical :: pred_corr_reached_max_steps, my_rank_writes_to_output, &
       density_needed, evolve_H0, take_screenshot
     character(len=:), allocatable :: string
@@ -165,7 +166,7 @@ contains
     my_rank_writes_to_output = ( rank == 0 ) 
     if( my_rank_writes_to_output ) then
       call open_rttddft_outputs( rt )
-      call write_file_info_header
+      call write_file_info_header()
     end if
 
     ! Initialization
@@ -181,17 +182,17 @@ contains
     
     call initialize_rttddft( rt, propagator, vec_pot, a_tot_save, &
         molecular_dynamics, psi, overlap, ham_init, ham_time, ham_past, apwalm, &
-        pmat, pmatmt, rhomt_frozen, rhoir_frozen, eps_occ )
+        pmat, pmatmt, rhomt_frozen, rhoir_frozen, occupations, eps_occ )
     call distribute_loop(mpi_env_k, nkpt, first_kpt, last_kpt)
     if( molecular_dynamics%on ) call init_MD( time, vec_pot%a_tot, dt, &
-        psi%active, overlap, ham_time, timeStepMultiplier, molecular_dynamics, &
+        psi%active, occupations, overlap, ham_time, timeStepMultiplier, molecular_dynamics, &
         MD_outputs, atom_positions, atom_velocities, e_field, forces )
     if ( rt%subtract_J0 ) then
-      call j_ind%evaluate_paramagnetic( psi, pmat, occsv(:, first_kpt:last_kpt), wkpt(first_kpt:last_kpt), mpi_env_k )
+      call j_ind%evaluate_paramagnetic( psi, pmat, occupations, wkpt(first_kpt:last_kpt), mpi_env_k )
       j_para_spurious = j_ind%paramagnetic
     end if
     if( rt%restart_previous_calculation() ) then
-      call j_ind%evaluate_paramagnetic( psi, pmat, occsv(:, first_kpt:last_kpt), wkpt(first_kpt:last_kpt), mpi_env_k )
+      call j_ind%evaluate_paramagnetic( psi, pmat, occupations, wkpt(first_kpt:last_kpt), mpi_env_k )
       call j_ind%evaluate_diamagnetic( chgval/Omega, vec_pot%a_tot )
     end if
 
@@ -216,29 +217,30 @@ contains
 
     ! Total energy
     if ( rt%calculate_total_energy .and. rt%do_from_scratch() ) then
[…]
