+++ src/src_xs/src_rttddft/rttddft_MD.f90
@@ -15,7 +15,6 @@ module rttddft_MD
     spcore, spmass, spocc, spr, spzn
   use mod_corestate, only: rhocr
   use mod_eigensystem, only: nmat, nmatmax
-  use mod_eigenvalue_occupancy, only: occsv
   use mod_gkvector, only: ngk, ngkmax, gkc, tpgkc, sfacgk, vgkc
   use mod_gvector, only: ngvec, vgc, sfacg
   use mod_kpoint, only: nkpt, wkpt
@@ -67,8 +66,8 @@ contains
     forall( is = 1:n_species ) charge_val(is) = sum( spocc(:, is), mask=(.not.spcore(:, is)) )
   end subroutine
 
-  subroutine force_rttdft( forces, a_tot, e_field, MD_input, evecfv_time, overlap, &
-    ham_time, printTimings, t_MD )
+  subroutine force_rttdft( forces, a_tot, e_field, MD_input, evecfv_time, occupations, &
+      overlap, ham_time, printTimings, t_MD )
     !> Object that packs information about the total forces
     type(force), intent(inout)      :: forces
     !> `x`, `y`, and `z` components of the (total) vector potential
@@ -79,6 +78,8 @@ contains
     type(MD_input_keys), intent(in) :: MD_input
     !> Basis-expansion coefficients of the KS-WFs at time \(t\)
     complex(dp), intent(in) :: evecfv_time(:, :, :)
+    !> State occupations array
+    real(dp), intent(in) :: occupations(:, :)
     !> Overlap matrix (of basis functions)
     complex(dp), intent(in) :: overlap(:, :, :)
     !> Hamiltonian matrix at current time \(t\)
@@ -127,8 +128,8 @@ contains
 
     ! Valence corrections: second part
     if( MD_input%valence_corrections ) &
-      call obtain_valence_corrections_part2( first_kpt, last_kpt, mpi_env_k, &
-        evecfv_time, overlap, ham_time, forces%val )
+      call obtain_valence_corrections_part2( first_kpt, mpi_env_k, &
+        evecfv_time, occupations, overlap, ham_time, forces%val )
     if( tDetail ) call timesec_RTTDDFT( ti, t_MD%t_MD_2nd )
     ! sum all contributions to total force and store it
     call forces%evaluate_total_force()
@@ -138,17 +139,17 @@ contains
   end subroutine
 
   !> Wrapper for calling val_corr_pt2_given_atom_and_kpt
-  subroutine obtain_valence_corrections_part2( first_kpt, last_kpt, mpi_env, &
-    evecfv_time, overlap, ham_time, forces_val )
+  subroutine obtain_valence_corrections_part2( first_kpt, mpi_env, &
+    evecfv_time, occupations, overlap, ham_time, forces_val )
     !> index of the first `k-point` to be considered in the sum
-    integer(i32),intent(in)        :: first_kpt
-    !> index of the last `k-point` considered
-    integer(i32),intent(in)        :: last_kpt
+    integer(i32),intent(in) :: first_kpt
     !> MPI environment
-    type(mpiinfo), intent(in)      :: mpi_env
+    type(mpiinfo), intent(in) :: mpi_env
     !> Basis-expansion coefficients of the KS-WFs at time \(t\)
-    !> (nmatmax, nstfv, first_kpt : last_kpt)
[…]
