+++ tools/excitingscripts/tests/test_free_from_dos.py
@@ -1,0 +1,34 @@
+import numpy as np
+from excitingscripts.free_from_dos import thermodynamic_properties
+
+
+def test_thermodynamic_properties():
+    omega_data = np.array([
+        14.32022378, 15.89544559, 17.4706674, 19.0458892, 20.62111101,
+        22.19633282, 23.77155461, 25.34677643, 26.92199823, 28.49722004,
+        30.07244184, 31.64766366, 33.22288545, 34.79810727, 36.37332907,
+        37.94855086, 39.52377268, 41.09899448, 42.6742163, 44.24943809
+    ])
+
+    dos_data = np.array([
+        1.80788643e-06, 2.10015495e-06, 2.49020227e-06, 3.04185444e-06,
+        3.82780724e-06, 4.70135038e-06, 5.56445164e-06, 6.34934710e-06,
+        7.11860044e-06, 7.91464869e-06, 8.74473099e-06, 9.60595723e-06,
+        1.04919210e-05, 1.14506355e-05, 1.25462421e-05, 1.38484668e-05,
+        1.52164186e-05, 1.65496607e-05, 1.77359841e-05, 1.89670755e-05
+    ])
+    temp = 100
+    energy_unit = "eV"
+
+    ref_results = np.array([
+        -1.6410342786222294e-06,
+        2.2309568192980417e-06,
+        3.871991097920271e-06,
+        3.871991097920272e-05,
+        0.00024871194893540474,
+        5.262182925446062e-07
+    ])
+
+    results = np.array(thermodynamic_properties(temp, omega_data, dos_data, energy_unit))
+
+    assert np.allclose(results, ref_results)
