+++ src/src_xs/src_rttddft/rttddft_density.f90
@@ -19,56 +19,60 @@ module rttddft_Density
   use rttddft_Wavefunction, only: wavefunction_set
   use mod_rhoir, only: genrhoir
   use mod_rhovalk, only: rhovalk
+  use general_matrix_multiplication, only: matrix_multiply
+  use mod_eigensystem, only: nmat
 
   implicit none
 
   private
 
-  public :: update_density, save_and_frozen, frozen, groundstate
+  public :: update_density, save_and_frozen, groundstate, frozen
 
   !> Enum with the density evaluation mode
-  !> There are 4 options: get active density, save density, ground density
+  !> There are 4 options available: get density from `active` and `frozen` states, 
+  !> `save` and `frozen` states, just `frozen` states, or all the unperturbed states.
   enum, bind(C)
     enumerator :: density_case
-    enumerator :: active_and_frozen, save_and_frozen, groundstate, frozen
+    enumerator :: active_and_frozen, save_and_frozen, frozen, groundstate
   end enum
 
 contains
   !> In `update_density`, we obtain the charge density at time \(t\) 
-  !> and put it in global 'rho_mt' and 'rho_ir' arrays.
-  !> It is calculated from the WFs, using the same scheme as in the GS
-  !> calcultations (refer to `scf_cycle.f90` for the case
-  !> `input%groundstate%useDensityMatrix` `.false.`)
+  !> and put it in global `rho_mt` and `rho_ir` arrays. This routine 
+  !> manages calls of `get_density_from_lapwlo_set` with appropriate arguments.
   subroutine update_density( first_kpt, psi, occupations, it, normalize, l_rad_step, &
-      rhomt_frozen, rhoir_frozen, printTimings, t_dens, dens_case )
+      rhomt_frozen, rhoir_frozen, ks_lapwo_transition_matrix, printTimings, t_dens, dens_case )
     !> The first k point
     integer(i32), intent(in) :: first_kpt
     !> Set of KS wavefunctions
     class(wavefunction_set), intent(in) :: psi
-    !> Initial occupations array
+    !> Initial occupations array (n_states, n_kpts)
     real(dp), intent(in) :: occupations(:, :)
-    !> number of the current iteration (employed to give possible warnings)
+    !> Number of the current iteration (employed to give possible warnings)
     integer(i32), intent(in) :: it
     !> If `.true.`, normalize the charge density
     logical, intent(in) :: normalize
-    !> radial step length
+    !> Radial step length
     integer(i32), intent(in) :: l_rad_step
     !> Frozen part of the muffin-tin density (lmmaxvr, nrmtmax, natmtot)
     real(dp), optional, intent(in) :: rhomt_frozen(:, :, :)
     !> Frozen part of the IR density (ngrtot)
     real(dp), optional, intent(in) :: rhoir_frozen(:)
-    !> Object that packs information about printing of timings [[Print_Timings]]
+    !> KS-LAPW+lo transition matrix (nmatmax, nstfv, first_kpt : last_kpt)
+    complex(dp), optional, intent(in):: ks_lapwo_transition_matrix(:, :, :)
+    !> Object that packs information about printing of timings
     type(Print_Timings), optional, intent(in) :: printTimings
     !> Object that packs information about timings to update the electronic density
     type(Timing_RTTDDFT_density), optional, intent(out) :: t_dens
-    !> Enum telling which part of the wavefuntion should be used for density evaluation
[…]
