### 2025-03-23 5335367
Fix excitingjupyter Notebook Version
Sets `notebook` version to `6.5.7` for the `excitingjupyter` package. Updates setting the custom CSS accordingly. Also updates the README files to include setting environment variables.
