### 2025-03-24 04f8d11
Update XC functionals tutorial
