+++ tools/exciting_tools/excitingtools/__init__.py
@@ -4,11 +4,10 @@ from excitingtools.constants.units import Unit
 # User-level objects
 from excitingtools.dataclasses import BandData, EigenValues
 
-# old deprecated API
 # Parsers returning to dicts
 # Questionable whether one should expose this - required for test framework recursive comparisons
 # Typically not the API one wants to expose to the user, as parsed dict keys are subject to change
-from excitingtools.exciting_dict_parsers import parse, parser_chooser
+from excitingtools.exciting_dict_parsers import parse
 
 # Parsers returning to objects
 from excitingtools.exciting_obj_parsers import parse_band_structure
@@ -39,7 +38,6 @@ __all__ = [
     "BandData",
     "EigenValues",
     "parse",
-    "parser_chooser",
     "parse_band_structure",
     "ExcitingGroundStateInput",
     "ExcitingXSInput",
