### 2025-03-25 669a9ad
General Lattice Optimization Jupyter Tutorial
