+++ tools/excitingscripts/excitingscripts/optimize/analyze.py
@@ -1,0 +1,580 @@
+"""Python script for fitting the energy-vs-volume and energy-vs-strain curves."""
+
+import warnings
+from argparse import ArgumentParser
+from pathlib import Path
+from typing import List, Callable
+
+import matplotlib.pyplot as plt
+import matplotlib.style
+import matplotlib.ticker as ptk
+import numpy as np
+from excitingscripts.optimize.setup import (
+    get_deformation_matrix,
+    optimization_map,
+)
+from excitingscripts.utils.utils import (
+    birch_murnaghan_eos,
+    pressure_birch_murnaghan_eos,
+    initial_guess,
+)
+from excitingscripts.utils.utils import sort_lists_by_first_list
+from excitingtools import ExcitingInputXML
+from excitingtools import parse
+from scipy.constants import physical_constants
+from scipy.optimize import leastsq
+
+if matplotlib.__version__.split(".")[0] == "2":
+    matplotlib.style.use("classic")
+
+bohr_radius = physical_constants["Bohr radius"][0]
+joule2hartree = physical_constants["hartree-joule relationship"][0]
+unitconv = joule2hartree / (bohr_radius ** 3) * 10 ** -9
+
+
+def get_energy_data(run_dir: Path, dir_name: str, vol_flag: bool) -> None:
+    """Extracts and saves the energy-volume data from a series of exciting calculations.
+
+    :param run_dir: Path to the directory containing the calculations.
+    :param dir_name: Name of the directory indicating the type of optimization.
+    :param vol_flag: Flag to indicate if the calculation is a volume optimization.
+    """
+    displ_points = sum(
+        1
+        for directory in run_dir.iterdir()
+        if directory.is_dir() and directory.name.startswith(f"{dir_name}_")
+    )
+
+    total_energy = []
+    x_values = []
+
+    if vol_flag:
+        outfile = run_dir / "energy-vs-volume"
+    else:
+        outfile = run_dir / "energy-vs-strain"
+
+    for i in range(displ_points):
+        root_directory = run_dir / f"{dir_name}_{i + 1}"
+        infofile = root_directory / "INFO.OUT"
+        parsed_info = parse(infofile.as_posix())
+        max_scf = max([int(i) for i in parsed_info["scl"].keys()])
+        converged_results = parsed_info["scl"][str(max_scf)]
+        total_energy.append(converged_results["Total energy"])
+
+        if vol_flag:
+            x_values.append(parsed_info["initialization"]["Unit cell volume"])
+
+    if not vol_flag:
+        strain_infofile = run_dir / f"{dir_name}-Parameters"
+        lines = strain_infofile.read_text().splitlines()
+        for line in lines:
+            if "Physical strain" in line:
+                x_values.append(float(line.split()[-1]))
+
+    energy_data = np.vstack((x_values, total_energy)).T
+
+    with open(outfile, "w") as f:
+        np.savetxt(f, energy_data)
+
+
+def murnaghan_eos(v: float, p: List[float]) -> float:
+    """Calculate energy using the Murnaghan equation of state.
+
+    :param v: Volume.
+    :param p: Parameters for the equation [v0, e0, b0, bp].
+    :return: Calculated energy.
+    """
+    eq_vol = p[0]  # Equilibrium volume
+    min_energy = p[1]  # Minimum energy
+    bulk_modulus = p[2]  # Bulk modulus at equilibrium volume
+    bulk_modulus_pressure_deriv = p[3]  # Pressure derivative of bulk modulus
+
+    energy = (
+            min_energy
+            + bulk_modulus
+            * v
+            / bulk_modulus_pressure_deriv
[…]
