+++ tools/excitingscripts/excitingscripts/plot/energy.py
@@ -150,7 +150,7 @@ def main() -> None:
 
     yfmt = ptk.ScalarFormatter(useOffset=True, useMathText=True)
 
-    fig = matplotlib.pyplot.figure(1, figsize=(8, 5.5))
+    fig = matplotlib.pyplot.figure(figsize=(8, 5.5))
 
     ax = fig.add_subplot(111)
 
