+++ tools/excitingscripts/excitingscripts/plot/centerofmass.py
@@ -133,3 +133,4 @@ def main() -> None:
 
 if __name__ == "__main__":
     main()
+

[…]
