### 2025-03-25 d7248b0
Update GW and BSE tutorials
