### 2025-03-25 2a38cc9
Fix assertions
