+++ tools/excitingjupyter/excitingjupyter/03_excited_states/02_BSE/assertions/excited_states_from_BSE.py
@@ -3,7 +3,7 @@ from os.path import dirname
 import numpy as np
 from excitingtools import parse
 
-TUTORIAL_EXCITED_STATES_FROM_BSE_RUNDIR = "run_LiF_BSE/BSE/EPSILON"
+TUTORIAL_EXCITED_STATES_FROM_BSE_RUNDIR = "../run_LiF_BSE/BSE/EPSILON"
 REFERENCE_DIR = "reference_excited_states_from_BSE/EPSILON"
 
 def test_tutorial_excited_states_from_bse(file_name: str):
@@ -11,7 +11,7 @@ def test_tutorial_excited_states_from_bse(file_name: str):
     """
     
     epsilon_reference = parse(f"{dirname(__file__)}/{REFERENCE_DIR}/{file_name}")
-    epsilon_result = parse(f"{dirname(__file__)}/../{TUTORIAL_EXCITED_STATES_FROM_BSE_RUNDIR}/{file_name}")
+    epsilon_result = parse(f"{dirname(__file__)}/{TUTORIAL_EXCITED_STATES_FROM_BSE_RUNDIR}/{file_name}")
 
     assert np.allclose(epsilon_result['frequency'],
                        epsilon_reference['frequency'], atol=2e-6), (
