+++ tools/excitingjupyter/static_web_config.py
@@ -11,9 +11,8 @@ os.makedirs(OUTDIR, exist_ok=True)
 c = get_config()  # type:ignore # pylint: disable=E0602
 
 c.NbConvertApp.notebooks = [
-    # Convert all tutorials in 01_getting_started and 02_ground_state (may be slow if combined with the --execute flag)
-    os.path.join(NB_ROOT, "01_getting_started", "*_tutorial_*.ipynb"),
-    os.path.join(NB_ROOT, "02_ground_state", "*_tutorial_*.ipynb")
+    # Convert all tutorials in 01_getting_started (may be slow if combined with the --execute flag)
+    os.path.join(NB_ROOT, "01_getting_started", "*.ipynb"),
     # for fast testing use, e.g., (because it has images etc.):
     # os.path.join(NB_ROOT, "01_getting_started", "electronic_band_structure_and_density_of_states.ipynb")
 ]
