### 2025-03-26 ec8afb5
Multiple BSE types in 1 Calculation
