+++ tools/exciting_tools/excitingtools/exciting_dict_parsers/input_parser.py
@@ -141,6 +141,7 @@ special_tags_to_parse_map = {
     "structure": parse_structure,
     "qpointset": lambda root: [[float(x) for x in qpoint.text.split()] for qpoint in root],
     "plan": lambda root: [doonly.attrib["task"] for doonly in root],
+    "BseTypeSet": lambda root: [bsetype.attrib["name"] for bsetype in root],
     "kstlist": lambda root: [[int(x) for x in pointstatepair.text.split()] for pointstatepair in root],
     "etCoeffComponents": lambda root: [int(x) for x in root.text.split()],
 }
