+++ src/xstring/xstring.f90
@@ -1,5 +1,6 @@
 module xstring
   use string_utils
   use to_char_conversion
+  use string_parser
   use xstring_format
 end module  
