### 2025-03-27 f134f01
Assertions for getting started tutorials
