### 2025-03-27 cb60b1b
Add assertions for groundstate methods tutorials
