### 2025-03-27 4c1483f
More assertions
