### 2025-03-27 a7a935c
Fix bug in molecules tutorial
