### 2025-03-27 b9f6244
Electronic properties assertions
