### 2025-03-27 213b4b7
Replace ZHEGVX by its wrapper in matrix_exp
