+++ src/math/matrix_exp.f90
@@ -1,11 +1,12 @@
-!> Module for advanced matrix operations
+!> Module computing the matrix exponential
 module matrix_exp
   use asserts, only: assert
   use constants, only: zone, zzero
   use math_utils, only: is_hermitian, is_positive_definite
   use general_matrix_multiplication, only: matrix_multiply
   use hermitian_matrix_multiplication, only: hermitian_matrix_multiply
-  use precision, only: dp
+  use precision, only: dp, i32
+  use xlapack, only: solve_generalized_hermitian_eigenproblem
 
   implicit none
 
@@ -18,55 +19,48 @@ module matrix_exp
   real(dp), parameter :: tol_default = 1e-6_dp
 
 contains
-  !> This subroutine obtains the exponential \( \exp(\alpha \hat{H}) \) applied
-  !> to a set of vectors: \( \exp(\alpha \hat{H})| v_{j} \rangle \).
-  !> The operator \( \hat{H} \) **must** be hermitian. \( \alpha \)
-  !> is a complex prefactor. The vectors \( | v_{j} \rangle \)
-  !> are described through the expansion coefficients \( C_{j\mu} \)
-  !> in terms of the (L)APW+lo basis.
+  !> Obtain the action of the exponential \( \exp(\alpha \hat{H}) \) operator on
+  !> a set of vectors, yielding \( \exp(\alpha \hat{H})| v_{j} \rangle \), 
+  !> for \(j=1,\ldots, N\). Here, \( \hat{H} \) is a hermitian operator, and \( \alpha \)
+  !> is a complex pre-factor. The vectors \( | v_{j} \rangle \)
+  !> are represented in terms of expansion coefficients \( C_{j\mu} \)
+  !> with respect to a non-orthonormal basis:
   !> \[
   !>    | v_{j} \rangle = \sum_\mu
   !>    C_{j\mu} | \phi_{\mu} \rangle
   !> \]
-  !> Since the basis is not orthonormal, we have
+  !> Given the overlap matrix \( S \), the exponential operation is expressed as: 
   !> \[
-  !>    \exp [ \alpha \hat{H} ]
-  !>    | v_{j} \rangle =
-  !>    \exp [ \alpha S^{-1}H ] \;
-  !>    C_{j} = \sum_{n=0}^{M} \frac{1}{n!}
-  !>    (\alpha S^{-1}H)^n C_{j}
+  !>    \exp [ \alpha \hat{H} ] | v_{j} \rangle =
+  !>    \exp [ \alpha S^{-1}H ] \; C_{j} = \sum_{n=0}^{M} \frac{1}{n!}
+  !>    (\alpha S^{-1}H)^n C_{j}, \quad j=1,\ldots, N
   !> \]
-  !> The exponential here is approximated by a Taylor expansion
-  !> up to the order defined by \( M \) (`order_taylor`)
+  !> Here, the matrix exponential is approximated using a Taylor expansion
+  !> up to the order defined by \( M \) ([[order_taylor]])
   subroutine exp_hermitian_matrix_times_vectors( order_taylor, alpha, &
     & H, S, vectors, tol )
-    !> The order of the Taylor expansion
-    integer, intent(in)           :: order_taylor
-    !> Complex prefactor
-    complex(dp), intent(in)       :: alpha
+    !> The order \( M \) of the Taylor expansion
+    integer(i32), intent(in)       :: order_taylor
+    !> Complex prefactor \( \alpha \)
+    complex(dp), intent(in)        :: alpha
     !> Hermitian matrix \( H \)
-    complex(dp),intent(in)        :: H(:, :)
[…]
