### 2025-03-28 ac8e5b6
More assertions
