### 2025-03-30 f208bf1
Another try for assertions
