+++ .gitlab/jupyter-exciting-python3-gnu-smp.yml
@@ -26,18 +26,18 @@ test-fedora-buster-gnu-smp-jupyter:
   - time jupyter nbconvert --execute 01_getting_started/how_to_start_an_exciting_calculation.ipynb --to markdown
   - time python3 01_getting_started/assertions/how_to_start_an_exciting_calculation.py
   - time jupyter nbconvert --execute 01_getting_started/simple_convergence_tests.ipynb --to markdown
-  - time python3 01_getting_started/assertions/simple_convergence_tests.py
-  - time jupyter nbconvert --execute 01_getting_started/electronic_band_structure_and_density_of_states.ipynb --to markdown
-  - time python3 01_getting_started/assertions/electronic_band_structure_and_density_of_states.py
-  - time jupyter nbconvert --execute 01_getting_started/understanding_the_exciting_species_files.ipynb --to markdown
-  - time jupyter nbconvert --execute 02_ground_state/01_Methods/exchange_correlation_functionals.ipynb --to markdown
-  - time python3 02_ground_state/01_Methods/assertions/exchange_correlation_functionals.py
-  - time jupyter nbconvert --execute 02_ground_state/01_Methods/how_to_visualize_kohn_sham_states.ipynb --to markdown
-  - time python3 02_ground_state/01_Methods/assertions/how_to_visualize_kohn_sham_states.py
-  - time jupyter nbconvert --execute 02_ground_state/01_Methods/hybrid_functional_calculations.ipynb --to markdown
-  - time python3 02_ground_state/01_Methods/assertions/hybrid_functional_calculations.py
-  - time jupyter nbconvert --execute 02_ground_state/01_Methods/spin_orbit_coupling.ipynb --to markdown
-  - time python3 02_ground_state/01_Methods/assertions/spin_orbit_coupling.py
+  - #time python3 01_getting_started/assertions/simple_convergence_tests.py
+  - #time jupyter nbconvert --execute 01_getting_started/electronic_band_structure_and_density_of_states.ipynb --to markdown
+  - #time python3 01_getting_started/assertions/electronic_band_structure_and_density_of_states.py
+  - #time jupyter nbconvert --execute 01_getting_started/understanding_the_exciting_species_files.ipynb --to markdown
+  - #time jupyter nbconvert --execute 02_ground_state/01_Methods/exchange_correlation_functionals.ipynb --to markdown
+  - #time python3 02_ground_state/01_Methods/assertions/exchange_correlation_functionals.py
+  - #time jupyter nbconvert --execute 02_ground_state/01_Methods/how_to_visualize_kohn_sham_states.ipynb --to markdown
+  - #time python3 02_ground_state/01_Methods/assertions/how_to_visualize_kohn_sham_states.py
+  - #time jupyter nbconvert --execute 02_ground_state/01_Methods/hybrid_functional_calculations.ipynb --to markdown
+  - #time python3 02_ground_state/01_Methods/assertions/hybrid_functional_calculations.py
+  - #time jupyter nbconvert --execute 02_ground_state/01_Methods/spin_orbit_coupling.ipynb --to markdown
+  - #time python3 02_ground_state/01_Methods/assertions/spin_orbit_coupling.py
   - time jupyter nbconvert --execute 02_ground_state/01_Methods/spin_polarized_calculations_for_bcc_Fe.ipynb --to markdown
   - time python3 02_ground_state/01_Methods/assertions/spin_polarized_calculations_for_bcc_Fe.py
   - time jupyter nbconvert --execute 02_ground_state/01_Methods/dft_0.5.ipynb --to markdown
