+++ tools/excitingjupyter/excitingjupyter/02_ground_state/01_Methods/assertions/dft_0.5.py
@@ -21,10 +21,10 @@ def test_dft_half():
                     [3.9, 1.1841189830],
                     [4.0, 1.1691102708],
                     [4.5, 1.0394795805],
-                    [5.0, 0.8522698691]]), (
-                    "bandgap values not equivalent to reference calculation")
+                    [5.0, 0.8522698691]])
 
-    assert np.allclose(bandgap_results, bandgap_reference,  atol=1.0e-7)
+    assert np.allclose(bandgap_results, bandgap_reference,  atol=1.0e-7), (
+                    "bandgap values not equivalent to reference calculation")
 
 
 def main():
