### 2025-03-31 e3cfdd2
Changed the CMake version to 3.5 in FoX and libXC.
