### 2025-03-31 97366f8
Another assertions test
