+++ tools/excitingjupyter/excitingjupyter/02_ground_state/05_Lattice_Dynamics/assertions/lattice_dynamics_of_diamond_and_zincblende_structure_crystals.py
@@ -1,20 +1,19 @@
 from os.path import dirname
 
-from excitingtools.exciting_dict_parsers.phonon_parser import parse_phonon_out
+import numpy as np
 
 TUTORIAL_RUNDIR = "../run_diamond_phonons"
 REFERENCE_DIR = "reference_lattice_dynamics_of_diamond_and_zincblende_structure_crystals"
 
-def test_phonon_data():
-    phonon_reference = parse_phonon_out(f"{dirname(__file__)}/{REFERENCE_DIR}/PHONON.OUT")
-    phonon_results = parse_phonon_out( f"{dirname(__file__)}/{TUTORIAL_RUNDIR}/PHONON.OUT")
+def test_phonon_dos(file_name: str):
+    dos_reference = np.loadtxt(f"{dirname(__file__)}/{REFERENCE_DIR}/{file_name}")
+    dos_results = np.loadtxt(f"{dirname(__file__)}/{TUTORIAL_RUNDIR}/{file_name}")
 
-    assert phonon_results == phonon_reference,(
-                       "PHONON.OUT data not equivalent to reference calculation")
-
+    assert np.allclose(dos_results, dos_reference, atol=1.0e-8),(
+                       "Phonon DOS not equivalent to reference calculation")
 
 def main():
-    test_phonon_data()
+    test_phonon_dos("PHDOS.OUT")
 
 if __name__=="__main__":
     main()
