+++ tools/excitingjupyter/set_env_vars.sh
@@ -2,5 +2,4 @@
 cd ../../
 export EXCITINGROOT=$(pwd)
 export TIMEFORMAT="   Elapsed time = %0lR"
-export WRITEMINMAX="1"
 cd tools/excitingjupyter
