+++ tools/excitingjupyter/excitingjupyter/03_excited_states/02_BSE/assertions/xray_absorption_spectra_using_bse.py
@@ -1,0 +1,242 @@
+from os.path import dirname
+
+import numpy as np
+from excitingtools import parse
+
+TUTORIAL_XANES_RUNDIR = "../run_XANES"
+REFERENCE_DIR = "reference_xray_absorption_spectra_using_bse"
+
+def test_groundstate_BN(converged_results_BN: dict):
+    """Test results (for cubic boron-nitride) in main output file INFO.OUT
+    containing the essential information on the material system, parameters of the calculation and results
+    """
+   
+    total_energy = -79.38465862
+    assert np.isclose(converged_results_BN['Total energy'], total_energy),\
+        f"Incorrect value for total energy. Expect {total_energy}"
+
+    fermi_energy = 0.42726809
+    assert np.isclose(converged_results_BN['Fermi energy'], fermi_energy),\
+        f"Incorrect value for fermi energy. Expect {fermi_energy}"
+
+    kinetic_energy = 79.03180996
+    assert np.isclose(converged_results_BN['Kinetic energy'], kinetic_energy),\
+        f"Incorrect value for kinetic energy. Expect {kinetic_energy}"
+
+    coulomb_energy = -147.71206057
+    assert np.isclose(converged_results_BN['Coulomb energy'], coulomb_energy),\
+        f"Incorrect value for coulomb energy. Expect {coulomb_energy}"
+
+    exchange_energy = -10.2100207
+    assert np.isclose(converged_results_BN['Exchange energy'], exchange_energy),\
+        f"Incorrect value for exchange energy. Expect {exchange_energy}"
+
+    correlation_energy = -0.49438731
+    assert np.isclose(converged_results_BN['Correlation energy'], correlation_energy),\
+        f"Incorrect value for correlation energy. Expect {correlation_energy}"
+
+    fermi_energy_dos = 0.0
+    assert np.isclose(converged_results_BN['DOS at Fermi energy (states/Ha/cell)'], fermi_energy_dos),\
+        f"Incorrect value for DOS at Fermi energy (states/Ha/cell). Expect {fermi_energy_dos}"
+
+    charge_core_electrons = 4.0
+    assert np.isclose(converged_results_BN['core'], charge_core_electrons),\
+        f"Incorrect value for charge of core electrons. Expect {charge_core_electrons}"
+
+    core_leakage_charge = 0.0014585
+    assert np.isclose(converged_results_BN['core leakage'], core_leakage_charge),\
+        f"Incorrect value for core leakage charge. Expect {core_leakage_charge}"
+
+    charge_valance_electrons = 8.0
+    assert np.isclose(converged_results_BN['valence'], charge_valance_electrons),\
+        f"Incorrect value for charge of valence electrons. Expect {charge_valance_electrons}"
+
+    interstitial_region_charge = 3.23877764
+    assert np.isclose(converged_results_BN['interstitial'], interstitial_region_charge),\
+        f"Incorrect value for charge in interstitial region. Expect {interstitial_region_charge}"
+
+    muffin_tins_charge = 8.76122236
+    assert np.isclose(converged_results_BN['total charge in muffin-tins'], muffin_tins_charge),\
+        f"Incorrect value for total charge in muffin-tins. Expect {muffin_tins_charge}"
+
+    total_charge = 12.0
+    assert np.isclose(converged_results_BN['total charge'], total_charge),\
+        f"Incorrect value for total charge. Expect {total_charge}"
+    
+    estimated_gap = 0.15663379
+    assert np.isclose(converged_results_BN['Estimated fundamental gap'], estimated_gap),\
[…]
