+++ tools/excitingscripts/excitingscripts/execute/volume_optimization.py
@@ -1,5 +1,4 @@
 import os
-import time
 from argparse import ArgumentParser
 from os.path import join
 
