### 2025-04-02 96aa86f
Include NEXC, ETOT, and JIND as tested files when restarting RT-TDDFT calculations
