+++ test/tests_config.yml
@@ -724,13 +724,18 @@ RT-TDDFT/LDA_PW-SiC-trapCos-AETRS-uKS-frozen-restart:
       - "OCCSV_TXT_60.OUT"
       - "PROJ_40.OUT"
       - "PROJ_60.OUT"
+      - "JIND.OUT"
+      - "NEXC.OUT"
+      - "ETOT_RTTDDFT.OUT"
     inputs:
       - "input.xml"
       - "C.xml"
       - "Si.xml"
-      - "JIND.OUT"
-      - "PVEC.OUT"
-      - "AVEC.OUT"
+      - "JIND.OUT.SAVE"
+      - "PVEC.OUT.SAVE"
+      - "AVEC.OUT.SAVE"
+      - "NEXC.OUT.SAVE"
+      - "ETOT_RTTDDFT.OUT.SAVE"
       - "EVECFV_RTTDDFT.OUT"
       - "EVECFV_PREVIOUS_RTTDDFT.OUT"
   
