### 2025-04-08 33d288b
More assertions
