+++ tools/excitingjupyter/excitingjupyter/03_excited_states/02_BSE/assertions/excited_states_from_bse.py
@@ -7,7 +7,7 @@ TUTORIAL_EXCITED_STATES_FROM_BSE_RUNDIR = "../run_LiF_BSE/BSE/EPSILON"
 REFERENCE_DIR = "reference_excited_states_from_BSE/EPSILON"
 
 def test_tutorial_excited_states_from_bse(file_name: str):
-    """Automatically test results of tutorial_excited_states_from_bse notebook, LiF bulk calculation.
+    """Automatically test results of excited_states_from_bse tutorial, LiF bulk calculation.
     """
     
     epsilon_reference = parse(f"{dirname(__file__)}/{REFERENCE_DIR}/{file_name}")
