### 2025-04-10 3e3db37
Fermi level calculation in GW taskgroup
