+++ src/src_gw/task_QPEigenvalues.f90
@@ -10,8 +10,9 @@ module task_QPEigenvalues
   use mod_selfenergy, only: evalks, evalqp, selfex, sigc, znorm, read_selfec_from_files, &
     read_selfex_from_files, generate_frequency_grid_for_correlation_self_energy
   use mod_vxc, only: read_vxcnn, vxcnn
+  use modgw, only: nvelgw
   use modinput, only: input, gw_type
-  use modmpi, only: terminate_if_false, mpiglobal
+  use modmpi, only: terminate_if_false, terminate, mpiglobal
   use precision, only: dp, i32
   use quasiparticle_energies, only: write_qp_energies_text_format
   use to_char_conversion, only: to_char
@@ -24,33 +25,54 @@ module task_QPEigenvalues
   character(len=*), parameter :: file_name_qp_energies = 'EVALQP'
   character(len=*), parameter :: extension_binary_format ='.OUT'
 
+  enum, bind(C)
+    enumerator :: method_to_obtain_Fermi_level
+    enumerator :: libbzint, DFT_VBM_CBm
+  end enum
+
   !> Interface to the parameters defined in the input file
   type task_QPEigenvalues_parameters
     private
     type(kpoints_sets) :: k_points
+    integer(kind(method_to_obtain_Fermi_level)) :: method_Fermi_level
   contains
-    procedure :: parse_input, sanity_checks
+    procedure :: parse_input, sanity_checks => task_QPEigenvalues_sanity_checks
   end type
 
   public :: execute_task_QPEigenvalues
 
 contains
+!> Convert a string to the enum [[method_to_obtain_Fermi_level]]
+pure function string_to_method_to_obtain_Fermi_level( string ) result(method)
+  !> String to be converted
+  character(len=*), intent(in) :: string
+  integer(kind(method_to_obtain_Fermi_level)) :: method
+
+  method = DFT_VBM_CBm
+  select case( trim( string ) )
+    case("from_DFT_VBM_CBM_indexes")
+      method = DFT_VBM_CBm
+    case("from_QP_Eigenvalues")
+      method = libbzint
+  end select
+end function
+
 !> Interface to the parameters defined in the input file
 subroutine parse_input( this, gw_inp, n_kpt )
   class(task_QPEigenvalues_parameters), intent(inout) :: this
   !> type with the variables given in the input file
   type(gw_type), intent(in):: gw_inp
-  !> maximum number of q-points
+  !> maximum number of k-points
   integer(i32), intent(in) :: n_kpt
 
   call this%sanity_checks( gw_inp )
   call this%k_points%parse_input( gw_inp%taskGroup%QPEigenvalues%kpointsarray, n_kpt )
-
+  this%method_Fermi_level = string_to_method_to_obtain_Fermi_level( gw_inp%taskGroup%QPEigenvalues%FermiLevel )
 end subroutine
 
 
 !> Perform sanity checks on the input parameters in the `gw` element
-subroutine sanity_checks( this, gw_inp )
+subroutine task_QPEigenvalues_sanity_checks( this, gw_inp )
   class(task_QPEigenvalues_parameters), intent(in) :: this
   !> type with the variables given in the input file
   type(gw_type), intent(in) :: gw_inp
@@ -59,7 +81,6 @@ subroutine sanity_checks( this, gw_inp )
     'Element QPEigenvalues must be present when executing '//'"'//task_name//'"' )
 
 end subroutine
-
 
 !> Execute a task QPEigenvalues calculation
 subroutine execute_task_QPEigenvalues( first_band, last_band, k_points_irreducible, file_format )
[…]
