### 2025-04-10 05697c1
More assertions
