+++ tools/excitingjupyter/excitingjupyter/03_excited_states/04_Others/assertions/graphene_from_ground_state_to_excitations.py
@@ -1,0 +1,144 @@
+from os.path import dirname
+
+import numpy as np
+
+from excitingtools.exciting_dict_parsers.groundstate_parser import parse_info_out
+
+TUTORIAL_RUNDIR = "../graphene"
+REFERENCE_DIR = "reference_graphene"
+
+
+def test_groundstate(converged_results: dict):
+    """Test results for groundstate calculation in main output file INFO.OUT.
+    """
+
+    total_energy = -76.11387789
+    assert np.isclose(converged_results['Total energy'], total_energy), \
+        f"Incorrect value for total energy. Expect {total_energy}"
+
+    fermi_energy = -0.06783556
+    assert np.isclose(converged_results['Fermi energy'], fermi_energy), \
+        f"Incorrect value for fermi energy. Expect {fermi_energy}"
+
+    kinetic_energy = 75.81190852
+    assert np.isclose(converged_results['Kinetic energy'], kinetic_energy), \
+        f"Incorrect value for kinetic energy. Expect {kinetic_energy}"
+
+    coulomb_energy = -141.13032000
+    assert np.isclose(converged_results['Coulomb energy'], coulomb_energy), \
+        f"Incorrect value for coulomb energy. Expect {coulomb_energy}"
+
+    exchange_energy = -10.37420659
+    assert np.isclose(converged_results['Exchange energy'], exchange_energy), \
+        f"Incorrect value for exchange energy. Expect {exchange_energy}"
+
+    correlation_energy = -0.42125982
+    assert np.isclose(converged_results['Correlation energy'], correlation_energy), \
+        f"Incorrect value for correlation energy. Expect {correlation_energy}"
+
+    fermi_energy_dos = 0.00002749
+    assert np.isclose(converged_results['DOS at Fermi energy (states/Ha/cell)'], fermi_energy_dos), \
+        f"Incorrect value for DOS at Fermi energy (states/Ha/cell). Expect {fermi_energy_dos}"
+
+    charge_core_electrons = 4.00000000
+    assert np.isclose(converged_results['core'], charge_core_electrons), \
+        f"Incorrect value for charge of core electrons. Expect {charge_core_electrons}"
+
+    core_leakage_charge = 0.00129555
+    assert np.isclose(converged_results['core leakage'], core_leakage_charge), \
+        f"Incorrect value for core leakage charge. Expect {core_leakage_charge}"
+
+    charge_valance_electrons = 8.00000000
+    assert np.isclose(converged_results['valence'], charge_valance_electrons), \
+        f"Incorrect value for charge of valence electrons. Expect {charge_valance_electrons}"
+
+    interstitial_region_charge = 4.85400965
+    assert np.isclose(converged_results['interstitial'], interstitial_region_charge), \
+        f"Incorrect value for charge in interstitial region. Expect {interstitial_region_charge}"
+
+    muffin_tins_charge = 7.14599035
+    assert np.isclose(converged_results['total charge in muffin-tins'], muffin_tins_charge), \
+        f"Incorrect value for total charge in muffin-tins. Expect {muffin_tins_charge}"
+
+    total_charge = 12.00000000
+    assert np.isclose(converged_results['total charge'], total_charge), \
+        f"Incorrect value for total charge. Expect {total_charge}"
+
+    estimated_gap = 0.07363036
+    assert np.isclose(converged_results['Estimated fundamental gap'], estimated_gap), \
+        f"Incorrect value for Estimated fundamental gap. Expect {estimated_gap}"
+
+def test_energy_vs_strain():
+    energy_vs_strain_results = np.loadtxt(f"{dirname(__file__)}/{TUTORIAL_RUNDIR}/opt/opt-run/energy-vs-strain")
+    energy_vs_strain_reference = np.array([[-0.10000000, -76.03156387],
[…]
