+++ .gitlab/jupyter-exciting-python3-gnu-smp.yml
@@ -47,8 +47,8 @@ test-fedora-buster-gnu-smp-jupyter:
  # - time python3 02_ground_state/01_Methods/assertions/van_der_waals_corrections.py
  # - time jupyter nbconvert --execute 02_ground_state/02_Electronic_Properties/fermi_surface_visualization.ipynb --to markdown
  # - time python3 02_ground_state/02_Electronic_Properties/assertions/fermi_surface_visualization.py
-  - time jupyter nbconvert --execute 02_ground_state/02_Electronic_Properties/spin_texture_calculation.ipynb --to markdown
-  - time python3 02_ground_state/02_Electronic_Properties/assertions/spin_texture_calculation.py
+ # - time jupyter nbconvert --execute 02_ground_state/02_Electronic_Properties/spin_texture_calculation.ipynb --to markdown
+ # - time python3 02_ground_state/02_Electronic_Properties/assertions/spin_texture_calculation.py
   - time jupyter nbconvert --execute 02_ground_state/03_Lattice_Optimization/simple_examples_of_structure_optimization.ipynb --to markdown
   - time python3 02_ground_state/03_Lattice_Optimization/assertions/simple_examples_of_structure_optimization.py
   - time jupyter nbconvert --execute 02_ground_state/03_Lattice_Optimization/volume_optimization_for_cubic_systems.ipynb --to markdown
@@ -76,8 +76,8 @@ test-fedora-buster-gnu-smp-jupyter:
   - time python3 03_excited_states/02_BSE/assertions/xray_absorption_spectra_using_bse.py
   - time jupyter nbconvert --execute 03_excited_states/02_BSE/xray_emission_spectra.ipynb --to markdown
   - time python3 03_excited_states/02_BSE/assertions/xray_emission_spectra.py
-  - time jupyter nbconvert --execute 03_excited_states/02_BSE/q_dependent_bse_calculations.ipynb --to markdown
-  - time python3 03_excited_states/02_BSE/assertions/q_dependent_bse_calculations.py
+ # - time jupyter nbconvert --execute 03_excited_states/02_BSE/q_dependent_bse_calculations.ipynb --to markdown
+ # - time python3 03_excited_states/02_BSE/assertions/q_dependent_bse_calculations.py
   - time jupyter nbconvert --execute 03_excited_states/02_BSE/optical_spectra_from_fastBSE.ipynb --to markdown
   - time python3 03_excited_states/02_BSE/assertions/optical_spectra_from_fastBSE.py
   - time jupyter nbconvert --execute 03_excited_states/03_TDDFT/excited_states_from_tddft.ipynb --to markdown
