+++ src/src_xs/src_rttddft/rttddft_main.f90
@@ -102,7 +102,7 @@ contains
     ! k-dependent Hamiltonian's dimensions array (first_kpt : last_kpt)
     integer(i32), allocatable :: k_dependent_dims(:)
     ! KS-LAPW+lo transition matrix (nmatmax, nstfv, first_kpt : last_kpt)
-    complex(dp), allocatable :: ks_lapwo_transition_matrix(:, :, :)
+    complex(dp), allocatable :: ks_lapwlo_transition_matrix(:, :, :)
 
     integer(i32) :: it, first_kpt, last_kpt, first_step, last_step, i_print, &
       time_step_multiplier, lmax_potential
@@ -179,7 +179,7 @@ contains
     
     call initialize_rttddft( rt, propagator, vec_pot, a_tot_save, &
       molecular_dynamics, psi, overlap, ham_init, ham_time, ham_past, effective_potential_init, apwalm, &
-      pmat, pmatmt, rhomt_frozen, rhoir_frozen, occupations, k_dependent_dims, eps_occ, Gkset, Gset, ks_lapwo_transition_matrix )
+      pmat, pmatmt, rhomt_frozen, rhoir_frozen, occupations, k_dependent_dims, eps_occ, Gkset, Gset, ks_lapwlo_transition_matrix )
     call distribute_loop( mpi_env_k, nkpt, first_kpt, last_kpt )
     if( molecular_dynamics%on ) call init_MD( time, vec_pot%a_tot, dt, &
       psi%active, occupations, overlap, ham_time, time_step_multiplier, molecular_dynamics, &
@@ -217,7 +217,7 @@ contains
     ! Total energy
     if ( rt%calculate_total_energy .and. rt%do_from_scratch() ) then
       if ( psi%has_frozen() ) call update_density( first_kpt, psi, occupations, 0, &
-        .false., rt%l_rad_step, rhomt_frozen, rhoir_frozen, ks_lapwo_transition_matrix )
+        .false., rt%l_rad_step, rhomt_frozen, rhoir_frozen, ks_lapwlo_transition_matrix )
       call potcoul()
       call potxc()
       call obtain_energy_rttddft( first_kpt, ham_time, psi, occupations, mpi_env_k, etotstore(1) )
@@ -234,7 +234,7 @@ contains
     if ( rt%screenshots%on ) then
       if ( rt%screenshots%density%on ) then
         call update_density( first_kpt, psi, occupations, 0, rt%normalize_WF, rt%l_rad_step, &
-          rhomt_frozen, rhoir_frozen, ks_lapwo_transition_matrix, dens_case=groundstate )
+          rhomt_frozen, rhoir_frozen, ks_lapwlo_transition_matrix, dens_case=groundstate )
         rhomt_init = rhomt
         rhoir_init = rhoir
       end if
@@ -293,7 +293,7 @@ contains
 
       ! DENSITY
       if ( density_needed ) call update_density( first_kpt, psi, occupations, it, rt%normalize_WF, &
-        rt%l_rad_step, rhomt_frozen, rhoir_frozen, ks_lapwo_transition_matrix, rt%printTimings, timing%t_RTTDDFT%dens )
+        rt%l_rad_step, rhomt_frozen, rhoir_frozen, ks_lapwlo_transition_matrix, rt%printTimings, timing%t_RTTDDFT%dens )
       
       ! KS-POTENTIAL
       if ( .not. rt%eeInteraction%ipa ) call update_potential( rt%printTimings, timing%t_RTTDDFT%pot )
@@ -332,7 +332,7 @@ contains
           call update_hamiltonian_without_pa_term_lapw( first_kpt, vec_pot%a_tot, ham_time, apwalm, &
           rt%printTimings, timing%t_RTTDDFT%ham )
[…]
