### 2025-04-14 e402005
Update assertions
