### 2025-04-15 d44ad8b
Parsing of convergence criteria and allow pathlib.Path input for parser
