+++ .gitlab/jupyter-exciting-python3-gnu-smp.yml
@@ -55,8 +55,8 @@ test-fedora-buster-gnu-smp-jupyter:
   - time python3 02_ground_state/03_Lattice_Optimization/assertions/volume_optimization_for_cubic_systems.py
   - time jupyter nbconvert --execute 02_ground_state/04_Molecules/how_to_run_calculations_for_simple_molecules.ipynb --to markdown
   - time python3 02_ground_state/04_Molecules/assertions/how_to_run_calculations_for_simple_molecules.py
-  - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/born_effective_charges_of_zincblende_structure_crystals.ipynb --to markdown
-  - time python3 02_ground_state/05_Lattice_Dynamics/assertions/born_effective_charges_of_zincblende_structure_crystals.py
+ # - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/born_effective_charges_of_zincblende_structure_crystals.ipynb --to markdown
+ # - time python3 02_ground_state/05_Lattice_Dynamics/assertions/born_effective_charges_of_zincblende_structure_crystals.py
   - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/how_to_visualize_phonons.ipynb --to markdown
   - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/lattice_dynamics_of_diamond_and_zincblende_structure_crystals.ipynb --to markdown
   - time python3 02_ground_state/05_Lattice_Dynamics/assertions/lattice_dynamics_of_diamond_and_zincblende_structure_crystals.py
