+++ .gitlab/jupyter-exciting-python3-gnu-smp.yml
@@ -58,8 +58,8 @@ test-fedora-buster-gnu-smp-jupyter:
  # - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/born_effective_charges_of_zincblende_structure_crystals.ipynb --to markdown
  # - time python3 02_ground_state/05_Lattice_Dynamics/assertions/born_effective_charges_of_zincblende_structure_crystals.py
   - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/how_to_visualize_phonons.ipynb --to markdown
-  - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/lattice_dynamics_of_diamond_and_zincblende_structure_crystals.ipynb --to markdown
-  - time python3 02_ground_state/05_Lattice_Dynamics/assertions/lattice_dynamics_of_diamond_and_zincblende_structure_crystals.py
+ # - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/lattice_dynamics_of_diamond_and_zincblende_structure_crystals.ipynb --to markdown
+ # - time python3 02_ground_state/05_Lattice_Dynamics/assertions/lattice_dynamics_of_diamond_and_zincblende_structure_crystals.py
   - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/phonons_at_gamma_in_diamond_structure_crystals.ipynb --to markdown
   - time python3 02_ground_state/05_Lattice_Dynamics/assertions/phonons_at_gamma_in_diamond_structure_crystals.py
   - time jupyter nbconvert --execute 02_ground_state/05_Lattice_Dynamics/phonons_at_X_in_diamond_structure_crystals.ipynb --to markdown
