### 2025-04-17 f03ad90
RT-TDDFT - refactor of IO operations
