### 2025-04-17 be35394
Update module exciting_mpi
