### 2025-04-17 f1bae56
GPU: src/src_gw/expand_products.f90
