+++ src/src_gw/expand_products_ifort.f90
@@ -1,295 +1,0 @@
-subroutine expand_products(ik,iq,nstart,nend,nsplit,mstart,mend,msplit,minm)
-
-    use modmain, only : zzero
-    use modgw,   only : mbsiz
-    implicit none
-
-    ! input variables
-    integer, intent(in) :: ik
-    integer, intent(in) :: iq
-    integer, intent(in) :: nstart, nend
-    integer, intent(in) :: nsplit         ! last index of valence states
-    integer, intent(in) :: mstart ,mend
-    integer, intent(in) :: msplit         ! last index of valence states
-    complex(8), intent(out) :: minm(mbsiz,nstart:nend,mstart:mend)
-    
-    ! local variables
-    integer :: cstart, cend
-    integer :: iflag
-    
-    minm(:,:,:) = zzero
-    
-    if ((nsplit>0).and.(msplit<=0)) then
-
-        !======================================================================
-        ! When calculating the dielectric function:
-        !
-        ! n -> {1:nomax}{1:ncg}, so nsplit=nomax, when core states are treated
-        ! m -> {numin:nstfv},       msplit is ignored
-        !
-        ! Important to specify msplit<=0 
-        ! to ignore the self energy related part (below)
-        ! 
-        !======================================================================    
-        if (nend<=nsplit) then
-        
-            ! calculate M^i_{nm}
-            iflag = 1
-            call expand_products_block(ik,iq,nstart,nend,mstart,mend, &
-            &                          minm(:,nstart:nend,mstart:mend), &
-            &                          iflag)
-        
-        else if (nstart>nsplit) then
-        
-            ! calculate M^i_{cm}
-            iflag = 2
-            cstart = nstart-nsplit
-            cend = nend-nsplit
-            call expand_products_block(ik,iq,cstart,cend,mstart,mend, &
-            &                          minm(:,nstart:nend,mstart:mend), &
-            &                          iflag)
-        
-        else
-        
-            ! calculate M^i_{nm}
-            iflag = 1
-            call expand_products_block(ik,iq,nstart,nsplit,mstart,mend, &
-            &                          minm(:,nstart:nsplit,mstart:mend), &
-            &                          iflag)
-            ! calculate M^i_{cm}
-            iflag = 2
-            cstart = 1
-            cend = nend-nsplit
-            call expand_products_block(ik,iq,cstart,cend,mstart,mend, &
-            &                          minm(:,nsplit+cstart:nend,mstart:mend), &
-            &                          iflag)
-        
-        end if ! dielectric function part     
-    
-    else if ((nsplit<=0).and.(msplit>0)) then
-      
-      !======================================================================
-      ! When calculating the self energy:
-      !
-      ! n -> {ibgw:nbgw},      nsplit is ignored
-      ! m -> {1:nstfv}{1:ncg}, so msplit=nstfv, when core states are treated
-      !
-      ! Important to specify nsplit<=0 
-      ! to ignore the dielectric function part (above)
-      !
-      !======================================================================    
-      if (mend<=msplit) then
-    
-        ! calculate M^i_{nm}
-        iflag = 1
-        call expand_products_block(ik,iq,nstart,nend,mstart,mend, &
-        &                          minm(:,nstart:nend,mstart:mend), &
-        &                          iflag)
-      
-      else if (mstart>msplit) then
-    
-        ! calculate M^i_{nc}
-        iflag = 3
-        cstart = mstart-msplit
-        cend = mend-msplit
-        call expand_products_block(ik,iq,nstart,nend,cstart,cend, &
[…]
