### 2025-04-17 35e648a
Define new type (trajectory) and refactor existing code
