+++ src/src_xs/src_rttddft/rttddft_MD.f90
@@ -8,7 +8,7 @@ module rttddft_MD
   use asserts, only: assert
   use constants, only: zone, zzero
   use exciting_mpi, only: mpiinfo, xmpi_allreduce
-  use MD, only: MD_input_keys, MD_timing, force, obtain_core_corrections, force_ext, &
+  use MD, only: trajectory, MD_input_keys, MD_timing, force, obtain_core_corrections, force_ext, &
     obtain_Hellmann_Feynman_force, obtain_valence_corrections_part1, &
     val_corr_pt2_given_atom_and_kpt => obtain_valence_corrections_part2
   use mod_atoms, only: atposc, idxas, natoms, natmtot, nspecies, &
@@ -27,7 +27,7 @@ module rttddft_MD
   use physical_constants, only: c
   use precision, only: dp, i32
   use rttddft_electric_field, only: Electric_Field
-  use rttddft_GlobalMDVariables, only: mathcalH, mathcalB
+  use rttddft_GlobalMDVariables, only: mathcalH, mathcalB, update_exciting_globals_for_new_ions_positions
   use rttddft_timings, only: Print_Timings, timesec_RTTDDFT
   use rttddft_VectorPotential, only: Vector_Potential_Field
   use vector_multiplication, only: dot_multiply
@@ -36,8 +36,12 @@ module rttddft_MD
 
   private
 
-  public :: force_rttdft, move_ions, allocate_global_arrays, deallocate_global_arrays, &
-    evaluate_charge_val
+  public :: allocate_global_arrays, &
+            deallocate_global_arrays, &
+            evaluate_charge_val, &
+            force_rttdft, &
+            move_ions, &
+            update_basis_derivative
 
   !> valence charge of each species
   real(dp), allocatable :: charge_val(:)
@@ -195,37 +199,37 @@ contains
       forces_val = forces_val + sumaux
   end subroutine
 
-  subroutine move_ions( first_kpt, forces, forces_old, dt, atoms_velocities, apwalm, &
+  subroutine move_ions( first_kpt, forces, forces_old, dt, nuclei_motion, apwalm, &
     printTimings, t_MD )
     !> The first k point
     integer(i32), intent(in) :: first_kpt
     !> Forces acting on each atom at time \( t \)
-    real(dp), intent(in)            :: forces(:, :)
+    real(dp), contiguous, intent(in) :: forces(:, :)
     !> Forces acting on each atom at time \( t - \Delta t \)
-    real(dp), intent(in)            :: forces_old(:, :)
+    real(dp), contiguous, intent(in) :: forces_old(:, :)
     !> Time step for the molecular dynamics
-    real(dp), intent(in)            :: dt
-    !> Velocities of the nuclei at time \( t \)
-    real(dp), intent(inout)         :: atoms_velocities(:,:)
+    real(dp), intent(in) :: dt
+    !> This argument packs nuclei positions and velocities at time \(t\)
+    class(trajectory), intent(inout) :: nuclei_motion
     !> Matching coefficients of the (L)APWs
     !> (ngkmax, apwordmax, lmmaxapw, natmtot, first_kpt : last_kpt)
-    complex(dp), intent(inout) :: apwalm(:, :, :, :, first_kpt :)
+    complex(dp), contiguous, intent(inout) :: apwalm(:, :, :, :, first_kpt :)
     !> Object that packs information about printing of timings [[Print_Timings]]
     type(Print_Timings), optional, intent(in) :: printTimings
     !> Object that packs information about timings spent in MD
[…]
