+++ src/src_gw/calcmwm.f90
@@ -4,63 +4,107 @@ subroutine calcmwm(nstart, nend, mstart, mend, minm)
     use modinput
     use constants, only: zone, zzero, pi
     use modgw,   only: vi, kqset, Gamma, singc1, singc2, mbsiz, &
-    &                  epsilon, epsh, epsw1, epsw2, freq, mwm
+                       epsilon, epsh, epsw1, epsw2, mwm
+    use precision, only: dp, i32
+    use mod_pointer_remapping, only: remap_fortran_pointer 
+    use mod_device_offload,    only: device_world
+    use device_linalg_common_interface, only: zgemm_batched_gpu, zhemm_gpu, zgemm_gpu, zhemv_gpu
+    use m_memory_device,       only: allocate_device_memory, deallocate_device_memory, &
+                                   bytes_double_complex, bytes_int, get_device_pointer
+    use iso_c_binding, only: c_ptr, c_size_t, c_f_pointer
+#include "offload.fpp"
+
     implicit none
 
     ! input variables
-    integer(4), intent(in) :: nstart, nend
-    integer(4), intent(in) :: mstart, mend
-    complex(8), intent(in) :: minm(mbsiz, nstart:nend, mstart:mend)
+    integer(i32), intent(in) :: nstart, nend
+    integer(i32), intent(in) :: mstart, mend
+    complex(dp), intent(in) :: minm(mbsiz, nstart:nend, mstart:mend)
 
     ! local variables
-    integer(4) :: iom
-    integer(4) :: ie1, ie2
-    real(8)    :: wkq
-    real(8)    :: vi4pi, coefs1, coefs2, coefs3
-    complex(8), allocatable :: wm(:)
-    complex(8), external    :: zdotu, zdotc
-    external zhemm
+    integer(i32) :: iom
+    integer(i32) :: ndim, mdim, nmdim, nomeg
+    integer(i32) :: ie1, ie2, iemin, iemax, ieg, imixcb
+    integer(i32) :: my_device
+    real(dp)     :: vi4pi, wkq
+    complex(dp)  :: coefs1, coefs2
+    type(c_ptr)  :: wm_cptr
+    complex(dp), contiguous, pointer :: wm(:,:,:)
 
-    wkq    = 1.d0/dble(kqset%nkpt)
-    
+    my_device = device_world%get_device()
+
+    wkq    = 1.0_dp/real(kqset%nkpt,kind=dp)
+    ndim   = nend - nstart + 1
+    mdim   = mend - mstart + 1
+    nmdim  = mdim*ndim
+    nomeg  = size( epsilon, 3 )
+
     if (Gamma) then
-      vi4pi  = 4.d0*pi*vi
+      vi4pi  = 4.0_dp*pi*vi
       coefs1 = singc1*sqrt(vi4pi)
       coefs2 = singc2*vi4pi
+      iemin  = max(nstart,mstart)
+      iemax  = min(nend,mend)
     end if 
 
     !-------------------------------------------------
     ! calculate \sum_{ij} M^i_{nm}* W^c_{ij} M^j_{nm}
     !-------------------------------------------------
-#ifdef USEOMP
-!$OMP PARALLEL DEFAULT(SHARED) PRIVATE(wm,iom,ie1,ie2)
-#endif
-    allocate(wm(mbsiz))
-    do iom = 1, freq%nomeg
+    call allocate_device_memory(wm_cptr, int(mbsiz*nmdim,kind=c_size_t)*bytes_double_complex, my_device)
+    call c_f_pointer(wm_cptr, wm, int([mbsiz,ndim,mdim],kind=c_size_t))
+    ! We are remapping the boundaries of a pointer with default Fortran lower bound 
[…]
