### 2025-04-18 2560e97
Add new species file functions
