### 2025-04-22 9f8867e
Update assertions
