+++ src/src_xs/src_rttddft/rttddft_main.f90
@@ -37,7 +37,7 @@ module rttddft_main
     open_file_nexc, close_file_nexc, write_nexc, &
     open_file_etot, close_file_etot, write_total_energy, &
     open_file_info, close_file_info, write_file_info, write_file_info_header, &
-    write_wavefunction, t, t_minus_dt, copy_files
+    write_wavefunction, t, t_minus_dt, copy_files, write_state_Ehrenfest_MD, read_state_Ehrenfest_MD
   use rttddft_MD, only: force_rttdft, move_ions, update_basis_derivative, &
     MD_allocate_global_arrays => allocate_global_arrays, &
     MD_deallocate_global_arrays => deallocate_global_arrays, &
@@ -154,11 +154,18 @@ contains
     if( rt%restart_previous_calculation() ) then
       if( rt%restart_extension /= "" ) then
         ! Copy files: sources are files ending with `rt%restart_extension`, dest. are to the default file names
-        if( my_rank_writes_to_output ) call copy_files( rt%restart_extension, rt%calculate_n_exc, rt%calculate_total_energy )
+        if( my_rank_writes_to_output ) then 
+          call copy_files( rt%restart_extension, rt%calculate_n_exc, rt%calculate_total_energy )
+          if( molecular_dynamics%on ) call MD_outputs%copy_files( rt%restart_extension, molecular_dynamics%print_all_force_components )
+        end if
         ! Before reading, ensure that copying has been finished
         call barrier()
       end if
       call read_time_and_fields( time, p_vec, vec_pot, a_ind_save, a_tot_save )
+      if( molecular_dynamics%on ) then
+        call MD_outputs%read_time_and_forces_from_files( time_aux, forces )
+        call terminate_if_false( time == time_aux, "Last time t is not the same across RT-TDDFT and MD output files")
+      end if
     else
       time = 0._dp
     end if
@@ -175,9 +182,25 @@ contains
       pmat, pmatmt, rhomt_frozen, rhoir_frozen, occupations, k_dependent_dims, eps_occ, Gkset, Gset, ks_lapwlo_transition_matrix )
     call distribute_loop( mpi_env_k, nkpt, first_kpt, last_kpt )
     if( molecular_dynamics%on ) then
-      call init_MD( time, vec_pot%a_tot, dt, &
+      call init_MD( rt%do_from_scratch(), time, vec_pot%a_tot, dt, &
         psi%active, occupations, overlap, ham_time, time_step_multiplier, molecular_dynamics, &
         MD_outputs, nuclei_motion, e_field, forces )
+      if( rt%restart_previous_calculation() ) then
+        call nuclei_motion%allocate_arrays( natmtot )
+        call read_state_Ehrenfest_MD( nuclei_motion )
+        call nuclei_motion%update_globals( )
+        call update_exciting_globals_for_new_ions_positions( first_kpt, apwalm )
+        if( molecular_dynamics%update_overlap .or. allocated(mathcalH) .or. &
+            & allocated(mathcalB) .or. molecular_dynamics%update_pmat ) then
+          if( molecular_dynamics%update_pmat ) &
+            call obtain_pmat_LAPWLOBasis( first_kpt, rt%pmat%force_pmat_hermitian, apwalm, pmat, pmatmt )
+          if( molecular_dynamics%update_overlap ) call update_overlap_lapw( first_kpt, vec_pot%a_tot, overlap, &
+            apwalm, pmatmt, update_mathcalH=allocated( mathcalH ), update_mathcalB=allocated( mathcalB ) )
+          if( propagator%extrapolation_needed() ) ham_past = ham_time  
[…]
