### 2025-04-27 c091c4f
Update assertions
