### 2025-04-29 1c0ec6c
Update assertions
