### 2025-04-30 25c5c41
Update assertions
