### 2025-05-02 f6bbe57
BUGFIX: Fourier interpolation failure
