+++ src/src_gw/task_band.f90
@@ -13,7 +13,7 @@ subroutine task_band()
   real(8)       :: tstart, tend
   integer       :: i, j, is, ia, ias, l
   character(80) :: fname, s
-  logical       :: exist, bandchar
+  logical       :: exist
   character(128) :: buffer
   Type (xmlf_t), Save :: xf
   real(8), allocatable :: bc(:,:,:,:)
@@ -58,14 +58,7 @@ subroutine task_band()
     fname = 'EVALQP.OUT'
   end if
 
-  bandchar = .false.
-  if( bandchar) then
-    allocate(bc(0:3, natmtot, nstsv, nkpt))
-    bc = 0.d0
-    call getevalqp(fname, nkpt, vkl, evalsv, bc)
-  else
-    call getevalqp(fname, nkpt, vkl, evalsv)
-  end if
+  call getevalqp(fname, nkpt, vkl, evalsv)
 
   !----------------------------------
   ! write QP bandstructure to disk
@@ -75,86 +68,32 @@ subroutine task_band()
   call xml_AddXMLPI(xf,"xml-stylesheet", 'href="'//trim(input%xsltpath)//&
        &'/visualizationtemplates/bandstructure2html.xsl" type="text/xsl"')
 
-  if( .not. bandchar) then
-      open( 50, file='BAND-QP.OUT', action='WRITE', form='FORMATTED')
-      call xml_NewElement( xf, "bandstructure")
-      call xml_NewElement( xf, "title")
-      call xml_AddCharacters( xf, trim( input%title))
-      call xml_endElement( xf, "title")
-      open(51, File="bandstructure-qp.dat", Action='Write', Form='Formatted')
-      write(51,*) "# ", ibgw, min(nbgw,nstsv), nkpt
-      do ib = ibgw, min(nbgw,nstsv)
-        call xml_NewElement( xf, "band")
-        do ik = 1, nkpt
-          write(50,'(2G18.10)') dpp1d(ik), evalsv(ib,ik)
-          write(51,'(2I6, 3F12.6, 2G18.10)') ib, ik, vkl(:,ik), dpp1d(ik), evalsv(ib,ik)
-          call xml_NewElement( xf, "point")
-          write( buffer, '(5G18.10)') dpp1d (ik)
-          call xml_AddAttribute( xf, "distance", trim( adjustl( buffer)))
-          write( buffer, '(5G18.10)') evalsv( ib, ik)
-          call xml_AddAttribute (xf, "eval", trim( adjustl( buffer)))
-          call xml_endElement( xf, "point")
-        end do !ik
-        call xml_endElement( xf, "band")
-        write(50,*)
-        write(51,*)
-      end do
-      close(50)
-      close(51)
-  else
-      call xml_NewElement (xf, "bandstructure")
-      call xml_AddAttribute (xf, "character", "true")
-      call xml_NewElement (xf, "title")
-      call xml_AddCharacters (xf, trim(input%title))
-      call xml_endElement (xf, "title")
-      do is = 1, nspecies
-        call xml_NewElement (xf, "species")
-        call xml_AddAttribute (xf, "name", trim(spname(is)))
-        call xml_AddAttribute (xf, "chemicalSymbol", trim(input%structure%speciesarray(is)%species%chemicalSymbol))
-        do ia = 1, natoms (is)
-          call xml_NewElement (xf, "atom")
-          write (buffer, '(5G18.10)') atposc (:, ia, is)
-          call xml_AddAttribute (xf, "coord", &
-               & trim(adjustl(buffer)))
-          ias = idxas (ia, is)
-          write (fname, '("BAND-QP_S", I2.2, "_A", I4.4, ".OUT")') is, ia
[…]
