+++ src/bandstructure.f90
@@ -1,58 +1,64 @@
-! Copyright (C) 2002-2005 J. K. Dewhurst, S. Sharma and C. Ambrosch-Draxl.
-! This file is distributed under the terms of the GNU General Public License.
-! See the file COPYING for license details.
-!
-!BOP
-! !ROUTINE: bandstr
-! !INTERFACE:
-!
-!
-Subroutine bandstr
-  ! !USES:
-  use modinput
-  use modmain
-  use modmpi
-  use FoX_wxml
+
+!> Module for bandstructure calculations
+  
+module bandstructure
+  use constants, only: zzero, twopi
+  use modinput, only: input
+  use precision, only: dp, i32, sp, str_128, str_256
+
+  implicit none
+
+  private
+
+  public :: bandstr, fourintp
+
+contains
+  !>   Produce a band structure along the path in reciprocal-space which connects
+  !>   the vertices in the array `vvlp1d`. The band structure is obtained from
+  !>   the second-variational eigenvalues and is written to the file `BAND.OUT`
+  !>   with the Fermi energy set to zero. If required, band structures are plotted
+  !>   to files `BAND\_Sss\_Aaaaa.OUT` for atom `aaaa` of species `ss`,
+  !>   which include the band characters for each \(l\) component of that atom in
+  !>   columns 4 onwards. Column 3 contains the sum over \(l\) of the characters.
+  !>   Vertex location lines are written to `BANDLINES.OUT`.  
+  !> REVISION HISTORY
+  !> - Created June 2003 (JKD)
+  !> - Modified June 2012 (DIN)
+  !> - Modified March 2014 (UW)
+  !> - Modified June 2018 (SeTi)
+  !> - Modified April 2025 (Ronaldo)
+subroutine bandstr
+  use mod_APW_LO, only: apwordmax
+  use mod_atoms, only: atposc, idxas, natoms, natmtot, nspecies, spname
+  use mod_Gkvector, only: gkc, ngk, ngkmax, sfacgk, tpgkc
+  use mod_eigensystem, only: mt_hscf, MTInitAll, MTNullify, nmatmax
+  use mod_eigenvalue_occupancy, only: efermi, evalsv, nstfv, nstsv
+  use mod_kpoint, only: nkpt, vkl
+  use mod_muffin_tin, only: idxlm, lmmaxapw
+  use mod_plotting, only: dpp1d, dvp1d, nvp1d
+  use mod_potential_and_density, only: meffig, m2effig, xctype
+  use mod_spin, only: nspinor, nspnfv
+  use modmpi, only: barrier, firstk, lastk, mpi_allgatherv_ifc, mpiglobal, rank, splittfile, terminate
+  use FoX_wxml, only: xmlf_t, xml_AddAttribute, xml_AddCharacters, xml_AddXMLPI, xml_Close, &
+                      xml_EndElement, xml_NewElement, xml_OpenFile
   use m_write_hdf5, only: hdf5_bandstructure_output
-  use precision, only: dp 
 
-
-  ! !DESCRIPTION:
-  !   Produces a band structure along the path in reciprocal-space which connects
-  !   the vertices in the array {\tt vvlp1d}. The band structure is obtained from
-  !   the second-variational eigenvalues and is written to the file {\tt BAND.OUT}
[…]
