### 2025-05-02 e148e1f
Update assertions
