+++ .gitlab/jupyter-exciting-python3-gnu-smp.yml
@@ -90,8 +90,8 @@ test-fedora-buster-gnu-smp-jupyter:
  # - time python3 03_excited_states/03_TDDFT/assertions/many_body_kernels_for_tddft_calculations.py
  # - time jupyter nbconvert --execute 03_excited_states/03_TDDFT/simulating_pump_probe_spectroscopy_with_rt_tddft.ipynb --to markdown
  # - time python3 03_excited_states/03_TDDFT/assertions/simulating_pump_probe_spectroscopy_with_rt_tddft.py
-  - time jupyter nbconvert --execute 03_excited_states/03_TDDFT/studying_higher_harmonics_generation_using_rt_tddft.ipynb --to markdown
-  - time python3 03_excited_states/03_TDDFT/assertions/studying_higher_harmonics_generation_using_rt_tddft.py
+ # - time jupyter nbconvert --execute 03_excited_states/03_TDDFT/studying_higher_harmonics_generation_using_rt_tddft.ipynb --to markdown
+ # - time python3 03_excited_states/03_TDDFT/assertions/studying_higher_harmonics_generation_using_rt_tddft.py
   - time jupyter nbconvert --execute 03_excited_states/03_TDDFT/real_time_tddft_combined_with_molecular_dynamics.ipynb --to markdown
   - time python3 03_excited_states/03_TDDFT/assertions/real_time_tddft_combined_with_molecular_dynamics.py
   - time jupyter nbconvert --execute 03_excited_states/04_Others/graphene_from_ground_state_to_excitations.ipynb --to markdown
