### 2025-05-05 321fd7f
Add support for search_E parameter in excitingtools
