+++ tools/exciting_tools/excitingtools/species/species_file.py
@@ -203,21 +203,22 @@ class SpeciesFile(ECTObject):
 
         return ns_per_l_with_mO_over_1
 
-    def add_number_los_for_all_valence_semicore_states(self, number_lo: int):
+    def add_number_los_for_all_valence_semicore_states(self, number_lo: int, *, search_e: bool = False):
         """Adds a certain number `number_lo` of local orbitals with increasing matching order for every valence and
         semicore state to the basis `self.basis[lo]`. Here the local orbital always consists of 2 wave function
         elements.
 
         :param number_lo: number of local orbitals added for every valence and semicore state
+        :param search_e: if True, the linearization energy is searched for
         """
         valence_semicore_states = self.get_atomicstates_ns_per_l(lambda x: not x["core"])
 
         for l, ns in valence_semicore_states.items():
             for n in ns:
                 for _ in range(number_lo):
-                    self.add_lo_higher_matching_order(l, n)
+                    self.add_lo_higher_matching_order(l, n, search_e=search_e)
 
-    def add_basic_lo_all_semicore_states(self):
+    def add_basic_lo_all_semicore_states(self, *, search_e: bool = False):
         """Adds one local orbital for every semicore state present in the atomic states.
 
         Moving a state from core to valence requires the addition of at least one local orbital to accurately
@@ -225,38 +226,42 @@ class SpeciesFile(ECTObject):
         The local orbital consists of two radial functions at matchingOrder 0:
             - one at the same principal quantum number 'n' of the (L)APW  for the given angular momentum ('l') channel.
             - the other at the principal quantum number 'n' of the semi core state.
+
+        :param search_e: if True, the linearization energy is searched for
         """
         valence_and_semicore_states = self.get_valence_and_semicore_atomicstate_ns_per_l()
         semicore_states = {l: data["semicore"] for l, data in valence_and_semicore_states.items()}
 
         for l, ns in semicore_states.items():
             for n in ns:
-                self.add_lo(l, (n + 1, n), (0, 0))
+                self.add_lo(l, (n + 1, n), (0, 0), search_e=search_e)
 
-    def add_custom_for_all_valence_states(self, custom_type: str):
+    def add_custom_for_all_valence_states(self, custom_type: str, *, search_e: bool = False):
         """Adds a custom element to the basis for every valence state present in the atomic states.
 
         :param custom_type: Type of the custom basis function. It can be either `apw`, `lapw`, or `apw+lo`.
+        :param search_e: if True, the linearization energy is searched for
         """
         valence_and_semicore_states = self.get_valence_and_semicore_atomicstate_ns_per_l()
         valence_states = {l: data["valence"] for l, data in valence_and_semicore_states.items()}
 
         for l, ns in valence_states.items():
             for n in ns:
-                self.basis["custom"].append({"l": l, "type": custom_type, "n": n, "searchE": False})
+                self.basis["custom"].append({"l": l, "type": custom_type, "n": n, "searchE": search_e})
 
-    def add_helos(self, l: int, number: int, skip_lin_dep: bool = True):
+    def add_helos(self, l: int, number: int, skip_lin_dep: bool = True, *, search_e: bool = False):
         """Adds a specific 'number' of High Energy Local Orbitals (HELOs) to the basis
         for a given angular momentum channel 'l'.
 
         :param l: angular momentum number l
         :param number: the number of HELOs to be added to the l-channel
         :param skip_lin_dep: If set to True and matchingOrder > 1 for HELO l and n, skip n+1 HELO
[…]
