+++ src/src_gw/mod_kpointset.f90
@@ -42,9 +42,6 @@ MODULE mod_kpointset
         integer(i32), allocatable :: tnodes(:,:) ! coordinates of tetrahedron
         integer(i32), allocatable :: wtet(:)     ! weight of each tetrahedron
         real(dp) :: tvol                     ! volume of the tetrahedra relative to the BZ volume
-
-        contains
-          procedure, public :: map_to_device => map_to_device_kset, map_delete_device => map_delete_device_kset
     end type k_set
 
 !-------------------------------------------------------------------------------
@@ -78,8 +75,6 @@ MODULE mod_kpointset
           procedure :: change_set_real => Gset_change_set_real
           procedure :: change_set_complex => Gset_change_set_complex
           generic :: change_set => change_set_real, change_set_complex
-          ! This is to map to the device
-          procedure, public :: map_to_device => map_to_device_Gset, map_delete_device => map_delete_device_Gset
     end type G_set
 
 !-------------------------------------------------------------------------------
@@ -109,8 +104,6 @@ MODULE mod_kpointset
         real(dp), allocatable :: gknrc(:,:,:)    ! length of G+k-vectors
         real(dp), allocatable :: tpgknrc(:,:,:,:)! (theta, phi) coordinates of G+k-vectors
         complex(dp), allocatable :: sfacgknr(:,:,:,:) ! structure factor for the G+k-vectors
-        contains
-          procedure, public :: map_to_device => map_to_device_Gkset, map_delete_device => map_delete_device_Gkset
     end type Gk_set
 
 !-------------------------------------------------------------------------------
@@ -137,8 +130,6 @@ MODULE mod_kpointset
         integer, allocatable :: ikp2ikq(:,:) ! map the irreducible k-point index to the corresponding from the non-reduced set
         integer, allocatable :: nsymkstar(:,:)   ! number of symmetry operations which form the star
         integer, allocatable :: isymkstar(:,:,:) ! index of the symmetry operations
-        contains
-          procedure, public :: map_to_device => map_to_device_kqset, map_delete_device => map_delete_device_kqset
     end type kq_set
 
 !-------------------------------------------------------------------------------
@@ -2571,449 +2562,5 @@ CONTAINS
         end function i1dnr
 
     end subroutine ikpik2iqivgnr
-
-    subroutine map_to_device_kset(this)
-
-      use mod_device_offload, only: device_world
-      use iso_c_binding,   only: c_loc, c_intptr_t
-
-      class(k_set), target, intent(inout) :: this
-      integer(c_intptr_t)                 :: address
-      character(str_256)                  :: address_tag
-
-      address = int(loc(this), kind = c_intptr_t)
-      write(address_tag, *) address
-
-      call device_world%register%alloc(trim(adjustl(address_tag))//"%nkptnr"     , bytes_size(this%nkptnr     ), device_world%get_device())
-      call device_world%register%alloc(trim(adjustl(address_tag))//"%nkpt"       , bytes_size(this%nkpt       ), device_world%get_device())
-      call device_world%register%alloc(trim(adjustl(address_tag))//"%isreduced"  , bytes_size(this%isreduced  ), device_world%get_device())
-      call device_world%register%alloc(trim(adjustl(address_tag))//"%usedlibzint", bytes_size(this%usedlibzint), device_world%get_device())
[…]
