+++ src/src_inputparser/groundstatetasklauncher.F90
@@ -8,18 +8,14 @@ Subroutine groundstatetasklauncher
     Use modinput
     Use modmain,     only: task, xctype
     Use inputdom
-    use modmpi
+    use modmpi, only: terminate_if_false, splittfile
     use mod_misc, only: filext
 
     Implicit None
 
     character(len=:), allocatable :: string
     logical :: is_cdft_calculation, skip_gnd_in_cdft_calculation
-
-    ! Interface to input elements defined for a constrained DFT calculation
-    is_cdft_calculation = associated(input%groundstate%constrainedDFT)
-    skip_gnd_in_cdft_calculation = .false.
-    if ( is_cdft_calculation ) skip_gnd_in_cdft_calculation = input%groundstate%constrainedDFT%skipgnd
+    logical :: is_hybrid_calculation, is_exchange_hartree_fock
 
     call delete_warnings
     splittfile= .true.
@@ -32,7 +28,7 @@ Subroutine groundstatetasklauncher
         input%groundstate%Hybrid => getstructHybrid(emptynode)
     End If
     ! EXX-OEP-based hybrid functionals
-    If (associated(input%groundstate%Hybrid)) Then
+    If (is_hybrid_calculation) Then
         If (input%groundstate%Hybrid%exchangetypenumber .Eq. 2) Then
             If (.not.associated(input%groundstate%OEP)) Then
                input%groundstate%OEP => getstructOEP (emptynode)
@@ -44,6 +40,17 @@ Subroutine groundstatetasklauncher
            input%groundstate%OEP => getstructOEP(emptynode)
         End If
     End If
+
+    ! Interface to input elements defined for a hybrid calculation
+    is_hybrid_calculation = associated(input%groundstate%Hybrid)
+    is_exchange_hartree_fock = .false.
+    if ( is_hybrid_calculation ) is_exchange_hartree_fock = ( input%groundstate%Hybrid%exchangetypenumber == 1 )
+
+    ! Interface to input elements defined for a constrained DFT calculation
+    is_cdft_calculation = associated(input%groundstate%constrainedDFT)
+    skip_gnd_in_cdft_calculation = .false.
+    if ( is_cdft_calculation ) skip_gnd_in_cdft_calculation = input%groundstate%constrainedDFT%skipgnd
+
     If (input%groundstate%do .Eq. "fromscratch") Then
         If (associated(input%relax)) Then
             task = 2
@@ -65,25 +72,29 @@ Subroutine groundstatetasklauncher
         If  (associated(input%groundstate%HartreeFock)) Then
             task = 5
             Call hartfock
-        ! DFT / OEP
-        Else If (associated(input%groundstate%Hybrid)) Then
-                If (input%groundstate%Hybrid%exchangetypenumber == 1) Then
-                    Call hybrids
-                Else
-                    Call gndstate
-                End If
+
         ! Constrained DFT calculation
         Else If ( is_cdft_calculation ) Then
             string = filext
+            if ( is_hybrid_calculation ) then
+                call terminate_if_false(  is_exchange_hartree_fock , " ERROR: The combination of OEP and CDFT is not implemented yet." )
+            end if
             if ( .not. skip_gnd_in_cdft_calculation ) then
                 filext = file_extension_GS
-                call gndstate
+                call hybrid_or_gndstate( is_hybrid_calculation .and. is_exchange_hartree_fock )
+            end if
+            if (task==7) then  ! for hybrid calculation
+                task=0
             end if
             filext = file_extension_CDFT
             call set_status_to_running_CDFT
-            call gndstate
+            call hybrid_or_gndstate( is_hybrid_calculation .and. is_exchange_hartree_fock )
             call set_status_to_finished_CDFT
[…]
