+++ src/src_hybrids/calc_vxnl.f90
@@ -18,12 +18,15 @@ subroutine calc_vxnl()
     use mod_APW_LO, only: apwordmax
     use mod_muffin_tin, only: lmmaxapw
     use mod_eigensystem, only: nmatmax
-    use mod_eigenvalue_occupancy, only: efermi, nstfv
+    use mod_eigenvalue_occupancy, only: efermi, nstfv, occsv
     use mod_coulomb_potential, only: delete_coulomb_potential
     use modmpi
     use mod_bands, only: evalfv, nomax, numin, ikvbm, ikcbm, ikvcm, eveck, eveckalm, eveckp, eveckpalm
+    use cdft, only: cdft_input_keys
+    use general_find_vbm_cbm, only: find_vbm_cbm
     use precision, only: i32, dp
 #include "offload.fpp"
+
 !
 ! !DESCRIPTION:
 !   Calculates the non-local exchange potential
@@ -33,7 +36,7 @@ subroutine calc_vxnl()
 !BOC
     implicit none
 
-    integer(i32) :: ikp, ik, jk, iq, ikq
+    integer(i32) :: ikp, ik, jk, iq, ikq, jkp
     integer(i32) :: ie12, ie12tot, ie1, ie2, ie3, icg
     integer(i32) :: ist, l, im
     integer(i32) :: i, j, k, jst, ispn
@@ -47,6 +50,8 @@ subroutine calc_vxnl()
     integer(i32), allocatable :: idxpair(:,:)
     complex(dp), allocatable :: minm(:,:,:)
     complex(dp), allocatable :: evecsv(:,:)
+    type(cdft_input_keys) :: cdft_calculation
+    real(dp), parameter :: tolerance = 1.0e-8_dp
 
     call cpu_time(tstart)
 
@@ -60,9 +65,13 @@ subroutine calc_vxnl()
       call getevalfv(kset%vkl(:,ik), evalfv(:,ik))
     end do
 
-    ! VB / CB state index
-    call find_vbm_cbm(1, nstfv, kset%nkpt, evalfv, efermi, nomax, numin, ikvbm, ikcbm, ikvcm)
-    ! write(*,*) 'calc_vxnl: ', nomax, numin, efermi
+    call cdft_calculation%read_input_keys( input%groundstate )
+    if (cdft_calculation%is_on()) then
+      call find_vbm_cbm(1, nstfv, kset%nkpt, occsv, evalfv, nomax, numin, ikvbm, ikcbm, ikvcm)
+    else
+      ! VB / CB state index in ground state
+      call find_vbm_cbm(1, nstfv, kset%nkpt, evalfv, efermi, nomax, numin, ikvbm, ikcbm, ikvcm)
+    end if
 
     ! BZ integration weights
     call kintw()
@@ -128,6 +137,7 @@ subroutine calc_vxnl()
         Gamma = gammapoint(kqset%vqc(:,iq))
         ik  = kset%ikp2ik(ikp)
         jk  = kqset%kqid(ik,iq)
+        jkp = kset%ik2ikp(jk)
 
         !=======================================
         ! distribute (k,q)-pair over processors
@@ -204,7 +214,12 @@ subroutine calc_vxnl()
                   ie1 = idxpair(1,ie12)
                   ie2 = idxpair(2,ie12)
                   mvm = dot_product(minm(1:mbsiz,ie1,ie3), minm(1:mbsiz,ie2,ie3))
-                  vxnl(ie1,ie2,ikp) = vxnl(ie1,ie2,ikp) - kiw(ie3,jk)*mvm
+                  if (cdft_calculation%is_on()) then
[…]
