+++ src/scf_cycle.f90
@@ -5,50 +5,80 @@
 !
 subroutine scf_cycle(verbosity)
 ! !USES:
-    use modinput
-    use modmain
-    use modmpi
-    use scl_xml_out_Module
-    use TS_vdW_module, only: C6ab, R0_eff_ab
-    use sirius_init,   only: sirius_options
+    use cdft, only: cdft_input_keys, deallocate_cdft_global_arrays, ExcitonCoefficients, &
+      file_extension_GS, initialize_cdft_global_arrays, occupy_cdft, update_occupations_with_the_maximum_overlap_method
+    use exciting_mpi, only: xmpi_bcast, xmpi_allreduce
+    use lo_recommendation, only: recommend_local_orbital_trial_energies
+    use mod_APW_LO, only: apwn, apwe0, lorbe0, lorbl, lorbord, lorbn, maxapword, maxlapw, nlorb
+    use mod_atoms, only: atposc, idxas, natoms, natmtot, nspecies, spr, spsymb
+    use mod_charge_and_moment, only: chgdst, chgpart, momtot
+    use mod_convergence, only: currentconvergence, vchgdst, vcurrentconvergence, vdeltae
+    use mod_eigenvalue_occupancy, only: evalsv, fermidos, occsv, nstfv, nstsv
+    use mod_eigensystem, only: mt_hscf, MTInitAll, MTNullify, nmatmax
+    use mod_energy, only: engytot, engyknst
+    use mod_force, only: forcemax, forcetot
+    use mod_Gvector, only: ngrid, ngrtot
+    use mod_Gkvector, only: vgkl
+    use mod_kpoint, only: nkpt, vkl, wkpt
+    use mod_LDA_LU, only: ldapu, lmmaxlu
+    use mod_misc, only: filext, task, tlast, tstop
+    use mod_muffin_tin, only: lmmaxvr, nrmt, nrmtmax
+    use mod_OEP_HF, only: resoep
+    use mod_potential_and_density, only: generate_density_and_magnetization, m2effig, magir, magmt, meffig, rhomt, rhoir, veffmt
+    use mod_spin, only: ndmag, nspinor, nspnfv
+    use mod_timing, only: stopwatch, time_density_init, time_pot_init, timefor, timefv, &
+      timeinit, timeio, timemat, timemixer, timemt, timepot, timerho, timesv
+    use modinput, only: input, getfixspinnumber
+    use modmpi, only: barrier, firstofset, ierr, lastofset, mpiglobal, mpi_allgatherv_ifc, &
+      procs, rank, splittfile
+    use precision, only: dp, i32
+    use scl_xml_out_Module, only: deltae, dforcemax, iscl, scl_iter_xmlout, scl_xml_out_write, scl_xml_write_moments
+    use secular_equation, only: seceqn
     use sirius_api,    only: set_radial_functions_sirius, solve_seceqn_sirius, get_eval_sirius, get_evec_sirius, &
                              put_occ_sirius, generate_density_sirius, get_periodic_function_sirius
-    use mod_potential_and_density, only: generate_density_and_magnetization
-    use cdft, only: cdft_input_keys, ExcitonCoefficients, file_extension_GS, occupy_cdft
-    use precision, only: dp
-    use lo_recommendation, only: recommend_local_orbital_trial_energies
-    use trial_energy_selection, only: select_apw_trial_energies, select_local_orbital_trial_energies
+    use sirius_init,   only: sirius_options
     use total_energy, only: energy
+    use trial_energy_selection, only: select_apw_trial_energies, select_local_orbital_trial_energies
+    use TS_vdW_module, only: C6ab, R0_eff_ab
+    
+    
 !
 
 ! !DESCRIPTION:
 !
[…]
