### 2025-05-09 1d81b70
Update moke tutorial name
