+++ .gitlab/jupyter-exciting-python3-gnu-smp.yml
@@ -96,7 +96,7 @@ test-fedora-buster-gnu-smp-jupyter:
   - time python3 03_excited_states/03_TDDFT/assertions/real_time_tddft_combined_with_molecular_dynamics.py
   - time jupyter nbconvert --execute 03_excited_states/04_Others/graphene_from_ground_state_to_excitations.ipynb --to markdown
   - time python3 03_excited_states/04_Others/assertions/graphene_from_ground_state_to_excitations.py
-  - time jupyter nbconvert --execute 04_additional_features/magneto_optical_kerr_effect_(moke).ipynb --to markdown
+  - time jupyter nbconvert --execute 04_additional_features/magneto_optical_kerr_effect.ipynb --to markdown
   - time python3 04_additional_features/assertions/magneto_optical_kerr_effect.py
   - time jupyter nbconvert --execute 04_additional_features/second_harmonic_generation.ipynb --to markdown
   - time python3 04_additional_features/assertions/second_harmonic_generation.py
