### 2025-05-10 dd9753b
Update shg assertions
