### 2025-05-11 3f68fee
Update shg assertions
