### 2025-05-12 e69ca89
Update shg assertions
