### 2025-05-12 cda0210
RTTDDFT: length gauge
