+++ src/src_xs/src_rttddft/rttddft_hamiltonianoverlap.f90
@@ -22,11 +22,9 @@ module rttddft_HamiltonianOverlap
   use mod_muffin_tin, only: idxlm, rmt
   use mod_potential_and_density, only: veffig, meffig, m2effig, veffmt
   use modinput, only: input
-  use modmpi
   use physical_constants, only: alpha, c
   use precision, only: dp, i32
   use rttddft_GlobalMDVariables, only: mathcalH, mathcalB
-  use rttddft_pmat, only: obtain_pmat_LAPWLOBasis
   use rttddft_timings, only: Print_Timings, Timing_RTTDDFT_hamiltonian, &
     Timing_Ehrenfest, timesec_RTTDDFT
   use rttddft_VectorPotential, only: Vector_Potential_Field
@@ -37,8 +35,9 @@ module rttddft_HamiltonianOverlap
 
   private
   public :: update_hamiltonian_without_pa_term_lapw, update_overlap_lapw, &
-    update_hamiltonian_without_pa_term_ks, add_external_coupling_vgauge
-
+    add_external_coupling_length_gauge, update_hamiltonian_without_pa_term_ks, &
+    add_external_coupling_vgauge
+    
   real(dp) :: fact, atot(3)
   type(MTHamiltonianList) :: mt_h
 
@@ -47,7 +46,7 @@ contains
   !> In `update_overlap_lapw`, we obtain the overlap of the (L)APWs at time \( t \).
   subroutine update_overlap_lapw( first_kpt, a_tot, overlap, apwalm, pmatmt, printTimings, &
     t_ham, t_MD, update_mathcalH, update_mathcalB )
-    !> The first k point
+    !> The first \( \mathbf{k} \) point
     integer(i32), intent(in) :: first_kpt
     !> Total vector potential
     type(Vector_Potential_Field), intent(in) :: a_tot
@@ -70,7 +69,7 @@ contains
     !> if `.True.`, update `mathcalB`
     logical, intent(in), optional :: update_mathcalB
 
-    integer(i32) :: ik, nmatp, last_kpt
+    integer(i32) :: ik, last_kpt
     real(dp) :: ti, tf
     logical :: timings_general, timings_detailed, get_mathcalH, get_mathcalB
 
@@ -97,23 +96,22 @@ contains
 
     if( timings_general ) call timesec( ti ) 
 
-    !$OMP PARALLEL DEFAULT(NONE), PRIVATE(ik, nmatp), &
-    !$OMP& SHARED(first_kpt, last_kpt, apwalm, pmatmt), &
-    !$OMP& SHARED(overlap, nmat, get_mathcalH, get_mathcalB)
-    !$OMP DO
+    !$omp parallel default(none), private(ik), &
+    !$omp& shared(first_kpt, last_kpt, apwalm, pmatmt), &
+    !$omp& shared(overlap, nmat, get_mathcalH, get_mathcalB)
+    !$omp do
     do ik = first_kpt, last_kpt
-      nmatp = nmat(1, ik)
       
       if ( get_mathcalB .or. get_mathcalH ) then
         call overlapsetup( ik, overlap(:, :, ik), apwalm(:, :, :, :, ik), &
-          nmatp, get_mathcalB, get_mathcalH, pmatmt(:, :, :, :, ik) )
+          nmat(1, ik), get_mathcalB, get_mathcalH, pmatmt(:, :, :, :, ik) )
       else
         call overlapsetup( ik, overlap(:, :, ik), apwalm(:, :, :, :, ik), &
-          nmatp, get_mathcalB, get_mathcalH )
+          nmat(1, ik), get_mathcalB, get_mathcalH )
       end if
 
     end do
-    !$OMP END PARALLEL
[…]
