+++ test/tests_config.yml
@@ -780,8 +780,9 @@ RT-TDDFT/GGA_PBE-BN-kick-EH-restart:
       - "PVEC.OUT"
       - "AVEC.OUT"
       - "rt.h5"
+      - "ELECTRIC_FIELD.OUT"
 
-RT-TDDFT/LDA_PW-SiC-trapCos-AETRS-uKS-frozen-restart:
+RT-TDDFT/LDA_PW-SiC-trapCos-AETRS-KS-frozen-restart-length-gauge:
     files_under_test:
       - "EIGVAL_40.OUT"
       - "EIGVAL_60.OUT"
@@ -803,6 +804,8 @@ RT-TDDFT/LDA_PW-SiC-trapCos-AETRS-uKS-frozen-restart:
       - "ETOT_RTTDDFT.OUT.SAVE"
       - "EVECFV_RTTDDFT.OUT"
       - "EVECFV_PREVIOUS_RTTDDFT.OUT"
+      - "ELECTRIC_FIELD.OUT.SAVE"
+      - "PHASES_RTTDDFT.OUT"
   
 RT-TDDFT/GGA_PBE-C-trapCos-EHM-predictorCorrector:
     files_under_test:
@@ -902,8 +905,9 @@ RT-TDDFT/LDA_PW-Ehrenfest-BN-restart:
       - "EVECFV_RTTDDFT.OUT"
       - "RHO_VKS_RTTDDFT.OUT"
       - "TRAJECTORY_RTTDDFT.OUT"
+      - "ELECTRIC_FIELD.OUT.SAVE"
 
-RT-TDDFT/LDA_PW-Si-sinSq-CFM4:
+RT-TDDFT/LDA_PW-Si-sinSq-CFM4-length-gauge:
    files_under_test:
       - "atoms.xml"
       - "eigval.xml"

[…]
