### 2025-05-13 7c59968
Update assertions
