### 2025-05-14 0ce3fdb
Compiling exciting using an apptainer container
