### 2025-05-14 90aedfb
Symmetry Option for Phonon Interpolation. Phonons can be returned now on a non reduced q-grid if wished.
