+++ tools/exciting_tools/excitingtools/utils/valid_attributes.py
@@ -670,7 +670,7 @@ realTimeTDDFT_valid_subtrees = ["predictorCorrector", "screenshots", "laser", "p
 predictorCorrector_attribute_types = {"maxIterations": (int, 1), "tol": (float, 1)} 
 
 screenshots_attribute_types = {"niter": (int, 1)} 
-screenshots_valid_subtrees = ["eigenvalues", "projectionCoefficients", "occupations", "density"] 
+screenshots_valid_subtrees = ["eigenvalues", "projectionCoefficients", "occupations", "deltadensityplot"] 
 
 eigenvalues_attribute_types = {"nEigenvalues": (int, 1), "tolerance": (float, 1)} 
 
@@ -678,7 +678,7 @@ projectionCoefficients_attribute_types = {"format": (str, 1), "printAbsoluteValu
 
 occupations_attribute_types = {"binaryFormat": (bool, 1), "format": (str, 1), "textFormat": (bool, 1)} 
 
-density_valid_subtrees = ["plot3d"] 
+deltadensityplot_valid_subtrees = ["plot3d"] 
 
 laser_attribute_types = {"fieldType": (str, ["external", "total"])} 
 laser_valid_subtrees = ["kick", "trapCos", "sinSq"] 
