### 2025-05-21 #7809 @zhubonan
Hi @maurerben could you please suggest some reviewers?

If my understanding is correct the low-scaling BSE code has been merged into the neon branch of the exciting code base. Could you please provide some instructions for testing particularly this functionality/part of the code? The code base of exiting is quite large. This would help the reviewing for this submission.
