### 2025-05-22 998ebe7
Comment out hybrid functional assertions
