### 2025-05-26 29db902
Uncomment rule to only execute test when merging to development
