+++ test/src/tolerance/templates/rt_tddft.py
@@ -9,12 +9,12 @@ Keys were generated by using the tolerance.py script. See that for command line
 details.
 
 Tolerances in default_tolerances cover files
-  * JIND.OUT
-  * NEXC.OUT
-  * ETOT_RTTDDFT.OUT
+  * CURRENT.OUT
+  * N_EXCITATIONS.OUT
+  * TOTENERGY_RTTDDFT.OUT
   * EIGVAL_*.OUT
   * OCCSV_TXT_*.OUT
-  * PROJ_*.OUT
+  * PROJECTION_COEFFS_*.OUT
   * ATOM_????.OUT
   * FCR_????.OUT
   * FEXT_????.OUT
@@ -99,12 +99,12 @@ rt_tddft_force = {
     'Fz': default.force,
 }
 
-rt_tddft_tolerances = {'JIND.OUT': rt_tddft_jind,
-                       'NEXC.OUT': rt_tddft_nexc,
-                       'ETOT_RTTDDFT.OUT': rt_tddft_etot,
+rt_tddft_tolerances = {'CURRENT.OUT': rt_tddft_jind,
+                       'N_EXCITATIONS.OUT': rt_tddft_nexc,
+                       'TOTENERGY_RTTDDFT.OUT': rt_tddft_etot,
                        'EIGVAL_??.OUT': rt_tddft_eigval,
                        'OCCSV_TXT_??.OUT': rt_tddft_occupations,
-                       'PROJ_??.OUT': rt_tddft_proj,
+                       'PROJECTION_COEFFS_??.OUT': rt_tddft_proj,
                        'ATOM_????.OUT': rt_tddft_atom,
                        'FCR': rt_tddft_force,
                        'FEXT': rt_tddft_force,
