+++ tools/exciting_tools/excitingtools/exciting_dict_parsers/parser_factory.py
@@ -94,11 +94,11 @@ _file_to_parser = {
     "SIGMAC_K*": gw_taskgroup_parser.parse_sigmac,
     "SIGMAX_K*": gw_taskgroup_parser.parse_sigmax,
     "POLARIZABILITY_FACTOR_Q*": gw_taskgroup_parser.parse_polarizability_factor,
-    "JIND.OUT": RT_TDDFT_parser.parse_jind,
-    "NEXC.OUT": RT_TDDFT_parser.parse_nexc,
-    "ETOT_RTTDDFT.OUT": RT_TDDFT_parser.parse_etot,
+    "CURRENT.OUT": RT_TDDFT_parser.parse_jind,
+    "N_EXCITATIONS.OUT": RT_TDDFT_parser.parse_nexc,
+    "TOTENERGY_RTTDDFT.OUT": RT_TDDFT_parser.parse_etot,
     "EIGVAL_*": RT_TDDFT_parser.parse_eigval_screenshots,
-    "PROJ_*": RT_TDDFT_parser.parse_proj_screenshots,
+    "PROJECTION_COEFFS_*": RT_TDDFT_parser.parse_proj_screenshots,
     "OCCSV_TXT_*.OUT": RT_TDDFT_parser.parse_occupations,
     "ATOM_*": RT_TDDFT_parser.parse_atom_position_velocity_force,
     "FCR_*": RT_TDDFT_parser.parse_force,
