### 2025-05-28 8494649
Update test suite defaults and documentation
