+++ test/src/exciting_settings/constants.py
@@ -39,7 +39,7 @@ settings = Defaults(max_time       = 1800,
                     input_file     = 'input.xml',
                     run_dir        = 'run',
                     ref_dir        = 'ref',
-                    exe_dir        =  os.path.normpath(os.path.join(os.getcwd(), '../bin')),
+                    exe_dir        =  os.path.normpath(os.path.join(os.getcwd(), '../install/bin')),
                     ignored_output = ['STATE.OUT', 'OCC', 'EVEC', 'EVALSV', 'EVALFV', 'APWCMT', 'SYM',
                                       'PMAT', 'FERMISURF', 'RMSDVEFF', 'LOCMT', 'EXCLI', 'SCCLI'],
                     binary_names    = _binary_names,
