+++ .gitlab/exciting-fedora-buster-gnu-serial.yml
@@ -33,4 +33,4 @@ test-fedora-buster-gnu-serial:
     # Install test framework parsers
      - pip3 install -e tools/exciting_tools
     # Run exciting app tests
-     - cd test/ && python3 runtest.py -r $PWD/../install/bin/ -e exciting_serial -repeat-tests 2
+     - cd test/ && python3 runtest.py -bp $PWD/../install/bin/ -e exciting_serial -repeat-tests 2
