+++ .gitlab/exciting-fedora-buster-intel2025-serial.yml
@@ -36,4 +36,4 @@ test-fedora-buster-Intel2025-serial:
     # Install test framework parsers
      - pip3 install -e tools/exciting_tools
     # Run exciting app tests
-     - cd test/ && python3 runtest.py -r $PWD/../install/bin/ -e exciting_serial -repeat-tests 2
+     - cd test/ && python3 runtest.py -bp $PWD/../install/bin/ -e exciting_serial -repeat-tests 2
