+++ .gitlab/exciting-fedora-buster-gnu-smp.yml
@@ -33,4 +33,4 @@ test-fedora-buster-gnu-smp:
     # Install test framework parsers
      - pip3 install -e tools/exciting_tools
     # Run exciting app tests
-     - cd test/ && export OMP_NUM_THREADS=2 && export OPENBLAS_NUM_THREADS=1 && python3 runtest.py -r $PWD/../install/bin/ -e exciting_smp -repeat-tests 2
+     - cd test/ && export OMP_NUM_THREADS=2 && export OPENBLAS_NUM_THREADS=1 && python3 runtest.py -bp $PWD/../install/bin/ -e exciting_smp -repeat-tests 2
