### 2025-05-30 f8b19a9
RT-TDDFT: GS potential and density
