+++ tools/exciting_tools/excitingtools/utils/valid_attributes.py
@@ -661,7 +661,7 @@ realTimeTDDFT_attribute_types = {"TaylorOrder": (int, 1),
                                  "do": (str, ["fromfile", "fromscratch"]),
                                  "eeInteraction": (str, ["IPA", "aDFT"]),
                                  "endTime": (float, 1),
-                                 "gauge": (str, ["length", "velocity"]),
+                                 "fieldCoupling": (str, ["berryPhase", "velocityGauge"]),
                                  "nEigenvectorsEH": (int, 1),
                                  "normalizeWF": (bool, 1),
                                  "numberOfFrozenStates": (int, 1),
