+++ src/src_xs/src_rttddft/rttddft_init.f90
@@ -41,7 +41,7 @@ module rttddft_init
   use muffin_tin_basis, only: mt_basis_type
   use precision, only: dp, i32, str_128
   use propagators, only: propagator_type => propagator, create_propagator
-  use rttddft_berry, only: get_td_overlap_det_and_length_gauge_term
+  use rttddft_berry, only: get_td_overlap_det_and_berry_coupling_term
   use rttddft_CurrentDensity, only: Current_Density, Current_Density_Field
   use rttddft_Density, only: update_density, save_and_frozen, frozen, ground_state
   use rttddft_electric_field, only: Electric_Field
@@ -49,7 +49,7 @@ module rttddft_init
   use rttddft_GlobalMDVariables, only: B_past, B_time, mathcalH, mathcalB
   use rttddft_HamiltonianOverlap, only: update_hamiltonian_without_pa_term_lapw, &
     update_overlap_lapw, update_hamiltonian_without_pa_term_ks, &
-    add_external_coupling_vgauge, add_external_coupling_length_gauge
+    add_external_coupling_velocity_gauge, add_external_coupling_berry_phase
   use rttddft_hybrids, only: hybrids_used, Set_Dimension_mixed_product_basis, set_barecoul_basis
   use rttddft_input, only: rttddft_input_keys
   use rttddft_io, only: file_pmat_exists, read_pmat, write_pmat, file_pmat_mt_exists, &
@@ -75,8 +75,8 @@ contains
 subroutine initialize_rttddft( rt_inp, propagator, vec_pot, a_tot_t_minus_dt, molecular_dynamics, &
     psi, overlap, ham_init, ham_time, ham_past, effective_potential_init, apwalm, &
     pmat, pmatmt, rhomt_frozen, rhoir_frozen, occupations, k_dependent_dims, &
-    occs_tol, kset_rttddft, Gkset, Gset, psi_gnd_lapwlo, pws_for_length_gauge, k_ptrs, &
-    td_overlap_det, length_gauge_term, prev_phases, e_vec, e_vec_save, j_para_spurious, p_vec_init )
+    occs_tol, kset_rttddft, Gkset, Gset, psi_gnd_lapwlo, pws_for_berry_phase, k_ptrs, &
+    td_overlap_det, berry_coupling_term, prev_phases, e_vec, e_vec_save, j_para_spurious, p_vec_init )
   !> Argument that encapsulates the input options of rttddft
   type(rttddft_input_keys), intent(in) :: rt_inp
   !> Argument that encapsulates the propagator
@@ -124,14 +124,14 @@ subroutine initialize_rttddft( rt_inp, propagator, vec_pot, a_tot_t_minus_dt, mo
   !> KS-LAPW+lo transition matrix (ground state set in the LAPW+lo basis) (to be allocated in `array_allocation` block)
   complex(dp), allocatable, intent(out) :: psi_gnd_lapwlo(:, :, :)
   !> Planewave matrix elements between neighbouring \( \mathbf{k} \) points (to be allocated in `array_allocation` block)
-  complex(dp), allocatable, intent(out) :: pws_for_length_gauge(:, :, :, :, :)
+  complex(dp), allocatable, intent(out) :: pws_for_berry_phase(:, :, :, :, :)
   !> Array containing indices of the neighbouring \( \mathbf{k} \) points (to be allocated in `array_allocation` block)
   integer(i32), allocatable, intent(out) :: k_ptrs(:, :, :)
[…]
