### 2025-06-02 679ce3b
RT: non-global KS energies
