+++ tools/exciting_tools/excitingtools/utils/valid_attributes.py
@@ -659,7 +659,7 @@ realTimeTDDFT_attribute_types = {"TaylorOrder": (int, 1),
                                  "calculateNExcitedElectrons": (bool, 1),
                                  "calculateTotalEnergy": (bool, 1),
                                  "do": (str, ["fromfile", "fromscratch"]),
-                                 "eeInteraction": (str, ["IPA", "aDFT"]),
+                                 "eeInteraction": (str, ["IPA", "aDFT", "tdH"]),
                                  "endTime": (float, 1),
                                  "fieldCoupling": (str, ["berryPhase", "velocityGauge"]),
                                  "nEigenvectorsEH": (int, 1),
