+++ src/src_gw/task_eps_r.f90
@@ -8,6 +8,7 @@ subroutine task_eps_r
     use invert_dielectric_function, only: calcinveps
     use modxs, only: symt2
     use mod_mpi_gw
+    use modmpi, only: rank, mpiglobal, distribute_loop
     use m_getunit
     use mod_hdf5
     use mod_rpath
@@ -98,20 +99,11 @@ subroutine task_eps_r
     ! Main loop: BZ integration
     !===========================================================================    
 
-#ifdef MPI
-    call set_mpi_group(kqset%nkpt)
-    call mpi_set_range(nproc_row, &
-    &                  myrank_row, &
-    &                  kset%nkpt, 1, &
-    &                  iqstart, iqend)
-#else
-    iqstart = 1
-    iqend = kqset%nkpt
-#endif
+    call distribute_loop( mpiglobal, kqset%nkpt, iqstart, iqend )
     iomstart = 1
     iomend = freq%nomeg
     
-    if (myrank==0) then
+    if (rank==0) then
       call boxmsg(fgw,'=','q-point cycle')
       call flushifc(fgw)
     end if
@@ -213,7 +205,7 @@ subroutine task_eps_r
     deallocate(tvec,wfmb)
 
 #ifdef MPI
-    call mpi_sum_array(0,eps_r,npt,mycomm_row)
+    call mpi_sum_array( eps_r, mpiglobal, .false. )
 #endif
 
     if (rank==0) then
