### 2025-06-05 a122332
introduce new generic expand_products subroutine
