+++ src/src_gw/expand_products.f90
@@ -1,411 +1,0 @@
-subroutine expand_products(ik,iq,nstart,nend,nsplit,mstart,mend,msplit,minm)
-
-    use modmain, only : zzero
-    use modgw,   only : mbsiz
-    use precision, only: i32, dp
-    use modgw,      only: kqset
-    use iso_c_binding,         only: c_ptr, c_loc, c_f_pointer, c_sizeof, c_size_t
-    use mod_pointer_remapping, only: remap_fortran_pointer
-    use mod_device_offload,    only: device_world
-    use device_linalg_common_interface, only: zgemm_gpu
-    use m_memory_device,       only: allocate_device_memory, deallocate_device_memory, &
-                                     bytes_double_complex, bytes_int, get_device_pointer
-#include "offload.fpp"
-    implicit none
-
-    ! input variables
-    integer(i32), intent(in) :: ik
-    integer(i32), intent(in) :: iq
-    integer(i32), intent(in) :: nstart, nend
-    integer(i32), intent(in) :: nsplit         ! last index of valence states
-    integer(i32), intent(in) :: mstart ,mend
-    integer(i32), intent(in) :: msplit         ! last index of valence states
-    complex(dp),  intent(inout) :: minm(mbsiz,nstart:nend,mstart:mend)
-
-    ! local variables
-    integer(i32) :: cstart, cend
-    integer(i32) :: im, ie1, ie2
-    integer(i32) :: iflag
-    integer(i32) :: my_device
-    integer(c_size_t) :: dim1, dim2, dim3
-    type(c_ptr)  :: minm_valence_cptr
-    complex(dp), pointer, contiguous :: minm_valence(:,:,:)
-    type(c_ptr)  :: minm_core_cptr
-    complex(dp), pointer, contiguous :: minm_core(:,:,:)
-
-    dim1 = mbsiz
-    my_device = device_world%get_device()
-
-    OMP_OFFLOAD target
-    ! IFX 2025.0.0 is not able to perform it efficiently except if a kernel is used
-    ! Cray compiler generates a runtime error if using the kernel for the loop.
-    ! GNU allows the latter so we go for it.
-    ! For core states, we keep the Cray way for all cases, as the number
-    ! of core states is always small in comparison with valence states
-    ! and I want to keep compiler branching as small as possible.
-    !
-    ! TODO: Check future IFX compiler to see if this is improved.
-    !
-#if __INTEL_COMPILER
-    !$omp teams distribute parallel do collapse(3) private(im,ie1,ie2)
-    do ie2 = mstart, mend
-      do ie1 = nstart, nend
-        do im = 1, mbsiz
-          minm(im,ie1,ie2) = zzero
-        end do
-      end do
-    end do
-    !$omp end teams distribute parallel do
-#else
-    minm(:,:,:) = zzero
-#endif
-    OMP_OFFLOAD end target
-
-    if ((nsplit>0).and.(msplit<=0)) then
-
-        !======================================================================
-        ! When calculating the dielectric function:
-        !
-        ! n -> {1:nomax}{1:ncg}, so nsplit=nomax, when core states are treated
-        ! m -> {numin:nstfv},       msplit is ignored
-        !
-        ! Important to specify msplit<=0
-        ! to ignore the self energy related part (below)
-        !
-        !======================================================================
-        if (nend<=nsplit) then
-
-            ! calculate M^i_{nm}
-            iflag = 1
-            dim2  = nend - nstart + 1
-            dim3  = mend   - mstart + 1
-            call allocate_device_memory(minm_valence_cptr, dim1*dim2*dim3*bytes_double_complex, my_device)
[…]
