### 2025-06-06 2abed85
RT: Warnings generation
