### 2025-06-10 bef8647
Kinetic Energy Density
