+++ src/kinetic_energy_density/kinetic_energy_density_cr.f90
@@ -1,0 +1,495 @@
+!> Module provides subroutines to calculate the kinetic energy density
+!> in the muffin-tin region for the core electrons `ked_cr`. 
+module kinetic_energy_density_cr
+    use asserts, only: assert
+    use precision, only: dp
+    use errors_warnings, only: terminate_if_true, terminate_if_false
+    use modmpi, only: mpiglobal
+    use modinput, only: input
+    use mod_muffin_tin, only: nrmt, nrmtmax, idxlm
+    use mod_atoms, only: natoms, idxas, nspecies, &
+        & spnst, spcore, spr, spk, spocc, spl, natmtot, spnstmax
+    use mod_SHT, only: zbshtvr
+    use kinetic_energy_density_vars
+    use constants, only: zzero, zone
+    use mod_corestate, only: rhocr, rwfcr
+
+    implicit none 
+    private
+
+    public :: gen_ked_cr, get_clebsch_gordan_coeffs
+
+    contains
+    
+    !> Returns the kinetic energy density for the core electrons in the muffin-tin region.
+    !> This is done by generating the four components of the core wavefunction and then taking the 
+    !> gradient for each component. 
+    !> From the gradient of the wavefunction, the total kinetic energy density can be 
+    !> constructed.
+    subroutine gen_ked_cr(ked_cr)
+        !> Kinetic energy density for the core electrons
+        real(dp), intent(inout) :: ked_cr(:,:,:)
+
+        integer :: ist, nr_core_state, wf_component
+        integer :: is, ia, ias
+        integer :: nr, m, lm, ir
+
+        complex(dp), allocatable :: ked_cr_per_state(:,:,:,:)
+        real(dp), allocatable :: wf_cr_per_state(:, :)
+
+        integer, allocatable :: core_states(:,:), core_states_count(:)
+
+        allocate(ked_cr_per_state(ked_lmmaxvr, nrmtmax, 4, natmtot), source=zzero)
+        allocate(wf_cr_per_state(nrmtmax, 4), source=0.0_dp)
+
+        allocate(core_states(spnstmax, nspecies))  
+        allocate(core_states_count(nspecies))
+
+        ! For each species, get the number of core states and their indices
+        do is = 1, nspecies
+            core_states_count(is) = COUNT(spcore(:, is))
+            core_states(1:core_states_count(is), is) = PACK([(ist, ist=1, spnst(is))], spcore(1:spnst(is), is))
+        end do
+    
+        ked_cr = 0.0_dp
+        do is = 1, nspecies
+            nr = nrmt(is)
+            do ia = 1, natoms(is)
+                ias = idxas(ia, is)
+                do nr_core_state = 1, core_states_count(is)
+                    ist = core_states(nr_core_state, is) 
+                    do m = -spk(ist, is), spk(ist, is)-1
+                        call gen_wf_cr_per_state(is, ia, ist, m, wf_cr_per_state) 
+                        call gen_ked_cr_per_state(is, ia, ist, m, wf_cr_per_state, ked_cr_per_state(:, :, :, ias))
+                        do wf_component = 1, 4
+                            do lm = 1, ked_lmmaxvr
+                                do ir = 1, nrmt(is)
[…]
