### 2025-06-16 bd0841b
GW_INFO
