+++ src/src_gw/gw_info.f90
@@ -1,0 +1,227 @@
+module gw_info
+  use asserts, only: assert
+  use gw_io, only: open_file, file_format_text
+  use mod_mpi_gw, only: indexes_parallelization
+  use precision, only: dp, i32, str_64, str_128
+
+  implicit none
+
+  private
+
+  ! Strings informing progress status 
+  character(len=*), public, parameter :: progress_coulomb = 'coulomb'
+  character(len=*), public, parameter :: progress_sgi = 'sgi'
+  character(len=*), public, parameter :: progress_ipw = 'ipw'
+  character(len=*), public, parameter :: progress_calculate_epsilon = 'calc_eps'
+  character(len=*), public, parameter :: progress_polarizability_tetrahedron = 'qdepwtet'
+  character(len=*), public, parameter :: progress_write_epsilon = 'write_eps'
+  character(len=*), public, parameter :: progress_epsilon_loop_kpoints = 'eps_loop_kpt'
+  character(len=*), public, parameter :: progress_epsilon_loop_blocks = 'eps_loop_blocks'
+
+  ! Strings to name symbols
+  character(len=*), public, parameter :: q_points_numbering = 'dummy index to list q-points'
+  character(len=*), public, parameter :: q_points_numbering_abbr = '(xq)'
+  character(len=*), public, parameter :: q_points_indexes = 'q-points indexes (as used in the code)'
+  character(len=*), public, parameter :: q_points_indexes_abbr = 'q'
+  character(len=*), public, parameter :: k_points_numbering = 'dummy index to list k-points'
+  character(len=*), public, parameter :: k_points_numbering_abbr = '(xk)'
+  character(len=*), public, parameter :: k_points_indexes = 'k-points indexes (as used in the code)'
+  character(len=*), public, parameter :: k_points_indexes_abbr = 'k'
+  character(len=*), public, parameter :: empty_bands_indexes = 'unoccupied states'
+  character(len=*), public, parameter :: empty_bands_abbr = 'm'
+  character(len=*), public, parameter :: bands_indexes = 'states (may include core states, depending on the choice given in input.xml)'
+  character(len=*), public, parameter :: bands_abbr = 'm'
+
+  !> Unit used for the general GW output file (`GW_INFO.OUT`)
+  integer(i32), public, protected :: fgw
+  !> Default name of the general GW output file
+  character(len=*), parameter :: filename_gwinfo = 'GW_INFO.OUT'
+
+  public :: open_gwinfo, write_to_gwinfo, write_to_gwinfo_boxmessage, &
+    write_to_gwinfo_parallelization_info, write_to_gwinfo_progress, &
+    write_to_gwinfo_progress_bar, write_to_gwinfo_table_with_index_map
+
+contains
+
+!> Open the `GW_INFO.OUT`
+subroutine open_gwinfo( )
+  
+  call open_file( filename_gwinfo , 'write', file_format_text, fgw )
+
+end subroutine
+
+
+!> Write a string into `GW_INFO.OUT`
+subroutine write_to_gwinfo( string )
+  character(len=*), intent(in) :: string
+
+  write( fgw, * ) string
+
+end subroutine
+
+
+!> Write information about the progress of subtasks in a `taskGroup` calculation
+subroutine write_to_gwinfo_progress( string )
+  character(len=*), intent(in) :: string
+  
+  select case( trim(string) )
+    case( progress_coulomb )
+      call write_to_gwinfo('*** Reading Coulomb matrix from file')
+    case( progress_sgi )
+      call write_to_gwinfo('*** Reading matrix with orthogonal IPW from file')
+    case( progress_ipw )
[…]
