### 2025-06-24 a71650b
Light CI
