### 2025-06-25 363f54a
Changed the solver to one allowing ill-matrices
