+++ src/bandstructure.f90
@@ -348,10 +348,12 @@ End Subroutine bandstr
 !> This subroutine interpolate function \(f1_n( k )\) defined on the kmesh 1, to kmesh 2
 !> using 3D Smooth Fourier transform according to PRB 38, 2721 (1988).
 subroutine fourintp(f1, nk1, kvecs1, f2, nk2, kvecs2, nb)
-  use constants, only: zone
-  use fouri, only: nrr, nst, rbas, rindex, rst, setrindex_done
+  use constants,    only: zone
+  use modmpi,       only: terminate_if_false
+  use linear_system_ill_defined_safe, only: ill_defined_safe_solve
+  use fouri,        only: nrr, nst, rbas, rindex, rst, setrindex_done
   use mod_symmetry, only: nsymcrys
-  use xlapack, only: matrix_multiply
+  use xlapack,      only: matrix_multiply
 
   integer(i32), intent(in) :: nk1,nk2,nb
   real(dp),    intent(in) :: kvecs1(3,nk1),kvecs2(3,nk2)
@@ -359,8 +361,12 @@ subroutine fourintp(f1, nk1, kvecs1, f2, nk2, kvecs2, nb)
   complex(dp), intent(out):: f2(nk2,1:nb) 
     
   integer(i32) :: i, ist, ib, ik, jk, ir
-  integer(i32) :: info
-  integer(i32), allocatable  :: ipiv(:)
+  integer(i32) :: info, rank, lwork, lrwork, liwork
+  real(dp), parameter   :: zero_tolerance = 1.0e-6_dp
+  real(dp), allocatable :: singular_values(:)
+  complex(dp), allocatable :: work(:)
+  real(dp), allocatable :: rwork(:)
+  integer(i32), allocatable :: iwork(:)
     
   real(dp) :: den, pref, kdotr
   real(dp) :: rmin,rlen,x2,x6,c1,c2
@@ -402,15 +408,15 @@ subroutine fourintp(f1, nk1, kvecs1, f2, nk2, kvecs2, nb)
   c2 = 0.25_dp
 
   allocate(smat1(nk1,nst), &
-  &        smat2(nk2,nst), &
-  &        rho(nst),       &
-  &        coef(nst,nb),   &
-  &        ipiv(1:nk1-1),  &
-  &        sm2(1:nk1-1,1:nst), &
-  &        h(1:nk1-1,1:nk1-1), &
-  &        dele(1:nk1-1,1:nb))
+           smat2(nk2,nst), &
+           rho(nst),       &
+           coef(nst,nb),   &
+           singular_values(1:nk1-1),  &
+           sm2(1:nk1-1,1:nst), &
+           h(1:nk1-1,1:nk1-1), &
+           dele(1:nk1-1,1:nb))
   
-  den = dble(nsymcrys)
+  den = real(nsymcrys, kind=dp)
 
   ! Calculate the star expansion function at each irreducible k-point
   smat1(1:nk1,1:nst) = zzero
@@ -418,8 +424,8 @@ subroutine fourintp(f1, nk1, kvecs1, f2, nk2, kvecs2, nb)
     kvec(1:3) = kvecs1(1:3,ik)
     do ir = 2, nrr
       ist = rst(1,ir)
-      pref = dble(rst(2,ir))
-      r(1:3) = dble(rindex(1:3,ir))
+      pref = real(rst(2,ir), kind=dp)
+      r(1:3) = real(rindex(1:3,ir), kind=dp)
       kdotr = twopi*sum( r(1:3)*kvec(1:3) ) 
[…]
