### 2025-06-30 8e1d828
mGGA: spin-unpolarised potential
