+++ src/mgga/mgga_poteff.f90
@@ -1,0 +1,295 @@
+!> Routines to calculate the mgga or gga potential. 
+module mgga_poteff
+    use precision, only: dp
+    use modinput, only: input
+    use asserts, only: assert
+    use mgga_potxc, only: potxc_ir_spinunpolarised, potxc_mt_spinunpolarised
+    use mod_potential_and_density, only: vclir, vclmt, vhalfir, vhalfmt
+    use mod_muffin_tin, only: nrmtmax, nrmt, lmmaxvr, nrmtinr, lmmaxinr
+    use mod_atoms, only: nspecies, natoms, idxas, natmtot
+    use mod_Gvector, only: ngrtot
+
+    implicit none 
+
+    private 
+
+    !> muffin-tin effective GGA potential used to construct the basis functions for meta-GGA calculations
+    real(dp), public, allocatable :: veffmt_gga(:, :, :)
+    !> muffin-tin GGA potential used to construct the basis functions for meta-GGA calculations
+    real(dp), public, allocatable :: vxcmt_gga(:, :, :)
+    !> muffin-tin GGA exchange energy density 
+    real(dp), public, allocatable :: exmt_gga(:, :, :)
+    !> muffin-tin GGA correlation energy density 
+    real(dp), public, allocatable :: ecmt_gga(:, :, :)
+    
+    !> interstitial effective GGA-potential used to construct the basis functions for meta-GGA calculations
+    real(dp), public, allocatable :: veffir_gga(:)
+    !> interstitial GGA-potential 
+    real(dp), public,  allocatable :: vxcir_gga(:)  
+    !> interstitial GGA exchange energy density 
+    real(dp), public, allocatable :: exir_gga(:)
+    !> interstitial GGA correlation energy density 
+    real(dp), public, allocatable :: ecir_gga(:)
+
+    public :: calc_poteff_mgga, calc_poteff_gga
+    public :: init_poteff_gga
+    public :: mgga_mixer
+
+    contains 
+
+        !> Allocate all arrays needed to calculate the GGA potential
+        subroutine init_poteff_gga()
+            if ( allocated( veffmt_gga) ) deallocate(veffmt_gga )
+            allocate(veffmt_gga(lmmaxvr, nrmtmax, natmtot))
+            if ( allocated( vxcmt_gga) ) deallocate( vxcmt_gga)
+            allocate(vxcmt_gga(lmmaxvr, nrmtmax, natmtot))
+            if ( allocated( exmt_gga) ) deallocate(exmt_gga )
+            allocate(exmt_gga(lmmaxvr, nrmtmax, natmtot))
+            if ( allocated( ecmt_gga) ) deallocate(ecmt_gga )
+            allocate(ecmt_gga(lmmaxvr, nrmtmax, natmtot))
+            if ( allocated( veffir_gga) ) deallocate(veffir_gga )
+            allocate(veffir_gga(ngrtot))
+            if ( allocated( vxcir_gga) ) deallocate(vxcir_gga )
+            allocate(vxcir_gga(ngrtot))
+            if ( allocated( exir_gga) ) deallocate(exir_gga )
+            allocate(exir_gga(ngrtot))
+            if ( allocated( ecir_gga) ) deallocate(ecir_gga)
+            allocate(ecir_gga(ngrtot))
+        end subroutine 
+
+        !> Calculate the effective meta-GGA potential. Returns the multiplicative potential `vxcmt_mgga` and `vxcir_mgga` and the mon-multiplicative potential 
+        !> `vxcmt_mgga_nonmult` and `vxcir_mgga_nonmult`. 
[…]
