### 2025-07-01 3e4e169
mGGA spinunpol eigensystem
