### 2025-07-03 56329dc
Task polarizability
