+++ src/src_gw/mod_head_and_wings.f90
@@ -1,0 +1,276 @@
+!> This module contains all the procedures related with the wings and heads of the dielectric/polarizability functions
+module mod_head_and_wings
+
+    use constants, only: fourpi, zzero, zone, pi
+    use precision, only: i32, dp
+    use mod_misc_gw,           only: vi
+#include "offload.fpp"
+    private
+    public :: calchead, calcwings
+
+contains
+
+    !> Computes the head.
+    !> Note: this procedure does not reset the head to zero.
+    !> It accumulates contributions over all k-points.
+    subroutine calchead(ik, iomstart, iomend, ndim, head)
+
+        use modinput,  only: input
+        use modmain,   only:  evalcr, idxas
+        use mod_bands, only: numin, nstdf, nomax, evalfv, metallic
+        use modgw, only: fdebug, kset, fnm, kwfer, time_dfhead, freq
+        use mod_core_states,   only: corind, ncg
+        use mod_dielectric_function, only: pmatvv, pmatcv 
+        implicit none
+
+        ! input/output
+        integer(i32), intent(in) :: ik
+        integer(i32), intent(in) :: iomstart, iomend
+        integer(i32), intent(in) :: ndim
+        complex(dp), intent(inout) :: head(3,3,iomstart:iomend)
+
+        ! local
+        integer(i32) :: ic, icg
+        integer(i32) :: ia, is, ias
+        integer(i32) :: ie1, ie2
+        integer(i32) :: ikp, iop, jop
+        integer(i32) :: iom
+        real(dp) :: edif    ! energy difference
+        real(dp) :: tstart, tend
+        complex(dp) :: coefh
+        complex(dp) :: pnm, zsum
+
+        if (input%gw%debug) write(fdebug,*) ' ---- calchead started ----'
+        call timesec(tstart)
+
+        ! position in the non-reducied grid
+        ikp = kset%ik2ikp(ik)
+
+        ! constant prefactor
+        coefh = cmplx(4.0_dp*pi*vi, 0.0_dp, kind=dp)
+
+        ! loop over tensor components
+        do jop = 1, 3
+        do iop = 1, 3
+
+            ! loop over frequencies
+            do iom = iomstart, iomend
+
+                ! Inter-band contribution
+                zsum = zzero
+                do ie2 = numin, nstdf
+                do ie1 = 1, ndim
+                    if (ie1 <= nomax) then
+                        ! valence-valence
+                        edif = evalfv(ie2,ikp)-evalfv(ie1,ikp)
+                        if (abs(edif) > 1.0e-6_dp) then
+                            pnm = pmatvv(ie1,ie2,iop)*conjg(pmatvv(ie1,ie2,jop))
+                            zsum = zsum + fnm(ie1,ie2,iom,ik)*pnm/(edif*edif)
+                        end if
+                    else
+                        ! core-valence
+                        icg  = ie1 - nomax
+                        is   = corind(icg,1)
+                        ia   = corind(icg,2)
+                        ias  = idxas(ia,is)
+                        ic   = corind(icg,3)
+                        edif = evalfv(ie2,ikp) - evalcr(ic,ias)
+                        if (abs(edif) > 1.0e-6_dp) then
+                            pnm = pmatcv(icg,ie2,iop)*conjg(pmatcv(icg,ie2,jop))
+                            zsum = zsum + fnm(ie1,ie2,iom,ik)*pnm/(edif*edif)
+                        end if
+                    end if
+                end do ! ie2
+                end do ! ie1
[…]
