### 2025-07-04 4eb069d
GW subtasks - computing the dielectric function with more parallelization options
