+++ src/src_gw/calcepsilon_ppm.f90
@@ -1,11 +1,15 @@
 
 subroutine calcepsilon_ppm(iq,iomstart,iomend)
-
+    use calculate_dielectric_function, only: calcepsilon, epsilon_indexes
+    use invert_dielectric_function, only: calcinveps
     use modinput
     use modmain
+    use modmpi, only: mpiglobal
     use modgw
-    use invert_dielectric_function, only: calcinveps
     use modxs, only: symt2
+    use mod_mpi_gw, only: indexes_parallelization
+    use mod_bands, only: numin, nstdf
+
     implicit none
     ! input parameters
     integer, intent(in) :: iq
@@ -27,7 +31,11 @@ subroutine calcepsilon_ppm(iq,iomstart,iomend)
 
     ! calculate the dielectric tensor at \omega=0 and \omega=\omega_p
     call init_dielectric_function(mbsiz,1,2,Gamma)
-    call calcepsilon(iq,1,2)
+    call calcepsilon(iq, epsilon_indexes( &
+                      indexes_parallelization( 1, kqset%nkpt, 1, kqset%nkpt ), &
+                      indexes_parallelization( numin, nstdf, numin, nstdf ), &
+                      indexes_parallelization( 1, 2, 1, 2 ) ) &
+                    )
     if( gamma ) then
       call calcinveps(1, 2, gamma, input%gw%scrcoul, freq%fconv, symt2,&
                       &epsilon, epsw1, epsw2, epsh, eps00, time_dfinv)
