### 2025-07-07 #7809 @JosePizarro3
> Hi [@JosePizarro3](https://github.com/JosePizarro3), [@kavanase](https://github.com/kavanase) and [@CompRhys](https://github.com/CompRhys), would you be able to help review this submission?

Hi @zhubonan, not sure if I understood the scope of this review, so just want to make sure: is it about reviewing all the exciting code and only the low-scaling BSE implementation?
