### 2025-07-07 #7809 @zhubonan
Hi @JosePizarro3, It is just about reviewing the low-scaling BSE implementation part.
