### 2025-07-07 10c7be2
script for auto-generating excitingscripts README with docstrings; increase excitingscripts version to 1.0.0
