+++ tools/excitingscripts/README.md
@@ -1,0 +1,1772 @@
+# excitingscripts
+
+`excitingscripts` is a collection of various Python scripts for executing various tasks with the [exciting](https://exciting-code.org) code from the command line.
+
+## Installation
+
+`excitingscripts` can be installed directly from PyPI:
+
+```bash
+pip install excitingscripts
+```
+
+Alternatively, it can be installed from the [exciting source][https://github.com/exciting/exciting]:
+
+```bash
+cd $EXCITINGROOT/tools/excitingscripts
+pip install -e .
+```
+
+## Script Listing
+
+### excitingscripts.checkfit
+
+
+#### excitingscripts.checkfit.checkfit
+
+Check-fit implementation.
+
+##### arg_parser
+
+Get the arg parser for checkfit scripts.
+
+:param quantity: what derivatives to check fit
+:return: the argparser
+
+##### fit
+
+Fit data to a polynomial.
+
+:param order: order of the fit
+:param x: x data
+:param y: y data
+:param order_of_derivative: determines which coefficient is taken from the fit
+:return: the chosen coefficient, if the fit is poorly conditioned "None"
+
+##### get_unit_conversion_factor
+
+Get the unit conversion factor.
+
+:param lattice_param_exp: exponent for the lattice parameter in the unit conversion.
+:return: unit conversion factor
+
+##### print_info_to_stdout
+
+Print some information to the terminal.
+
+:param max_displacement: maximum chosen displacement for the fit
+:param frequencies: the computed frequencies
+:param order_of_derivative: fit order of interest
+:param n_max: number of chosen displacement values
+
+##### quantity_specific_checkfit
+
+Specific checkfit implementation for either energy or force.
+
+:param quantity: name of the quantity key in the result file
+:param factor: factor for the unit conversion
+:param lattice_param_exp: exponential for the lattice parameter in the unit conversion
+return: checkfit function
+
+
+#### excitingscripts.checkfit.energy_vs_displacement
+
+Python script for extracting the derivatives at zero displacement of the energy-vs-displacement curves.
+
+##### arg_parser
+
+Get the arg parser for checkfit scripts.
+
+:param quantity: what derivatives to check fit
+:return: the argparser
+
+##### quantity_specific_checkfit
+
+Specific checkfit implementation for either energy or force.
+
+:param quantity: name of the quantity key in the result file
+:param factor: factor for the unit conversion
+:param lattice_param_exp: exponential for the lattice parameter in the unit conversion
+return: checkfit function
+
+
+#### excitingscripts.checkfit.energy_vs_strain
+
+Python script for extracting derivatives at zero strain of energy-vs-strain curves.
+
+##### fit
+
+Fit data to a polynomial.
+
+:param order: order of the fit
+:param x: x data
+:param y: y data
+:param order_of_derivative: determines which coefficient is taken from the fit
+:return: the chosen coefficient, if the fit is poorly conditioned "None"
+
+##### parse_energy_vs_strain
+
+Read the "energy_vs_strain" file
+
+:param directory: directory containing the file
+:param max_strain: value of maximum strain
+:return: strain and energy values
+
+##### parse_info_elastic_constants
+
+Gives the necessary info to be printed
+
+:param directory: path to the directory containing the file "INFO-elastic-constants"
+:return : dictionary containing the information
+
+##### print_info_to_stdout
+
+Print some information to the terminal.
+
+:param info:
+:param max_strain: maximum chosen strain for the fit
+:param derivatives: the computed derivatives
+:param order_of_derivative: fit order of interest
+:param n_max: number of chosen strain values
+
+##### save_derivatives_to_json
+
+Saving the data in fitted derivative data in a JSON format
+
+:param directory: path to save the file
[…]
