+++ tools/excitingscripts/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
 name = "excitingscripts"
 description = "Executable command-line scripts for basic tasks with the exciting code."
-version = "0.0.1"
+version = "1.0.0"
 
 requires-python = ">=3.6"
 
