### 2025-07-08 69716c3
Write script descriptions to their respective files
