+++ tools/excitingscripts/excitingscripts/setup/volume_optimization.py
@@ -1,3 +1,15 @@
+"""Generate structures at different volumes.
+
+Located at `excitingscripts/setup/volume_optimization.py`.
+
+Call as:
+
+```bash
+python3 -m excitingscripts.setup.volume_optimization nr_vol
+```
+Where <code>nr_vol</code> is the number of volume values for which structures are generated by varying the lattice constant.
+"""
+
 import os
 import pathlib
 from argparse import ArgumentParser
