+++ tools/excitingscripts/excitingscripts/execute/volume_optimization.py
@@ -1,3 +1,15 @@
+"""Run a series of **exciting** calculations for structures with different volumes.
+
+Located at `excitingscripts/execute/volume_optimization.py`.
+
+Call as:
+
+```bash
+python3 -m excitingscripts.execute.volume_optimization nr_vol
+```
+Where <code>nr_vol</code> is the number of volume values for which structures are generated by varying the lattice constant.
+"""
+
 import os
 from argparse import ArgumentParser
 from os.path import join
