### 2025-07-08 e5dfef4
add also resolved README
