+++ tools/excitingscripts/README.md
@@ -1,387 +1,22 @@
 # excitingscripts
 
-<<<<<<< HEAD
-**excitingscripts** is a collection of various python scripts for executing various tasks with the [`exciting`](https://exciting-code.org) code from
-the command line.
-
-## Installation
-
-**excitingscripts** can be installed directly from PyPI:
-=======
 `excitingscripts` is a collection of various Python scripts for executing various tasks with the [exciting](https://exciting-code.org) code from the command line.
 
 ## Installation
 
 `excitingscripts` can be installed directly from PyPI:
->>>>>>> tutorial-updates-sodium
 
 ```bash
 pip install excitingscripts
 ```
 
-<<<<<<< HEAD
-Alternatively, it can be installed from the `exciting` source:
-=======
 Alternatively, it can be installed from the [exciting source][https://github.com/exciting/exciting]:
->>>>>>> tutorial-updates-sodium
 
 ```bash
 cd $EXCITINGROOT/tools/excitingscripts
 pip install -e .
 ```
 
-<<<<<<< HEAD
-## Script Listing
-
-### SETUP
-
-#### <span style="color:#15317E">setup.band_structure</span>
-
-Add band structure element to given input file by getting the band path from the input structure.
-
-Located at `excitingscripts/setup/band_structure.py`.
-
-Call as:
-
-```bash
-python3 -m excitingscripts.setup.band_structure
-```
-
-#### <span style="color:#15317E">setup.convergence_test</span>
-
-Generate input files with different values of the main computational parameters.
-
-Located at `excitingscripts/setup/convergence_test.py`.
-
-Call as:
-
-```bash
-python3 -m excitingscripts.setup.convergence_test k_i k_f rgkmax_i rgkmax_f
-```
-Where <code>k_i</code> and <code>k_f</code> are the initial and final k-values for defining the 
-<code><span style="color:green">groundstate</span></code> attribute 
-<code><span style="color:MediumBlue">ngridk</span></code>, and <code>rgkmax_i</code> and <code>rgkmax_f</code> the 
-initial and final values for the 
-<code><span style="color:green">groundstate</span></code> attribute 
-<code><span style="color:MediumBlue">rgkmax</span></code>.
-
-
-#### <span style="color:#15317E">setup.dft_05</span>
-
-Generate a set of input files varying the attribute <code><span style="color:mediumblue">cut</span></code>.
-
-Located at `excitingscripts/setup/dft_05.py`.
-
-Call as:
-
-```bash
-python3 -m excitingscripts.setup.dft_05 r_cut_min r_cut_max number_r_cut_steps -s species -r root_dir
-```
-Where <code>r_cut_min</code> and <code>r_cut_max</code> are the minimum and maximum values for r_cut, 
-<code>number_r_cut_steps</code> is the number of r_cut values for which input files are generated, <code>species</code>
-is the species with regard to which r_cut is varied and <code>root_dir</code> is the root directory.
-
-
-#### <span style="color:#15317E">setup.dos_band_structure</span>
-
-Add DOS and band structure element to given input file by getting the band path from the input structure.
-
-Located at `excitingscripts/setup/dos_band_structure.py`.
-
-Call as:
-
-```bash
-python3 -m excitingscripts.setup.dos_band_structure
-```
-#### <span style="color:#15317E">setup.excitingroot</span>
-
-Replace placeholder "$EXCITINGROOT" in **input.xml** files by actual path.
-
-Located at `excitingscripts/setup/excitingroot.py`.
-
-Call as:
-
-```bash
-python3 -m excitingscripts.setup.excitingroot
-```
-
-#### <span style="color:#15317E">setup.interlayer_distance</span>
-
-Generate structures with different interlayer distances.
-
[…]
