### 2025-07-10 e859465
Restore RTTDDFT  compatibility with Cray 18
