+++ src/src_xs/src_rttddft/rttddft_numberexcitations_test.f90
@@ -1,163 +1,0 @@
-module rttddft_NumberExcitations_test
-  use constants, only: zone, zi, sqrt_two
-  use exciting_mpi, only: mpiinfo
-  use math_utils, only: all_close
-  use mock_arrays, only: complex_positive_definite_matrix_5x5, complex_hermitian_matrix_5x5, complex_matrix_5x5
-  use precision, only: dp, i32
-  use rttddft_NumberExcitations, only: obtain_number_excitations
-  use rttddft_Wavefunction, only: wavefunction_set, initialize_wavefunction_set
-  use unit_test_framework, only : unit_test_type
-  use xlapack, only: solve_generalized_hermitian_eigenproblem
-  use to_char_conversion, only: to_char  
-
-  implicit none
-
-  private
-
-  public :: rttddft_NumberExcitations_test_driver
-
-  real(dp), parameter :: tol = 1.0e-10_dp
- 
-contains
-
-  subroutine rttddft_NumberExcitations_test_driver( mpiglobal, kill_on_failure)
-    !> mpi information
-    type(mpiinfo), intent(in) :: mpiglobal
-    !> Kill the program before the test driver finishes
-    !> if an assertion fails
-    logical, optional, intent(in) :: kill_on_failure
-    
-    type(unit_test_type) :: test_report
-    integer(i32), parameter :: n_assertions_test_obtain_number_excitations = 6
-    integer(i32), parameter :: n_assertions = n_assertions_test_obtain_number_excitations
-
-    character(len=*), parameter :: module_tested = 'rttddft_NumberExcitations'
-
-    ! Initialize test object
-    call test_report%init(n_assertions, mpiglobal)
-
-    ! Run and assert tests
-    call test_obtain_number_excitations( mpiglobal, test_report )
-
-    ! report results
-    if ( present( kill_on_failure ) ) then
-      call test_report%report( module_tested, kill_on_failure )
-    else
-      call test_report%report( module_tested )
-    end if
-
-    ! Finalise test object
-    call test_report%finalise()
-
-  end subroutine
-
-  !> Unit tests for [[obtain_number_excitations]]
-  subroutine test_obtain_number_excitations( mpiglobal, test_report )
-    !> mpi information
-    type(mpiinfo), intent(in) :: mpiglobal
-    !> Our test object
-    type(unit_test_type), intent(inout) :: test_report
-
-    character(len=*), parameter :: test_identifier = "test_obtain_number_excitations"
-    integer(i32), parameter :: n_states_gnd = 5
-    integer(i32), parameter :: n_states = 3
-    integer(i32), parameter :: n_frozen = 1
-    integer(i32), parameter :: i_VBM = 2
-    integer(i32), parameter :: i_CBm = i_VBM + 1
-    integer(i32), parameter :: n_dim = size( complex_positive_definite_matrix_5x5, 1 )
-    integer(i32) :: ik, n_kpt, n_kpt_per_rank, first_k, last_k, test_counter
-    real(dp) :: n_exc, n_gs, n_exc_ref, n_gs_ref, eigs_gnd(n_states_gnd), eigs(n_states)
-    real(dp), allocatable :: wkpt(:), occ_gnd(:, :), occ(:)
-    complex(dp), allocatable :: H_gnd(:, :), H(:, :), S(:, :), overlap(:, :, :), aux(:, :)
[…]
