+++ src/src_xs/src_rttddft/rttddft_test_drivers.f90
@@ -2,7 +2,6 @@ module rttddft_test_drivers
   use modmpi, only: mpiinfo
   use propagators_test, only: propagators_test_driver
   use rttddft_io_test, only: rttddft_io_test_driver
-  use rttddft_NumberExcitations_test, only: rttddft_NumberExcitations_test_driver
   use rttddft_Wavefunction_test, only: rttddft_Wavefunction_test_driver
 
   implicit none
@@ -29,7 +28,6 @@ contains
     ! Call test drivers here
     call propagators_test_driver( mpiglobal, kill_on_failure_ )
     call rttddft_io_test_driver( mpiglobal, kill_on_failure_ )
-    call rttddft_NumberExcitations_test_driver( mpiglobal, kill_on_failure_ )
     call rttddft_Wavefunction_test_driver( mpiglobal, kill_on_failure_ )
   end subroutine
 end module
