### 2025-07-11 a87cef0
change neon to sodium in tutorials; layout fixes
