### 2025-07-14 0015ae6
Fix excitingtools parse_band_structure for wannier output
