### 2025-07-15 570c97f
GPU offload to task-based GW
