### 2025-07-16 f10458b
Solved issue for old bare BSE with ifx.
