### 2025-07-17 863d50b
Changed CMake so FFTW3 is used for FFT in exciting
