+++ src/elfplot.f90
@@ -18,6 +18,7 @@ Subroutine elfplot
       use modplotlabels
       use modmpi, only : rank
       use constants, only: pi
+      use m_zfftifc, only: zfftifc
 ! !DESCRIPTION:
 !   Outputs the electron localisation function (ELF) for 1D, 2D or 3D plotting.
 !   The spin-averaged ELF is given by
