### 2025-07-17 055ac4a
Enabling BSE Calculations from RT-TDDFT Occupations
