### 2025-07-18 fe305f6
Refactor genlofr
