+++ src/genlofr.f90
@@ -1,162 +1,0 @@
-!
-!
-!
-! Copyright (C) 2002-2005 J. K. Dewhurst, S. Sharma and C. Ambrosch-Draxl.
-! This file is distributed under the terms of the GNU General Public License.
-! See the file COPYING for license details.
-!
-!BOP
-! !ROUTINE: genlofr
-! !INTERFACE:
-!
-!
-Subroutine genlofr
-! !USES:
-      Use modinput
-      Use modmain
-      use mgga_poteff, only: veffmt_gga
-      use modmain, only: iscl
-! !DESCRIPTION:
-!> Generates the local-orbital radial functions. This is done by integrating
-!> the scalar relativistic Schr\"{o}dinger equation or Dirac equation
-!> (or its energy deriatives) at the current linearisation energies using the
-!> spherical part of the effective potential. Dirac-type local orbitals are
-!> useful in the context of spin-orbit coupling, but should only be used along
-!> with second variation with local orbitals. For more details see:
-!> arXiv:2306.02965 [cond-mat.mtrl-sci].
-!> For each local-orbital, a linear combination of {\tt lorbord} radial functions is
-!> constructed such that its radial derivatives up to order ${\tt lorbord}-1$ are zero
-!> at the muffin-tin radius.This function is normalised and the radial Hamiltonian applied
-!> to it. The results are stored in the global array {\tt lofr}.
-!
-!
-! !REVISION HISTORY:
-!   Created March 2003 (JKD)
-!EOP
-!BOC
-      Implicit None
-! local variables
-      Integer :: is, ia, ias, nr, ir
-      Integer :: ilo, io1, io2
-      Integer :: j, l, nn, info, np
-      Real (8) :: t1
-! automatic arrays
-      Real (8) :: vr (nrmtmax), fr (nrmtmax), gr (nrmtmax), cf (3, &
-     & nrmtmax)
-      Real (8) :: p0 (nrmtmax, maxlorbord), p1 (nrmtmax,maxlorbord)
-      Real (8) :: q0 (nrmtmax, maxlorbord), q1 (nrmtmax, maxlorbord)
-      Real (8) :: p0s (nrmtmax), p1s (nrmtmax) ,q0s (nrmtmax), q1s (nrmtmax)
-      Real (8) :: hp0 (nrmtmax)
-! allocatable arrays
-      Integer, Allocatable :: ipiv (:)
-      Real (8), Allocatable :: xa (:), ya (:)
-      Real (8), Allocatable :: a (:, :), b (:), c (:)
-! external functions
-      Real (8) :: polynom
-      External polynom
-! variables for the lo recommendation
-      Real (8) energy,energyp,tmp,tmp2,ens(0:20),elo,ehi,flo,fhi,emi,fmi
-      integer nodes
-      call stopwatch("exciting:genlofr", 1)
-      
-      np = Max (maxlorbord+1, 4)
-      Allocate (ipiv(np))
-      Allocate (xa(np), ya(np), c(np))
-      Allocate (a(np, np), b(np))
-      Do is = 1, nspecies
-         nr = nrmt (is)
-         Do ia = 1, natoms (is)
-            ias = idxas (ia, is)
-            if (associated(input%groundstate%mgga) .and. iscl > 1) then 
-               vr(1:nr) = veffmt_gga (1, 1:nr, ias) * y00
-            else 
-               vr (1:nr) = veffmt (1, 1:nr, ias) * y00
-            end if            
-            Do ilo = 1, nlorb (is)
-               l = lorbl (ilo, is)
-               Do io2 = 1, lorbord (ilo, is)
-                  if (wfkappa(io2, ilo, is) /= 0) then
[…]
