### 2025-07-18 edbbb5d
Module for Solving the Secular Equation in the Presence of Inversion Symmetry
