+++ src/src_sym/mod_secular_equation_inversion_symmetry.f90
@@ -1,0 +1,318 @@
+! This file is distributed under the terms of the GNU General Public License.
+! See the file COPYING for license details.
+! Copyright (C) exciting Code, SOL group. 2025
+
+! Created April 2025 (Mara)
+
+!> Module for taking advantage of inversion symmetry when solving the secular equation.
+!>
+!> In systems with inversion symmetry, the APWs transform in the following way:
+!> \[
+!>      \phi_{\textbf{G}+\textbf{k}}(-\textbf{r})=\phi^\ast_{\textbf{G}+\textbf{k}}(\textbf{r}) \: .
+!> \]
+!> This makes the generally complex Hermitian Hamiltonian and overlap matrices become real symmetric.
+!> Since LOs generally do not respect this property, they are attached to so-called fictitious planewaves by
+!> constructing a coefficient matrix to transform the Hamiltonian and overlap matrices into the real basis.
+!> Diagonalization is then performed in this real representation and the resulting eigenvectors are back-transformed
+!> to maintain compatibility with the rest of the code.
+!>
+!> This module gathers all the necessary code.
+module mod_secular_equation_inversion_symmetry
+  use precision, only: dp, i32
+  use constants, only: zzero, zil
+  use modinput, only: input
+  use matrix_rank, only: matrix_rank_SVD
+  use general_matrix_multiplication, only: matrix_multiply
+  use qr_factorization, only: qr_column_pivot
+  use asserts, only: assert
+  use modfvsystem, only: evsystem
+  use generalized_hermitian_eigenproblem, only: solve_generalized_hermitian_eigenproblem
+  use modfvsystem, only: deletesystem
+
+  implicit none
+  private
+  public :: get_lo_transformation_matrix_inv_sym, set_lo_transformation_matrix_inv_sym, build_coefficient_matrix_lo, &
+       solve_secular_equation_inversion_symmetry, transform_eigenvectors_inversion_symmetry
+
+
+  !> Coefficients for attaching the LOs to fictitious planewaves
+  complex(dp), allocatable :: lo_transformation_matrix_inv_sym (:,:,:,:)
+
+contains
+
+  function get_lo_transformation_matrix_inv_sym(ispn,ik) result(coeff_matrix)
+    integer(i32), intent(in) :: ispn
+    integer(i32), intent(in) :: ik
+    complex(dp), allocatable :: coeff_matrix(:,:)
+
+    call assert(allocated(lo_transformation_matrix_inv_sym), 'lo_transformation_matrix_inv_sym not allocated')
+    coeff_matrix = lo_transformation_matrix_inv_sym(:,:,ispn,ik)
+  end function get_lo_transformation_matrix_inv_sym
+
+  subroutine set_lo_transformation_matrix_inv_sym(coeff_matrix)
+    complex(dp), intent(in) :: coeff_matrix(:,:,:,:)
+
+    if (allocated(lo_transformation_matrix_inv_sym)) deallocate (lo_transformation_matrix_inv_sym)
+    lo_transformation_matrix_inv_sym = coeff_matrix
+
+  end subroutine set_lo_transformation_matrix_inv_sym
+
+  !> This subroutine builds the coefficients used to attach the LOs to so-called fictitious planewaves in order to
+  !> transform the Hamiltonian and overlap matrices into the real basis. This is done by forming linear combinations
+  !> over the LOs,
+  !> \[
[…]
