+++ src/zfinp.f90
@@ -47,29 +47,19 @@ Complex (8) Function zfinp (tsh, zfmt1, zfmt2, zfir1, zfir2)
       Complex (8), Intent (In) :: zfir2 (ngrtot)
 ! local variables
       Integer :: is, ia, ias, ir
-      Complex (8) zsum,zsum0
+      Complex (8) zsum
 ! external functions
       Complex (8) zfmtinp
       External zfmtinp
       zsum = 0.d0
 ! interstitial contribution
-
-#ifdef USEOMP
-!$OMP PARALLEL PRIVATE (ir,zsum0)
- zsum0= 0.d0
-!$OMP DO  
-      Do ir = 1, ngrtot
-         zsum0 = zsum0 + cfunir (ir) * conjg (zfir1(ir)) * zfir2 (ir)
-      End Do
-!$OMP END DO
-!$OMP ATOMIC
-     zsum = zsum +zsum0
-!$OMP END PARALLEL  
-#else
+      !$omp parallel do default(none) private(ir) shared(ngrtot,cfunir,zfir1,zfir2) reduction(+:zsum)
       Do ir = 1, ngrtot
          zsum = zsum + cfunir (ir) * conjg (zfir1(ir)) * zfir2 (ir)
       End Do
-#endif
+      !$omp end parallel do
+
+
       zsum = zsum * omega / dble (ngrtot)
 ! muffin-tin contribution
       Do is = 1, nspecies
