### 2025-07-19 a7a10c3
Add new Externals
