+++ external/spglib-2.6.0/test/utils.c
@@ -1,0 +1,281 @@
+#include <stdio.h>
+#include <stdlib.h>
+
+#include "spglib.h"
+
+void show_symmetry_operations(int const (*rotations)[3][3],
+                              double const (*translations)[3], int const size) {
+    int i, j;
+    for (i = 0; i < size; i++) {
+        printf("--- %d ---\n", i + 1);
+        for (j = 0; j < 3; j++) {
+            printf("%2d %2d %2d\n", rotations[i][j][0], rotations[i][j][1],
+                   rotations[i][j][2]);
+        }
+        printf("%f %f %f\n", translations[i][0], translations[i][1],
+               translations[i][2]);
+    }
+}
+
+void show_matrix_3d(double const lattice[3][3]) {
+    for (int i = 0; i < 3; i++) {
+        printf("%f %f %f\n", lattice[0][i], lattice[1][i], lattice[2][i]);
+    }
+}
+
+void show_cell(double const lattice[3][3], double const positions[][3],
+               int const types[], int const num_atoms) {
+    printf("Lattice parameter:\n");
+    show_matrix_3d(lattice);
+    printf("Atomic positions:\n");
+    for (int i = 0; i < num_atoms; i++) {
+        printf("%d: %f %f %f\n", types[i], positions[i][0], positions[i][1],
+               positions[i][2]);
+    }
+}
+
+void show_spacegroup_type(SpglibSpacegroupType const spgtype) {
+    printf("Number:            %d\n", spgtype.number);
+    printf("International:     %s\n", spgtype.international_short);
+    printf("International:     %s\n", spgtype.international_full);
+    printf("International:     %s\n", spgtype.international);
+    printf("Schoenflies:       %s\n", spgtype.schoenflies);
+    printf("Hall symbol:       %s\n", spgtype.hall_symbol);
+    printf("Point group intl:  %s\n", spgtype.pointgroup_international);
+    printf("Point group Schoe: %s\n", spgtype.pointgroup_schoenflies);
+    printf("Arithmetic cc num. %d\n", spgtype.arithmetic_crystal_class_number);
+    printf("Arithmetic cc sym. %s\n", spgtype.arithmetic_crystal_class_symbol);
+}
+
+void show_magnetic_symmetry_operations(int const (*rotations)[3][3],
+                                       double const (*translations)[3],
+                                       int const *time_reversals,
+                                       int const size) {
+    int i, j;
+    for (i = 0; i < size; i++) {
+        printf("--- %d ---\n", i + 1);
+        for (j = 0; j < 3; j++) {
+            printf("%2d %2d %2d\n", rotations[i][j][0], rotations[i][j][1],
+                   rotations[i][j][2]);
+        }
+        printf("%f %f %f\n", translations[i][0], translations[i][1],
+               translations[i][2]);
+        printf("%2d\n", time_reversals[i]);
+    }
+}
+
+void show_magnetic_spacegroup_type(SpglibMagneticSpacegroupType const msgtype) {
+    printf("UNI Number:    %d\n", msgtype.uni_number);
+    printf("Litvin Number: %d\n", msgtype.litvin_number);
+    printf("BNS Number:    %s\n", msgtype.bns_number);
+    printf("OG Number:     %s\n", msgtype.og_number);
+    printf("Number:        %d\n", msgtype.number);
+    printf("Type:          %d\n", msgtype.type);
+}
+
+void show_spg_magnetic_dataset(SpglibMagneticDataset const *dataset) {
+    int i, p, s;
+    printf("UNI number: %d\n", dataset->uni_number);
+    printf("Type: %d\n", dataset->msg_type);
[…]
