+++ external/IDieL/benchmark/procedures_3d.f90
@@ -1,0 +1,158 @@
+! Copyright (C) 2020-2024 GreenX library
+!
+! Licensed under the Apache License, Version 2.0 (the "License");
+! you may not use this file except in compliance with the License.
+! You may obtain a copy of the License at
+!
+!   http://www.apache.org/licenses/LICENSE-2.0
+!
+! Unless required by applicable law or agreed to in writing, software
+! distributed under the License is distributed on an "AS IS" BASIS,
+! WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+! implied. See the License for the specific language governing
+! permissions and limitations under the License.
+
+!> @file
+!> Contains program to see procedure 3d for a real case (black phosphorous)
+
+!> Program showcasing the 3D routines
+program procedures_3d
+
+    use idiel_constants, only: aip, i32, pi, twopi, zzero
+    use idiel, only: idiel_t
+
+    implicit none
+
+    ! GW data
+    integer(i32), parameter :: nomega = 32_i32  ! Number of frequencies
+    integer(i32) :: iom
+
+    ! Mesh data
+    integer(i32), parameter :: ngrid(3) = [6_i32, 6_i32, 8_i32]
+
+    ! Silicon crystal data
+    integer(i32), parameter :: natoms = 4_i32
+    real(aip) :: a(3), b(3), c(3), lattice(3,3)
+    real(aip), allocatable :: redpos(:,:)
+    integer(i32), allocatable :: types(:)
+
+    ! Dielectric data
+    complex(aip), allocatable :: head(:,:,:), wingL(:,:,:)
+    complex(aip), allocatable :: wingU(:,:,:), body(:,:,:)
+
+    ! Computation object
+    type(idiel_t) :: inv_diel
+
+    ! In case that no SPG we set only Identity (i.e. no symmetry)
+    integer(i32) :: nsym = 8
+    real(aip)    :: crot(3,3,8)
+
+    ! Lattice vectors
+    a(:) = [3.13061478636775_aip, -9.89555085794508_aip, 0.00000000000000_aip]
+    b(:) = [3.13061478636775_aip,  9.89555085794508_aip, 0.00000000000000_aip]
+    c(:) = [0.00000000000000_aip,  0.00000000000000_aip, 8.26566207441072_aip]
+
+    lattice(1,:) = a(:)
+    lattice(2,:) = b(:)
+    lattice(3,:) = c(:)
+
+    ! Atomic types (they might differ that the atomic number)
+    allocate(types(natoms))
+    types(:) = [15_i32, 15_i32, 15_i32, 15_i32]
+
+    ! Reduced positions
+    allocate(redpos(3,natoms))
+    redpos(:,1) = [0.8966_aip, 0.1034_aip, 0.0806_aip]
+    redpos(:,2) = [0.6034_aip, 0.3966_aip, 0.5806_aip]
+    redpos(:,3) = [0.3966_aip, 0.6034_aip, 0.4194_aip]
+    redpos(:,4) = [0.1034_aip, 0.8966_aip, 0.9194_aip]
+
+    ! Reading files
+    call load_from_file('diel3d/head',  head)
+    call load_from_file('diel3d/wingL', wingL)
+    call load_from_file('diel3d/wingU', wingU)
+    call load_from_file('diel3d/body',  body)
+
+    !!!!!!!!!!!! Here it goes the main programm !!!!!!!!!!!!!!!!!!!
+
+    ! Init common objects
+#ifdef USE_SPGLIB
+    call inv_diel%init_common(lattice, redpos, types, ngrid, 3_i32)
+#else 
[…]
