### 2025-07-19 20fd3f1
Test Case: Real Secular Equation Inversion Symmetry
