### 2025-07-20 bb08ba8
GW subtasks parallel sigmac
