### 2025-07-20 #7809 @zhubonan
Hi @JosePizarro3 just want to check if you are happy to review this submission. It is just for the low-scaling BSE implementation part of the code.
