### 2025-07-21 9eb98c7
Preparation for merge of eph code
