+++ src/src_wannier/mod_wannier_helper.f90
@@ -41,14 +41,14 @@ module mod_wannier_helper
             wf_kset%vkl = vkl
             wf_kset%vkc = vkc
           end if
-          call generate_k_vectors( wf_kset_red, bvec, input%groundstate%ngridk, input%groundstate%vkloff, .true., .false.)
+          call generate_k_vectors( wf_kset_red, bvec, input%groundstate%ngridk, input%groundstate%vkloff, .true.)
           nstfv = min( minval( nmat_ptr), int( chgval/2.d0) + input%groundstate%nempty + 1)
         case( "gw")
           input%groundstate%stypenumber = -1 ! turn on LIBBZINT
           if (xctype(1) >= 400) then
             ! GW@hybrids
             nstfv = min( minval( nmat_ptr), int( chgval/2.d0) + input%groundstate%nempty + 1)
-            call init1()
+            call init1
             call generate_k_vectors( wf_kset, bvec, input%groundstate%ngridk, input%groundstate%vkloff, .false.)
             call generate_k_vectors( wf_kset_red, bvec, input%groundstate%ngridk, input%groundstate%vkloff, .true.)
           else
@@ -61,7 +61,7 @@ module mod_wannier_helper
             input%groundstate%vkloff = input%gw%vqloff
             nempty = input%groundstate%nempty
             input%groundstate%nempty = input%gw%nempty
-            call init1()
+            call init1
             input%groundstate%reducek = reducek
             input%groundstate%ngridk = ngridk
             input%groundstate%vkloff = vkloff
@@ -87,7 +87,7 @@ module mod_wannier_helper
         case( "hybrid")
           nstfv = min( minval( nmat_ptr), int( chgval/2.d0) + input%groundstate%nempty + 1)
           input%groundstate%stypenumber = -1 ! turn on LIBBZINT
-          call init1()
+          call init1
           call generate_k_vectors( wf_kset, bvec, input%groundstate%ngridk, input%groundstate%vkloff, .false.)
           call generate_k_vectors( wf_kset_red, bvec, input%groundstate%ngridk, input%groundstate%vkloff, .true.)
         case default
@@ -317,12 +317,13 @@ module mod_wannier_helper
     ! for a given set of eigenenergies
     subroutine wfhelp_occupy( kset, eval, fst, lst, efermi, occ, tetra)
       use mod_opt_tetra
+      use mod_occupy, only: find_fermi
 
       type( k_set), intent( in)           :: kset
       integer, intent( in)                :: fst, lst
-      real(8), intent( in)                :: eval( lst-fst+1, kset%nkpt)
+      real(8), intent( in)                :: eval(fst:lst, kset%nkpt)
       real(8), intent( out)               :: efermi
-      real(8), intent( out)               :: occ( lst-fst+1, kset%nkpt)
+      real(8), intent( out)               :: occ(fst:lst, kset%nkpt)
       type( t_set), optional, intent( in) :: tetra
 
       integer, parameter :: maxit = 1000
@@ -353,7 +354,7 @@ module mod_wannier_helper
           write( *, '("Warning (wfhelp_occupy): No valence bands given. All bands are considered to be unoccupied. Fermi energy set to lowest energy given.")')
         end if
         occ = 0.d0
-        efermi = minval( eval(1,:))
+        efermi = minval( eval(fst,:))
         return
       end if
       if( (lst .le. nvm)) then
@@ -362,12 +363,12 @@ module mod_wannier_helper
[…]
