+++ src/dfpt/phonons_eigensystem.f90
@@ -146,18 +146,14 @@ module phonons_eigensystem
     !> are the response of the overlap and Hamiltonian matrix, respectively, as obtained from
     !> [[ph_eig_gen_dSHmat(subroutine)]].
     !>
-    !> If `projector=.true.`, then the eigenvector response corresponding to the projection of the
-    !> wavefunction response onto the manifold of unoccupied states, 
-    !> \(\hat{\bf P}_u \delta^{\bf q}_{I \mu} \psi_{n{\bf k}}({\bf r})\), is computed. This is 
-    !> sufficient to compute the density response and ensures (given that \(n\) corresponds to
-    !> an occupied state) that the energie differences \(\epsilon_{m{\bf k+q}} - \epsilon_{n{\bf k}}\)
-    !> are always greater than zero.
+    !> This routine implements a solver that also works for fractional occupations (metals) and treats 
+    !> degenerate states explicitely.
     subroutine ph_eig_sternheimer( ik, Gkqset, fst, lst, evalk, occk, evalkq, occkq, eveckq, dSmat, dHmat, gamma, deval, devec, &
-        projector, eps_deg )
+        devecf )
       use constants, only: zzero, zone
       use mod_kpointset, only: Gk_set
       use mod_APW_LO, only: nlotot
-      use mod_eigenvalue_occupancy, only: occmax
+      use mod_eigenvalue_occupancy, only: occmax, efermi
       use modinput
       !> index of the \({\bf k}\) point
       integer, intent(in) :: ik
@@ -179,63 +175,79 @@ module phonons_eigensystem
       real(dp), intent(out) :: deval(:)
       !> eigenvector response at \({\bf k}\)
       complex(dp), intent(out) :: devec(:,:)
-      !> calculate the eigenvector response corresponding to the projection
-      !> of the wavefunction response onto the manifold of unoccupied states
-      !> (default: `.true.`)
-      logical, optional, intent(in) :: projector
-      !> tolerance for degenerate eigenvalues (default: `1e-5`)
-      real(dp), optional, intent(in) :: eps_deg
+      !> additional eigenvector response for force response at \({\bf k}\)
+      complex(dp), optional, intent(out) :: devecf(:,:)
 
-      integer :: nmatkq, ist, jst, nst
-      real(dp) :: eps, dev, sig, t1, t2
-      logical :: proj
+      integer :: nmatkq, ist, jst, nst, stype
+      real(dp) :: dev, sigma, t1, t2
 
       complex(dp), allocatable :: dX(:,:)
 
-      proj = .true.
-      if( present( projector ) ) proj = projector
-      eps = 1e-5_dp
-      if( present( eps_deg ) ) eps = eps_deg
+      real(8), external :: stheta, sdelta
 
       nmatkq = Gkqset%ngk(1, ik) + nlotot
       nst = lst - fst + 1
+      stype = input%groundstate%stypenumber
+      if (stype < 0) stype = 3 ! Fermi-Dirac
+      sigma = 1.0_dp / input%groundstate%swidth
 
       allocate( dX(nmatkq, fst:lst) )
 
       deval = 0.0_dp
+      dX = zzero
       do ist = fst, lst
-        dX(:, ist) = dHmat(1:nmatkq, ist) - evalk(ist) * dSmat(1:nmatkq, ist)
-        if( gamma ) deval(ist) = dble( dX(ist, ist) )
-        if( occk(ist) < input%groundstate%epsocc ) cycle
+        if (gamma) deval(ist) = dble( dHmat(ist, ist) - evalk(ist) * dSmat(ist, ist) )
[…]
