### 2025-07-21 d9efe71
For text IO, only used in the CI the Fortran
