### 2025-07-22 f9cbec0
increase tolerances for assertions to pass tests
