### 2025-07-22 dc19879
Tutorial section for parallel BSE execution
