+++ tools/exciting_tools/excitingtools/utils/valid_attributes.py
@@ -963,7 +963,7 @@ barecoul_attribute_types = {"barcevtol": (float, 1),
                             "pwm": (float, 1),
                             "stctol": (float, 1)} 
 
-scrcoul_attribute_types = {"averaging": (str, ["2d", "isotropic"]),
+scrcoul_attribute_types = {"averaging": (str, ["2d", "anisotropic", "anisotropic-2d", "isotropic"]),
                            "omegap": (float, 1),
                            "q0eps": (float, 3),
                            "scrtype": (str, 1),
