+++ tools/excitingjupyter/excitingjupyter/00_compiling_exciting.md
@@ -1,0 +1,297 @@
+In this tutorial you will lear how to compile of **`exciting`** based on your needs. Further technical details can be found in the `INSTALL.md` file in the `exciting` source directory.
+
+## Table of Contents
+
+1. [Basic Compilation](#basic-compilation)  
+   - 1.1. [Requirements](#requirements)  
+   - 1.2. [Basic compilation using GNU compilers (gfortran)](#basic-compilation-using-gnu-compilers-gfortran)  
+   - 1.3. [Basic compilation using Intel compilers (ifx)](#basic-compilation-using-intel-compilers-ifx)  
+
+2. [How to Tune Your Compile Options for Better Performance](#how-to-tune-your-compile-options-for-better-performance)  
+   - 2.1. [Selecting a performant library](#selecting-a-performant-library)  
+     - 2.1.1. [Linear Algebra Libraries](#linear-algebra-libraries)  
+     - 2.1.2. [FFT Libraries](#fft-libraries)  
+   - 2.2. [Passing Extra Options to the Fortran Compiler](#passing-extra-options-to-the-fortran-compiler)  
+   - 2.3. [Extra Options for Further Tuning with the Modern Intel Compiler (IFX) for Intel Processors](#extra-options-for-further-tuning-with-the-modern-intel-compiler-ifx-for-intel-processors)  
+
+3. [Advanced: Other Production Options](#advanced-other-production-options)  
+   - 3.1. [HDF5](#hdf5)  
+   - 3.2. [ScaLAPACK](#scalapack)  
+   - 3.3. [SIRIUS](#sirius)  
+
+4. [Advanced: GPU Compilation](#advanced-gpu-compilation)  
+   - 4.1. [Compiling for AMD and NVIDIA GPUs using `Crayftn`](#compiling-for-amd-and-nvidia-gpus-using-crayftn)  
+   - 4.2. [Compiling for Intel GPUs using `ifx`](#compiling-for-intel-gpus-using-ifx)  
+     - 4.2.1. [Unified Shared Memory Support](#unified-shared-memory-support)  
+
+## 1. Basic Compilation
+
+**`exciting`** uses **CMake** (version 3.25 or later) for building and installation.
+You can download a precompiled version of CMake suitable for your system from [here](https://cmake.org/download/).
+For **x86-64 Linux**, bundled CMake 3.31.3 binaries are provided with **exciting** under `external/cmake-3.31.3-linux-x86_64/`.
+
+### 1.1 Requirements
+
+**`exciting`** requires `xsltproc` to preprocess its XML schema into code.
+
+Additionally, the code depends on the installation of:
+
+- **FFTW3** implementations (such as `oneMKL`, `AOCL-FFTW`, `FFTW3`, `Cray-FFTW`, etc.)
+- A **BLAS/LAPACK** implementation (such as `oneMKL`, `BLIS` + `libFLAME`, `OpenBLAS`, `LibSci`, etc.)
+
+> **Please ensure you have these libraries and binaries installed before proceeding.**
+
+> **Note:** For performant compilation, the choice of these libraries is essential.
+
+The **exciting** package includes the following external libraries which are necessary for compilation:
+
+- [FoX XML](https://github.com/andreww/fox) — Library for parsing input (2012 version).
+- [LIBXC V7](https://libxc.gitlab.io/) — Library of DFT exchange and correlation functionals. External libXC versions 5.0.0 or higher are also supported.
+- [BSPLINE-FORTRAN](https://github.com/jacobwilliams/bspline-fortran) — Multidimensional B-spline interpolation on a regular grid.
+
[…]
