### 2025-07-23 11b306d
Restoring CI working
