### 2025-07-23 589f639
Restore Cray compatibility prerelease
