### 2025-07-24 1c80847
Update MT Potential Routine to Support Lattice Harmonics
