### 2025-07-24 dbfdce3
activate CI tests for tutorials
