### 2025-07-29 d2beadf
Excitingtools/base/use pydantic
