### 2025-08-05 #7809 @jjkas
make test does not work. Throws the error:
ModuleNotFoundError: No module named 'excitingtools'
