### 2025-08-07 6292adc
Fixes for Unit Test Failures and Compiler Compatibility (Issues #240, #239, #238)
