+++ src/src_sym/mod_lattice_harmonics.F90
@@ -6,14 +6,14 @@
 
 !> Module for computing lattice harmonics (symmetrized real spherical harmonics).
 module mod_lattice_harmonics
-  use modinput, only: input
-  use constants, only: zi, zzero, sqrt_two, real_zero, zone
-  use precision, only: dp
-  use vector_multiplication, only: norm
-  use general_matrix_multiplication, only: matrix_multiply
-  use qr_factorization, only: qr_column_pivot
   use asserts, only: assert
+  use constants, only: real_zero, sqrt_two, zi, zone, zzero
+  use general_matrix_multiplication, only: matrix_multiply
   use grid_utils, only: mesh_1d
+  use qr_factorization, only: qr_column_pivot
+  use modinput, only: input
+  use precision, only: dp, i32
+  use vector_multiplication, only: norm
 
   implicit none
 
@@ -28,7 +28,7 @@ module mod_lattice_harmonics
   !> detailed description of how these arrays are constructed.
   type :: lattice_harmonics_type
      !> Number of lattice harmonics for a specific \( l \)-value and the given index of atoms and species
-     integer, allocatable :: number(:,:)
+     integer(i32), allocatable :: number(:,:)
      !> Coefficients for the lattice harmonic expansion, depending on  \( l \), \( m_1 \) and \( m_2 \) values,
      !> as well as the given index of atoms and species
      real(dp), allocatable :: coefficients(:,:,:,:)
@@ -46,7 +46,7 @@ contains
   function construct_lattice_harmonics_type(num_lattice_harmonics, lattice_harmonics_coeffs, &
        lattice_harmonics_gaunt_coeffs) result(this)
     !> Number of lattice harmonics for a specific \( l \)-value and the given index of atoms and species
-    integer, intent(in) :: num_lattice_harmonics(:,:)
+    integer(i32), intent(in) :: num_lattice_harmonics(:,:)
     !> Coefficients for the lattice harmonic expansion, depending on  \( l \), \( m_1 \) and \( m_2 \) values,
     !> as well as the given index of atoms and species
     real(dp), intent(in) :: lattice_harmonics_coeffs(:,:,:,:)
@@ -79,7 +79,7 @@ contains
     !> Input matrix
     real(dp), intent(in) :: C(:, :)
     !> Number of non-zero rows
-    integer, intent(out) :: num_non_zero_rows
+    integer(i32), intent(out) :: num_non_zero_rows
     !> Reduced matrix
     real(dp), allocatable, intent(out) :: C_reduced(:, :)
 
@@ -98,39 +98,72 @@ contains
   !> \]
   subroutine generate_matrix_complex_to_real_spherical_harmonics(lmax, A)
     !> Maximum angular momentum
-    integer, intent(in) :: lmax
+    integer(i32), intent(in) :: lmax
     !> Matrix which relates complex to real spherical harmonics
     complex(dp), allocatable, intent(out) :: A(:,:,:)
 
-    ! Local variables
-    integer :: l, m1
-    complex(dp):: zone_over_sqrt_two, zi_over_sqrt_two
-
-    zone_over_sqrt_two = zone / sqrt_two
-    zi_over_sqrt_two = zi / sqrt_two
+#if !defined(__INTEL_LLVM_COMPILER)
+    integer(i32) :: l, m1
+#endif
+    complex(dp), parameter :: zone_over_sqrt_two = zone / sqrt_two
+    complex(dp), parameter :: zi_over_sqrt_two = zi / sqrt_two
 
[…]
