### 2025-08-13 70d4ef1
Bug fix BSE from occupations
