+++ src/density/mod_rhovalk.f90
@@ -1,27 +1,29 @@
-! !REVISION HISTORY:
-!   Created April 2003 (JKD)
-
 module mod_rhovalk
 
   use asserts, only: assert
-  use modinput, only: input, isspinspiral, issvlo
-  use mod_kpoint, only: wkpt
-  use mod_eigensystem, only: idxlo
-  use mod_eigenvalue_occupancy, only: nstfv, nstsv
-  use modmpi, only: mpi_env_k, distribute_loop
-  use mod_timing, only: timerho
-  use precision, only: dp
-  use svlo, only: get_num_of_basis_functions_sv
-  use constants, only: zone, zzero
-  use mod_spin, only: ncmag, nspnfv, nspinor, ndmag
-  use mod_muffin_tin, only: nrcmtmax, lmmaxapw, idxlm, nrcmt, nrmt, lmmaxvr, nrmtmax
-  use mod_gkvector, only: ngk, ngkmax, gkc, tpgkc, sfacgk
+  use constants, only: real_zero, zone, zzero
+  
   use mod_APW_LO, only: apwordmax, lofr, nlorb, lorbl
   use mod_atoms, only: natmtot, nspecies, natoms, idxas
-  use mod_SHT, only: zbshtvr, rfshtvr
+  use mod_eigensystem, only: idxlo
+  use mod_eigenvalue_occupancy, only: nstfv, nstsv
+  use mod_gkvector, only: ngk, ngkmax, gkc, tpgkc, sfacgk
   use mod_Gvector, only: ngrtot
+  use mod_kpoint, only: wkpt
+  use mod_muffin_tin, only: nrcmtmax, lmmaxapw, idxlm, nrcmt, nrmt, lmmaxvr, nrmtmax
+  use mod_SHT, only: zbshtvr, rfshtvr
+  use mod_spin, only: ncmag, nspnfv, nspinor, ndmag
+  use mod_timing, only: timerho
+  use modinput, only: input, isspinspiral, issvlo
+  use modmpi, only: mpi_env_k, distribute_loop
+  use precision, only: dp, i32
+  use svlo, only: get_num_of_basis_functions_sv
 
   implicit none
+
+  private
+
+  public :: rhovalk
 
   interface rhovalk
     module procedure :: rhovalk_spin_polarized
@@ -34,19 +36,19 @@ contains
   !> to be used as input arguments for `rhovalk_spin_polarized`
   subroutine rhovalk_non_spin_polarized ( ik, evecfv, occupations, rhomt, magmt, evecsv )
     !> k-point number
-    integer, intent (in) :: ik
+    integer(i32), intent (in) :: ik
     !> First-variational eigenvectors (nmatmax, nstfv)
     complex(dp), contiguous, target, intent (in) :: evecfv(:, :)
     !> State occupations (nstsv)
-    real(dp), intent(in) :: occupations(:)
+    real(dp), contiguous, intent(in) :: occupations(:)
     !> Muffin-tin charge density
-    real(dp), intent(inout) :: rhomt(:, :, :)
+    real(dp), contiguous, intent(inout) :: rhomt(:, :, :)
     !> Muffin-tin magnetisation vector field
-    real(dp), optional, intent(inout) :: magmt(:, :, :, :)
+    real(dp), contiguous, optional, intent(inout) :: magmt(:, :, :, :)
     !> Second-variational eigenvectors (nstfv, nstsv)
-    complex(dp), optional, intent(in) :: evecsv(:, :)
+    complex(dp), contiguous, optional, intent(in) :: evecsv(:, :)
 
-    integer, parameter :: n_spin = 1
+    integer(i32), parameter :: n_spin = 1
     complex(dp), contiguous, pointer :: ptr(:, :, :)
 
[…]
