+++ src/density/mod_rhoir.f90
@@ -1,24 +1,23 @@
-! REVISION HISTORY:
-!   Created April 2003 (JKD)
-
 module mod_rhoir
 
   use asserts, only: assert
-  use modinput, only: input, isspinspiral
-  use mod_kpoint, only: wkpt
+  use constants, only: real_zero, zzero
+  use m_zfftifc, only: zfftifc
   use mod_eigenvalue_occupancy, only: nstfv, nstsv
-  use modmpi, only: mpi_env_k, distribute_loop
-  use mod_timing, only: timerho
-  use precision, only: dp
-  use constants, only: zzero
-  use svlo, only: get_num_of_basis_functions_sv
-  use mod_spin, only: ncmag, nspinor, ndmag
   use mod_gkvector, only: ngk, igkig
   use mod_Gvector, only: ngrtot, igfft, ngrid
+  use mod_kpoint, only: wkpt
   use mod_lattice, only: omega
-  use m_zfftifc, only: zfftifc
+  use mod_spin, only: ncmag, nspinor, ndmag
+  use mod_timing, only: timerho
+  use modinput, only: input, isspinspiral
+  use modmpi, only: mpi_env_k, distribute_loop
+  use precision, only: dp, i32
+  use svlo, only: get_num_of_basis_functions_sv
 
   implicit none
+
+  private
 
   interface genrhoir
     module procedure :: genrhoir_spin_polarized
@@ -33,19 +32,19 @@ contains
   !> to be used as input arguments for `genrhoir_spin_polarized`
   subroutine genrhoir_non_spin_polarized ( ik, evecfv, occupations, rhoir, magir, evecsv )
     !> k-point number
-    integer, intent (in) :: ik
+    integer(i32), intent (in) :: ik
     !> First-variational eigenvectors (nmatmax, nstfv)
     complex(dp), contiguous, target, intent (in) :: evecfv(:, :)
     !> State occupations (nstsv)
-    real(dp), intent(in) :: occupations(:)
+    real(dp), contiguous, intent(in) :: occupations(:)
     !> Interstitial charge density
-    real(dp), intent(inout) :: rhoir(:)
+    real(dp), contiguous, intent(inout) :: rhoir(:)
     !> Interstitial magnetisation vector field
-    real(dp), optional, intent(inout) :: magir(:, :)
+    real(dp), contiguous, optional, intent(inout) :: magir(:, :)
     !> Second-variational eigenvectors (nstfv, nstsv)
-    complex(dp), optional, intent(in) :: evecsv(:, :)
+    complex(dp), contiguous, optional, intent(in) :: evecsv(:, :)
 
-    integer, parameter :: n_spin = 1
+    integer(i32), parameter :: n_spin = 1
     complex(dp), contiguous, pointer :: ptr(:, :, :)
 
     ptr(1 : size( evecfv, 1 ), 1 : size( evecfv, 2 ), 1 : n_spin) => evecfv
@@ -60,22 +59,22 @@ contains
   !> {\tt wavefmt}, {\tt genshtmat} and {\tt seceqn}.
   subroutine genrhoir_spin_polarized ( ik, evecfv, occupations, rhoir, magir, evecsv )
     !> k-point number
-    integer, intent(in) :: ik
+    integer(i32), intent(in) :: ik
     !> First-variational eigenvectors (nmatmax, nstfv, n_spin)
-    complex(dp), intent (in) :: evecfv (:, :, :)
+    complex(dp), contiguous, intent (in) :: evecfv (:, :, :)
     !> State occupations (nstsv)
-    real(dp), intent(in) :: occupations(:)
+    real(dp), contiguous, intent(in) :: occupations(:)
[…]
