+++ src/src_xs/src_rttddft/rttddft_init.f90
@@ -46,15 +46,15 @@ module rttddft_init
   use rttddft_electric_field, only: Electric_Field
   use rttddft_file_names, only: filename_avec, filename_evec, filename_jind, filename_pvec, RTTDDFT_GND_sufix
   use rttddft_GlobalMDVariables, only: B_past, B_time, mathcalH, mathcalB
-  use rttddft_HamiltonianOverlap, only: update_hamiltonian_without_pa_term_lapw, &
-    update_overlap_lapw, update_hamiltonian_without_pa_term_ks, &
-    add_external_coupling_velocity_gauge, add_external_coupling_berry_phase
+  use rttddft_Hamiltonian, only: add_external_coupling_berry_phase, add_external_coupling_velocity_gauge, &
+    update_hamiltonian_without_pa_term_ks, update_hamiltonian_without_pa_term_lapw
   use rttddft_hybrids, only: hybrids_used, Set_Dimension_mixed_product_basis, set_barecoul_basis
   use rttddft_input, only: rttddft_input_keys
   use rttddft_io, only: file_pmat_exists, read_pmat, write_pmat, file_pmat_mt_exists, &
     read_pmat_mt, write_pmat_mt, write_file_info, write_file_info_fill_line_with_char, &
     get_filename_pmat, get_filename_pmat_mt, read_wavefunction, groundstate, t, t_minus_dt, &
     read_phases
+  use rttddft_Overlap, only: update_overlap_lapw
   use rttddft_pmat, only: obtain_pmat_LAPWloBasis, obtain_pmat_KSBasis
   use rttddft_Polarization, only: Polarization
   use rttddft_potential, only: update_potential
@@ -68,7 +68,7 @@ module rttddft_init
   
   private
   
-  public :: initialize_rttddft
+  public :: initialize_rttddft, initialize_me
 
 contains
 !> This subroutine initializes many variables used in a RT-TDDFT calculation.
@@ -171,12 +171,14 @@ subroutine initialize_rttddft( rt_inp, propagator, vec_pot, a_tot_t_minus_dt, mo
   ! a one shot GS calculation serves no purpose
   if ( .not. hybrids_used() ) call gndstateq( input%xs%vkloff, RTTDDFT_GND_sufix//filext )
 
+  ! Generate k, G, G+k vectors
   call generate_k_vectors( kset_rttddft, bvec, input%groundstate%ngridk, input%xs%vkloff, .false., .false. )
+  call generate_G_vectors( Gset, bvec, intgv, input%groundstate%gmaxvr )
+  call generate_Gk_vectors( Gkset, kset_rttddft, Gset, gkmax )
+  ! Distribute k-points over MPI ranks
   call distribute_loop( mpi_env_k, kset_rttddft%nkpt, first_kpt, last_kpt )
   
   evolve_H0 = ( molecular_dynamics%on .or. ( .not. rt_inp%eeInteraction%use_ipa() ) )
-  if ( (rt_inp%use_ks_basis() .and. evolve_H0) .or. rt_inp%use_berry_phase() ) &
-    call init_me_evaluation( kset_rttddft, Gkset, Gset )
   
   ! Neighbour is a k point from another MPI rank, which is reachable in 1 or 2 jumps
   ! by one of the k points controlled by the current MPI rank
@@ -220,6 +222,9 @@ subroutine initialize_rttddft( rt_inp, propagator, vec_pot, a_tot_t_minus_dt, mo
   if ( rt_inp%use_ks_basis() ) k_dependent_dims = ham_dimension
 
[…]
