+++ src/src_xs/src_rttddft/rttddft_hamiltonian.f90
@@ -2,14 +2,10 @@
 ! See the file COPYING for license details.
 ! Copyright (C) Exciting Code, SOL group. 2020
 
-! HISTORY
-! Created Apr 2019 (Ronaldo)
-! Improved documentation: July 2021 (Ronaldo)
 ! Reference: https://doi.org/10.1088/2516-1075/ac0c26
-
-!> Module that manages the hamiltonian and overlap in RT-TDDFT
-!> //TODO(Ronaldo): Refactor to reduce the number of global variables
-module rttddft_HamiltonianOverlap
+! TODO(Ronaldo): Refactor to reduce the number of global variables
+!> Module that manages the hamiltonian matrix in RT-TDDFT
+module rttddft_Hamiltonian
   use asserts, only: assert
   use constants, only: fourpi, zi, zone, zzero
   use mod_APW_LO, only: apword, nlorb, lorbl
@@ -34,7 +30,7 @@ module rttddft_HamiltonianOverlap
   implicit none
 
   private
-  public :: update_hamiltonian_without_pa_term_lapw, update_overlap_lapw, &
+  public :: update_hamiltonian_without_pa_term_lapw, &
     add_external_coupling_berry_phase, update_hamiltonian_without_pa_term_ks, &
     add_external_coupling_velocity_gauge
     
@@ -42,85 +38,6 @@ module rttddft_HamiltonianOverlap
   type(MTHamiltonianList) :: mt_h
 
 contains
-
-  !> In `update_overlap_lapw`, we obtain the overlap of the (L)APWs at time \( t \).
-  subroutine update_overlap_lapw( first_kpt, a_tot, overlap, apwalm, pmatmt, printTimings, &
-    t_ham, t_MD, update_mathcalH, update_mathcalB )
-    !> The first \( \mathbf{k} \) point
-    integer(i32), intent(in) :: first_kpt
-    !> Total vector potential
-    type(Vector_Potential_Field), intent(in) :: a_tot
-    !> Overlap matrix (of basis functions) (nmatmax, nmatmax, first_kpt : last_kpt)
-    complex(dp), contiguous, intent(out) :: overlap(:, :, first_kpt :)
-    !> Matching coefficients of the (L)APWs
-    !> (ngkmax, apwordmax, lmmaxapw, natmtot, first_kpt : last_kpt)
-    complex(dp), contiguous, intent(in) :: apwalm(:, :, :, :, first_kpt :)
-    !> Muffin-tin part of the Momentum matrix
-    !> (nmatmax, nmatmax, 3, natmtot, first_kpt : last_kpt)
-    complex(dp), contiguous, optional, intent(inout) :: pmatmt(:, :, :, :, first_kpt :)
-    !> Object that packs information about printing of timings
-    type(Print_Timings), optional, intent(in) :: printTimings
-    !> Object that packs information about timings to update the Hamiltonian
-    type(Timing_RTTDDFT_hamiltonian), optional, intent(out) :: t_ham
-    !> Object that packs information about timings related to MD
-    type(Timing_Ehrenfest), optional, intent(out) :: t_MD
-    !> if `.True.`, update `mathcalH`
-    logical, intent(in), optional :: update_mathcalH
-    !> if `.True.`, update `mathcalB`
-    logical, intent(in), optional :: update_mathcalB
-
-    integer(i32) :: ik, last_kpt
-    real(dp) :: ti, tf
-    logical :: timings_general, timings_detailed, get_mathcalH, get_mathcalB
-
[…]

[…]
