### 2025-09-05 3e1433b
New type to encapsulate overlap matrix in RT-TDDFT
