+++ src/src_xs/src_rttddft/rttddft_overlap.f90
@@ -2,15 +2,15 @@
 module rttddft_Overlap
   use asserts, only: assert
   use constants, only: y00, zi, zone, zzero
+  use math_utils, only: all_zero
   use matrix_elements, only: me_mt_alloc, me_mt_prepare, me_mt_mat, me_ir_mat
   use mod_atoms, only: nspecies, natoms, idxas
-  use mod_eigensystem, only: nmat
   use mod_gvector, only: cfunig
   use mod_kpointset, only: Gk_set
   use mod_muffin_tin, only: nrmtmax
   use physical_constants, only: c
   use precision, only: dp, i32
-  use rttddft_GlobalMDVariables, only: mathcalH, mathcalB
+  use rttddft_arrays, only: positive_matrix_set
   use rttddft_timings, only: timesec_RTTDDFT, Timing_RTTDDFT_overlap
   use rttddft_VectorPotential, only: Vector_Potential_Field
 
@@ -19,45 +19,125 @@ module rttddft_Overlap
   private
   integer(i32), parameter :: n_cartesian = 3
 
-  public :: update_overlap_lapw
-
+  type, public, extends(positive_matrix_set) :: overlap_set
+    private
+    logical :: identity = .false.
+  contains
+    private
+    procedure, public :: allocate => overlap_set_allocate
+    procedure, public :: calculate => overlap_set_calculate
+    procedure, public :: initialize => overlap_set_initialize
+    procedure, public :: is_identity => overlap_set_is_identity
+    procedure, public :: is_not_identity => overlap_set_is_not_identity
+    procedure :: initialize_as_identity => overlap_set_initialize_as_identity
+    final :: destructor
+  end type
+    
 contains
-  !> In `update_overlap_lapw`, we obtain the overlap of the (L)APWs at time \( t \).
-  subroutine update_overlap_lapw( first_kpt, overlap, apwalm, Gkset, t_overlap, pmatmt, &
-    a_tot, update_mathcalH, update_mathcalB )
-    !> The first \( \mathbf{k} \) point
-    integer(i32), intent(in) :: first_kpt
-    !> Overlap matrix (of basis functions) (nmatmax, nmatmax, first_kpt : last_kpt)
-    complex(dp), contiguous, intent(out) :: overlap(:, :, first_kpt :)
-    !> Matching coefficients of the (L)APWs (ngkmax, apwordmax, lmmaxapw, natmtot, first_kpt : last_kpt)
-    complex(dp), contiguous, intent(in) :: apwalm(:, :, :, :, first_kpt :)
-    !> Set of G+k vectors for LAPW expansion
+  !> Wrapper for calling `this%allocate_array`
+  pure subroutine overlap_set_allocate( this, use_lapwlo_basis, m, ki, kf )
+    class(overlap_set), intent(inout) :: this
+    !> If `.true.`, LAPW+LO basis is used (allocate matrix). Otherwise, overlap is identity
+    logical, intent(in) :: use_lapwlo_basis
+    !> Overlap matrix dimension
+    integer(i32), intent(in) :: m
+    !> First \( \mathbf{k} \)-point
+    integer(i32), intent(in) :: ki
+    !> Last \( \mathbf{k} \)-point
+    integer(i32), intent(in) :: kf
+
+    this%identity = .not. use_lapwlo_basis 
+    call this%allocate_array( [1, 1 , ki], [m, m, kf] )
+  end subroutine
+
+  pure elemental logical function overlap_set_is_identity( this )
+    class(overlap_set), intent(in) :: this
+
+    overlap_set_is_identity = this%identity
[…]
