+++ src/lapack_wrappers/multiplication/add_scaled_array_test.f90
@@ -1,0 +1,118 @@
+!> Module with tests for the [[add_scaled_array]] module
+module add_scaled_array_test
+  use add_scaled_array, only: scaled_add
+  use constants, only: zi, zone, zzero
+  use exciting_mpi, only: mpiinfo
+  use math_utils, only: all_close
+  use mock_arrays, only: complex_vector_5, complex_vector_7, complex_matrix_5x7, &
+    complex_matrix_7x5, complex_matrix_5x5
+  use precision, only: dp, i32
+  use to_char_conversion, only: to_char
+  use unit_test_framework, only : unit_test_type
+
+  implicit none
+
+  private
+
+  public :: add_scaled_array_test_driver
+
+contains
+  subroutine add_scaled_array_test_driver(mpiglobal, kill_on_failure)
+    !> mpi information
+    type(mpiinfo), intent(in) :: mpiglobal
+    !> Kill the program before the test driver finishes
+    !> if an assertion fails
+    logical, optional, intent(in) :: kill_on_failure
+    
+    type(unit_test_type) :: test_report
+    integer(i32), parameter :: n_assertions_test_scaled_add_vector = 3
+    integer(i32), parameter :: n_assertions_test_scaled_add_matrix = 3
+    integer(i32), parameter :: n_assertions = n_assertions_test_scaled_add_vector + &
+                                              n_assertions_test_scaled_add_matrix
+
+    character(len=*), parameter :: module_tested = "add_scaled_array"
+
+    ! Initialize test object
+    call test_report%init(n_assertions, mpiglobal)
+
+    ! Run and assert tests
+    call test_scaled_add_vector( test_report )
+    call test_scaled_add_matrix( test_report )
+
+    call test_report%report( module_tested, kill_on_failure )
+
+    call test_report%finalise()
+
+  end subroutine
+
+  subroutine test_scaled_add_vector( test_report )
+    !> The test object
+    type(unit_test_type), intent(inout) :: test_report
+
+    integer(i32) :: test_number
+    character(len=*), parameter :: test_id = "test_scaled_add_vector"
+    complex(dp) :: a
+    complex(dp), allocatable :: x(:), y(:), y_expected(:)
+    
+    test_number = 1
+    a = cmplx(1.0_dp, 2.0_dp, dp)
+    x = complex_vector_5
+    y = complex_vector_7(1:size(x))
+    y_expected = a*x + y
+    call scaled_add( a, x, y )
+    call test_report%assert( all_close( y, y_expected ), report_message( test_id, "complex", test_number ) )
+
+    test_number = test_number + 1
+    call scaled_add( zzero, x, y )
+    call test_report%assert( all_close( y, y_expected ), report_message( test_id, "complex", test_number ) )
+
+    test_number = test_number + 1
+    y = complex_vector_7(1:size(x))
+    y_expected = x + y
+    call scaled_add( zone, x, y )
+    call test_report%assert( all_close( y, y_expected ), report_message( test_id, "complex", test_number ) )
+  end subroutine
+
+  subroutine test_scaled_add_matrix( test_report )
+    !> The test object
+    type(unit_test_type), intent(inout) :: test_report
+
+    integer(i32) :: test_number
+    character(len=*), parameter :: test_id = "test_scaled_add_matrix"
+    complex(dp) :: a
+    complex(dp), allocatable :: x(:, :), y(:, :), y_expected(:, :)
+    
+    test_number = 1
+    a = cmplx(1.0_dp, 2.333_dp, dp)
+    x = complex_matrix_5x5
+    y = reshape( complex_matrix_7x5, shape(x) )
+    y_expected = a*x + y
+    call scaled_add( a, x, y )
+    call test_report%assert( all_close( y, y_expected ), report_message( test_id, "complex", test_number ) )
+
[…]
