### 2025-09-06 43b0e67
RT hamiltonian refactor
