+++ external/deviceacc/external/hipfort/hipfort_hiprand_enums.f90
@@ -25,51 +25,39 @@
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
           
            
-module hipfort_rocfft_enums
+module hipfort_hiprand_enums
   implicit none
 
   enum, bind(c)
-    enumerator :: rocfft_status_success
-    enumerator :: rocfft_status_failure
-    enumerator :: rocfft_status_invalid_arg_value
-    enumerator :: rocfft_status_invalid_dimensions
-    enumerator :: rocfft_status_invalid_array_type
-    enumerator :: rocfft_status_invalid_strides
-    enumerator :: rocfft_status_invalid_distance
-    enumerator :: rocfft_status_invalid_offset
-    enumerator :: rocfft_status_invalid_work_buffer
+    enumerator :: HIPRAND_STATUS_SUCCESS = 0
+    enumerator :: HIPRAND_STATUS_VERSION_MISMATCH = 100
+    enumerator :: HIPRAND_STATUS_NOT_INITIALIZED = 101
+    enumerator :: HIPRAND_STATUS_ALLOCATION_FAILED = 102
+    enumerator :: HIPRAND_STATUS_TYPE_ERROR = 103
+    enumerator :: HIPRAND_STATUS_OUT_OF_RANGE = 104
+    enumerator :: HIPRAND_STATUS_LENGTH_NOT_MULTIPLE = 105
+    enumerator :: HIPRAND_STATUS_DOUBLE_PRECISION_REQUIRED = 106
+    enumerator :: HIPRAND_STATUS_LAUNCH_FAILURE = 201
+    enumerator :: HIPRAND_STATUS_PREEXISTING_FAILURE = 202
+    enumerator :: HIPRAND_STATUS_INITIALIZATION_FAILED = 203
+    enumerator :: HIPRAND_STATUS_ARCH_MISMATCH = 204
+    enumerator :: HIPRAND_STATUS_INTERNAL_ERROR = 999
+    enumerator :: HIPRAND_STATUS_NOT_IMPLEMENTED = 1000
   end enum
 
   enum, bind(c)
-    enumerator :: rocfft_transform_type_complex_forward
-    enumerator :: rocfft_transform_type_complex_inverse
-    enumerator :: rocfft_transform_type_real_forward
-    enumerator :: rocfft_transform_type_real_inverse
-  end enum
-
-  enum, bind(c)
-    enumerator :: rocfft_precision_single
-    enumerator :: rocfft_precision_double
-  end enum
-
-  enum, bind(c)
-    enumerator :: rocfft_placement_inplace
-    enumerator :: rocfft_placement_notinplace
-  end enum
-
-  enum, bind(c)
-    enumerator :: rocfft_array_type_complex_interleaved
-    enumerator :: rocfft_array_type_complex_planar
-    enumerator :: rocfft_array_type_real
-    enumerator :: rocfft_array_type_hermitian_interleaved
-    enumerator :: rocfft_array_type_hermitian_planar
-    enumerator :: rocfft_array_type_unset
-  end enum
-
-  enum, bind(c)
-    enumerator :: rocfft_exec_mode_nonblocking
-    enumerator :: rocfft_exec_mode_nonblocking_with_flush
-    enumerator :: rocfft_exec_mode_blocking
+    enumerator :: HIPRAND_RNG_TEST = 0
[…]
