### 2025-09-12 ad05258
Add utility function for species symbol extraction
