+++ src/src_xs/src_rttddft/rttddft_sanity_checks.f90
@@ -30,16 +30,18 @@ contains
     type(input_type), intent(in) :: inp
 
     integer(i32) :: i
+    character(len=*), parameter :: procedure_name = "check_rttddft_input"
+    character(len=*), parameter :: warning_header = "Warning(" // procedure_name // "): "
     
     call terminate_if_false( .not. inp%groundstate%solver%packedmatrixstorage, &
-      & 'Error: RT-TDDFT does not work with matrices stored in a packed form.' )
+      & 'RT-TDDFT does not work with matrices stored in a packed form.' )
 
     ! Consistency check: check if no spin polarized calculations are requested.
     call terminate_if_false( .not. inp%groundstate%tevecsv, &
-      & 'Error: only spin unpolarised calculations are possible with RT-TDDFT now.' )
+      & 'Only spin unpolarised calculations are possible with RT-TDDFT now.' )
     ! iora*
     if( trim(inp%groundstate%ValenceRelativity) == "iora*" ) call terminate( &
-      & 'Error: RT-TDDFT not implemented for ValenceRelativity="iora*"'   )
+      & 'RT-TDDFT not implemented for ValenceRelativity="iora*"'   )
 
     ! Consistency check: laser has been defined?
     call terminate_if_false( associated( inp%xs%realTimeTDDFT%laser ), &
@@ -92,10 +94,15 @@ contains
       associate( kick_array => inp%xs%realTimeTDDFT%laser%kickarray )
         do i = 1, size( kick_array )
           if ( abs( kick_array(i)%kick%width ) < eps_kick_width ) &
-            call warning( 'Warning: electric field is ill-defined at time ' // &
+            call warning( warning_header // 'electric field is ill-defined at time ' // &
             to_char( real( kick_array(i)%kick%t0, sp) ) // ' for the kick number ' // to_char( i ) )
         end do
       end associate
+    end if
+
+    if ( inp%xs%realTimeTDDFT%numberOfFrozenStates > 0 .and. trim( inp%xs%realTimeTDDFT%fieldCoupling ) == "velocityGauge"  ) then
+      if ( .not. inp%xs%realTimeTDDFT%subtractJ0 ) call warning( warning_header // &
+        'it is recommended to set subtractJ0="true" with velocity gauge if frozen states are present.'  )
     end if
 
   end subroutine
@@ -127,6 +134,8 @@ contains
       e_field_max_magnitude
     type(Electric_Field) :: e_aux
     integer(i32) :: i, j
+    character(len=*), parameter :: procedure_name = "check_rttddft_setup"
+    character(len=*), parameter :: warning_header = "Warning(" // procedure_name // "): "
 
     do i = 1, n_cartesian_directions
       lattice_vector_norm(i) = norm( lattice_vectors(:, i) )
@@ -143,8 +152,8 @@ contains
       end do
     end do
 
-    if ( all( e_field_max_lattice < eps_e_field ) ) call warning( "Warning(check_rttddft_setup): &
-      external field amplitude is zero." )
+    if ( all( e_field_max_lattice < eps_e_field ) ) call warning( warning_header // &
+      "external field amplitude is zero." )
 
     if ( use_berry_phase ) then
       call terminate_if_false( energy_gap > eps_energy_gap, " &
@@ -156,22 +165,22 @@ contains
 
         e_field_critical(i) = energy_gap / ( real( k_grid_dimensions(i), dp ) * &
           sqrt( dot_product( lattice_vectors(:, i), lattice_vectors(:, i) ) ) )
-        if ( e_field_max_lattice(i) > e_field_critical(i) ) call warning( 'Warning(check_rttddft_setup): &
-          field strength ' // to_char( real( e_field_max_lattice(i), sp) ) // ' in lattice direction ' &
[…]
