### 2025-09-26 cb826af
New type to encapsulate hamiltonian matrix in RT-TDDFT
