+++ src/src_xs/src_rttddft/rttddft_main.f90
@@ -29,10 +29,9 @@ module rttddft_main
   use rttddft_Density, only: update_density, ground_state
   use rttddft_file_names, only: RTTDDFT_suffix
   use rttddft_electric_field, only: Electric_Field, obtain_electric_field
-  use rttddft_Energy, only: TotalEnergy, obtain_energy_rttddft
-  use rttddft_GlobalMDVariables
-  use rttddft_Hamiltonian, only: add_external_coupling_berry_phase, add_external_coupling_velocity_gauge, &
-    update_hamiltonian_without_pa_term_ks, update_hamiltonian_without_pa_term_lapw
+  use rttddft_Energy, only: Total_Energy
+  use rttddft_GlobalMDVariables, only: B_past, B_time, mathcalB, update_exciting_globals_for_new_ions_positions
+  use rttddft_Hamiltonian, only: hamiltonian_set
   use rttddft_init, only: initialize_me, initialize_rttddft
   use rttddft_input, only: rttddft_input_keys
   use rttddft_io, only: open_files_vector_fields, close_files_vector_fields, read_vector_field, write_vector_field, &
@@ -81,24 +80,19 @@ contains
   !>    3. using these wavefunctions, determine the density, the potential, and the hamiltonian.
   !> 4. Evolve the wavefunctions, the density and the hamiltonian using the desired time step.
   subroutine coordinate_rttddft_calculation()
-    ! Basis-expansion coefficients of the KS-WFs
     class(wavefunction_set), allocatable :: psi
-    ! Overlap and Hamiltonian matrices (nmatmax, nmatmax, first_kpt : last_kpt)
-    complex(dp), allocatable :: ham_time(:, :, :), berry_coupling_term(:, :, :), &
-      ham_past(:, :, :), ham_init(:, :, :), effective_potential_init(:, :, :)
+    complex(dp), allocatable :: berry_coupling_term(:, :, :)
     type(overlap_set) :: overlap
+    type(hamiltonian_set) :: H
     ! Matching coefficients of the (L)APWs: (ngkmax, apwordmax, lmmaxapw, natmtot, first_kpt : last_kpt)
     complex(dp), allocatable :: apwalm(:, :, :, :, :)
-    ! Momentum matrix elements (ham_dimension, ham_dimension, 3, first_kpt : last_kpt)
-    ! and (nmatmax, nmatmax, 3, natmtot, first_kpt : last_kpt)
+    ! Momentum matrix elements 
     complex(dp), allocatable :: pmat(:, :, :, :), pmatmt(:, :, :, :, :)
     ! Electron density (lmmaxvr, nrmtmax, natmtot) and (ngrtot)
     real(dp), allocatable :: rhomt_frozen(:, :, :), rhomt_init(:, :, :), &
       rhoir_frozen(:), rhoir_init(:)
     ! Initial occupations and energies array (nstates, first_kpt : last_kpt)
-    real(dp), allocatable :: occupations(:, :), initial_ks_energies(:, :)
-    ! k-dependent Hamiltonian's dimensions array (first_kpt : last_kpt)
-    integer(i32), allocatable :: k_dependent_dims(:)
+    real(dp), allocatable :: occupations(:, :)
     ! KS-LAPW+lo transition matrix (nmatmax, nstfv, first_kpt : last_kpt)
     complex(dp), allocatable :: ks_lapwlo_transition_matrix(:, :, :)
     ! Planewave matrix elements between neighbouring k points
@@ -109,11 +103,11 @@ contains
     integer(i32) :: it, first_kpt, last_kpt, first_step, last_step, i_print, &
       time_step_multiplier, lmax_potential
     logical :: pred_corr_reached_max_steps, my_rank_writes_to_output, &
[…]

[…]
