+++ src/src_xs/src_rttddft/rttddft_wavefunction_test.f90
@@ -3,15 +3,13 @@ module rttddft_Wavefunction_test
   use exciting_mpi, only: mpiinfo
   use math_utils, only: all_close
   use mock_arrays, only: complex_positive_definite_matrix_5x5, complex_hermitian_matrix_5x5, complex_matrix_5x5
+  use mod_kpointset, only: k_set
   use precision, only: dp, i32
-  use constants, only: zone, zi, sqrt_two
-  use exciting_mpi, only: mpiinfo
-  use math_utils, only: all_close
   use rttddft_Overlap, only: overlap_set
   use rttddft_Wavefunction, only: obtain_occupations, wavefunction_set, initialize_wavefunction_set
+  use to_char_conversion, only: to_char
   use unit_test_framework, only : unit_test_type
   use xlapack, only: solve_generalized_hermitian_eigenproblem
-  use to_char_conversion, only: to_char
 
   implicit none
 
@@ -78,7 +76,7 @@ contains
                                       -(0.0_dp, 0.6_dp),  (0.0_dp, 0.8_dp), (0.0_dp, 0.0_dp), (0.0_dp, 0.0_dp), & !2nd kpt
                                        (0.0_dp, 0.0_dp),  (0.0_dp, 0.0_dp),      zi/sqrt_two,    zone/sqrt_two],& !2nd kpt
                                       [n_states_gnd, n_states, n_kpt] )
-    real(dp), allocatable :: occ(:, :)                                                                  
+    real(dp), allocatable :: occ(:, :) 
 
     call obtain_occupations( proj, occ_gnd, occ )
     call test_report%assert( all( shape(occ) == shape(occ_expected) ), &
@@ -96,24 +94,26 @@ contains
 
     character(len=*), parameter :: test_identifier = "test_obtain_number_excitations"
     integer(i32), parameter :: n_states_gnd = 5
-    integer(i32), parameter :: n_states = 3
+    integer(i32), parameter :: n_states = 3 ! occupied + empty states
     integer(i32), parameter :: n_frozen = 1
-    integer(i32), parameter :: i_VBM = 2
-    integer(i32), parameter :: i_CBm = i_VBM + 1
+    integer(i32), parameter :: i_VBM = 2 
+    integer(i32), parameter :: i_CBm = i_VBM + 1 
     integer(i32), parameter :: n_dim = size( complex_positive_definite_matrix_5x5, 1 )
     integer(i32) :: ik, n_kpt, n_kpt_per_rank, first_k, last_k, test_counter
-    real(dp) :: n_exc, n_gs, n_exc_ref, n_gs_ref, eigs_gnd(n_states_gnd), eigs(n_states)
+    real(dp) :: n_exc, n_gs, n_exc_ref, n_gs_ref, eigs_gnd(n_states_gnd), eigs(n_frozen + n_states)
     real(dp), allocatable :: wkpt(:), occ_gnd(:, :), occ(:)
     complex(dp), allocatable :: H_gnd(:, :), H(:, :), S_aux(:, :), aux(:, :), overlap(:, :, :)
     type(overlap_set) :: S
     complex(dp), allocatable :: psi_gnd(:, :, :), tmp(:, :), proj(:, :), psi_t(:, :, :)
     class(wavefunction_set), allocatable :: psi, psi_no_frozen
+    type(k_set) :: kset
 
     ! Initialization
     n_kpt_per_rank = 2
     n_kpt = n_kpt_per_rank*( mpiglobal%procs )
[…]
