### 2025-09-29 32ea977
Address the review by @ruiyiQM: - Format the equation between line 62 and 63 - Restructure Results part to explain the generation of the   interpolation grids in detail. Introduce a new figure to present   results.
