+++ paper/paper.md
@@ -55,7 +55,7 @@ u^*_{o'\mathbf{k}'}(\mathbf{r}')
 $$
 where $W_{\mathbf{k-k'}}(\mathbf{r}, \mathbf{r}')$ is the statically screened Coulomb potential. $u_{i\mathbf{k}}(\mathbf{r})$ is the periodic part of the one-particle wavefunction of state $i$ at the reciprocal lattice point $\mathbf{k}$. In general, we cannot see, which of the matrix elements will be zero, thus we need to compute all of them. Therefore, setting up the full BSH scales with $\mathcal{O}(N_o^2 N_u^2 N_{\mathbf{k}}^2)$ and diagonalizing it directly with $\mathcal{O}(N_o^3 N_u^3 N_{\mathbf{k}}^3)$. 
 	
-The primary scaling bottleneck stems from evaluating the matrix elements of the interaction kernels. To mitigate this, we reformulate the wavefunction products using Interpolative Separable Density Fitting (ISDF) [@Lu:2015]. Specifically, we approximate these products on a discrete real-space grid, $\{\mathbf{r}\}$, by expressing them as superpositions of values evaluated on a smaller interpolation grid, $\{\mathbf{r}_\mu\} \subset \{\mathbf{r}\}$:
+The primary scaling bottleneck stems from evaluating the matrix elements of the interaction kernels. To mitigate this, we reformulate the wavefunction products using Interpolative Separable Density Fitting (ISDF) [@Lu:2015,@Lu:2016]. Specifically, we approximate these products on a discrete real-space grid, $\{\mathbf{r}\}$, by expressing them as superpositions of values evaluated on a smaller interpolation grid, $\{\mathbf{r}_\mu\} \subset \{\mathbf{r}\}$:
 $$
 u_{i\mathbf{k}}^*(\mathbf{r})u_{j\mathbf{k}'}(\mathbf{r}) \approx \sum_{\mu=1}^{N_\mu} \zeta_\mu(\mathbf{r}) u_{i\mathbf{k}}^*(\mathbf{r}_\mu) u_{j\mathbf{k}}(\mathbf{r}_\mu) \:,
 $$
@@ -102,22 +102,23 @@ $$
 where we first compute the sums over $o'$, $u'$, and $\mathbf{k}'$ to get a term that depends only on $\mathbf{r}_\nu^V$ with a complexity of $\mathcal{O}(N_\mu^V(N_o N_u N_{\bf k} + N_uN_{\bf k})$. The remaining sums can be computed with $\mathcal{O}((N_\mu^V)^2 N_\mu^V N_o N_u N_{\bf k})$, so the complexity of computing $V\cdot X$ is bounded by $\mathbf{O}((N_\mu^V)^2 + N_\mu N_o N_u N_{\bf k})$.
 Applying the screened kernel to $X$, after reordering the sums, we get
 
-
 $$
 [W \cdot X]_{ou\mathbf k} = \frac{1} {N_k} \sum_{\nu=1}^{N_\mu^{W_o}}
 u_{o\mathbf k}(\mathbf r_\nu^{W_o})
-\left\{ \sum_{\mu=1}^{N_\mu^{W_u}}
+\Bigg\{ \sum_{\mu=1}^{N_\mu^{W_u}}
 u^*_{u\mathbf k}(\mathbf r_\mu^{W_u})
-\left[
-\sum_{\mathbf{k'}}
-\tilde{W}_{\mu\nu, \mathbf{k-k'}}
-\left(\sum_{u'}
[…]
