+++ paper/paper.bib
@@ -71,3 +71,21 @@
   year = {2017},
   journal = {Journal of Chemical Theory and Computation}
 }
+
+@article{Lu:2016,
+	title = {Fast algorithm for periodic density fitting for Bloch waves},
+	volume = {1},
+	issn = {2380-2898},
+	url = {https://link.intlpress.com/JDetail/1805564664274690050},
+	doi = {10.4310/AMSA.2016.v1.n2.a3},
+	abstract = {International Press of Boston - publishers of scholarly mathematical and scientific journals and books},
+	pages = {321--339},
+	number = {2},
+	journaltitle = {Annals of Mathematical Sciences and Applications},
+	shortjournal = {Annals of Math. Sci. and Appl.},
+	author = {Lu, Jianfeng and Ying, Lexing},
+	urldate = {2025-08-20},
+	date = {2016-07-25},
+	note = {Publisher: International Press of Boston},
+	file = {Submitted Version:/home/benedikt/Zotero/storage/HIKQ34SN/Lu and Ying - 2016 - Fast algorithm for periodic density fitting for Bloch waves.pdf:application/pdf},
+}
