### 2025-09-30 65350ac
fix exciting format in the last paragraph
