+++ paper/paper.md
@@ -150,6 +150,6 @@ In Fig.(\autoref{fig:error}) we compare the exciton binding energies and spectra
 In Fig.(\autoref{fig:scaling}) we show the wall times for solving the BSE with the new and direct implementations for increasing $N_\mathbf{k}$. The new algorithm massively outperforms the direct one, and the speedup increases more than linearly with $N_\mathbf{k}$. We also show the wall times for computing the RPA screening with increasing $N_\mathbf{k}$, which is now clearly the bottleneck in solving the BSE.
 ![Runtimes of the direct (black) and new (red) BSE implementations and the RPA screening (blue) as a function of $N_\mathbf{k}$. The speedup of the algorithm is shown by the gray dashed line.\label{fig:scaling}](Diamond_ngridk_scaling.png)
 
-Altogether, we have implemented a new, low-scaling BSE solver and fully integrated it into the all-electron, full-potential solver `exciting`. We demonstrate that the new implementation yields results equivalent to the direct solution of the BSE but with significantly reduced computational time. Consequently, it enables more precise calculations and facilitates the study of more complex problems.
+Altogether, we have implemented a new, low-scaling BSE solver and fully integrated it into the all-electron, full-potential package <span style="font-family: 'Courier New', monospace; font-weight: bold;">exciting</span>. We demonstrate that the new implementation yields results equivalent to the direct solution of the BSE but with significantly reduced computational time. Consequently, it enables more precise calculations and facilitates the study of more complex problems.
 
 # References
