### 2025-10-01 831afde
RT-TDDFT: update method to evaluate the number of excitations
