+++ src/src_xs/src_rttddft/rttddft_wavefunction_test.f90
@@ -1,15 +1,17 @@
 module rttddft_Wavefunction_test
-  use constants, only: zone, zi, sqrt_two
+  use constants, only: zzero,zone, zi, sqrt_two
   use exciting_mpi, only: mpiinfo
   use math_utils, only: all_close
-  use mock_arrays, only: complex_positive_definite_matrix_5x5, complex_hermitian_matrix_5x5, complex_matrix_5x5
+  use mock_arrays, only: complex_hermitian_matrix_5x5, complex_matrix_7x5, &
+    complex_positive_definite_matrix_5x5, complex_unitary_matrix_5x5
   use mod_kpointset, only: k_set
+  use normalize, only: norm_squared_with_positive_matrix
   use precision, only: dp, i32
   use rttddft_Overlap, only: overlap_set
   use rttddft_Wavefunction, only: obtain_occupations, wavefunction_set, initialize_wavefunction_set
   use to_char_conversion, only: to_char
   use unit_test_framework, only : unit_test_type
-  use xlapack, only: solve_generalized_hermitian_eigenproblem
+  use xlapack, only: solve_generalized_hermitian_eigenproblem, norm
 
   implicit none
 
@@ -29,8 +31,10 @@ contains
     logical, optional, intent(in) :: kill_on_failure
     
     type(unit_test_type) :: test_report
-    integer(i32), parameter :: n_assertions_test_obtain_occupations = 8
-    integer(i32), parameter :: n_assertions = n_assertions_test_obtain_occupations
+    integer(i32), parameter :: n_assertions_test_obtain_occupations = 2
+    integer(i32), parameter :: n_assertions_test_obtain_number_excitations = 10
+    integer(i32), parameter :: n_assertions = n_assertions_test_obtain_occupations + &
+                                              n_assertions_test_obtain_number_excitations
 
     character(len=*), parameter :: module_tested = 'rttddft_Wavefunction'
 
@@ -94,21 +98,22 @@ contains
 
     character(len=*), parameter :: test_identifier = "test_obtain_number_excitations"
     integer(i32), parameter :: n_states_gnd = 5
-    integer(i32), parameter :: n_states = 3 ! occupied + empty states
+    integer(i32), parameter :: n_states = 3 ! occupied states
     integer(i32), parameter :: n_frozen = 1
     integer(i32), parameter :: i_VBM = 2 
-    integer(i32), parameter :: i_CBm = i_VBM + 1 
-    integer(i32), parameter :: n_dim = size( complex_positive_definite_matrix_5x5, 1 )
-    integer(i32) :: ik, n_kpt, n_kpt_per_rank, first_k, last_k, test_counter
-    real(dp) :: n_exc, n_gs, n_exc_ref, n_gs_ref, eigs_gnd(n_states_gnd), eigs(n_frozen + n_states)
-    real(dp), allocatable :: wkpt(:), occ_gnd(:, :), occ(:)
+    integer(i32), parameter :: i_CBm = i_VBM + 1 ! metallic system: iCBm = n_states
+    integer(i32), parameter :: n_dim = 7
+    integer(i32) :: ik, j, n_kpt, n_kpt_per_rank, first_k, last_k, test_counter
[…]
