### 2025-10-06 #8866 @EderGio
@maurerben
The present work presents a fast implementation of BSE within the exciting code based. The goal is to enable better scaling of BSE calculations, particularly as the number of k-points increases in a calculation. The authors utilize Interpolative Separable Density Fitting  (ISDF) to extrapolate calculation kernels onto a finite grid. Coupled with the Lanczos algorithm, they can rearrange sums and exploit the separable properties of their ISDF kernels. This is a very impressive piece of work; I support the manuscript for publication after my brief comments are addressed.

I compiled the code and calculated the Diamond BSE spectra using the “regular” (direct) solver and the fastBSE solver in exciting. I used the following k-grids: 222, 444, and 888. The direct solver can do these calculations in 8 seconds, 54 seconds and 1 hour 17 minutes 17 seconds, respectively. The fastBSE solver can do the same in 9 seconds, 23 seconds and 4 minutes 26 seconds, respectively. The fastBSE solver is drastically faster at high k-point densities, as similarly shown in the figures of the manuscript. I could not compare the outputs directly; the output files from using each solver seems different in relation to the exciton energies in terms of formatting and number of files. I am not very familiar with the exicting code base, but comparing the calculated C K-edge spectra would be a good way to show you get the same result in each solver.

Minor comments:
-	Authors should add figure labels and captions. Currently, the main text refers to “Fig.(??)” in all instances when referencing a figure.  Captions should describe axes as well.
-	Schrödinger is misspelled in line 19.

Apologies for the delay getting the review. Summer travel and preparation for the new semester got the better of me.
