### 2025-10-08 aff3153
LAPACK eigensolver in the GPU
