+++ external/deviceacc/external/hipfort/hipfort_rocsolver.F90
@@ -22395,6 +22395,115 @@ module hipfort_rocsolver
       rocsolver_zhegv_rank_1
 #endif
   end interface
+
+  !> @brief HEGVX computes a subset of the eigenvalues and optionally the corresponding
+  !> eigenvectors of a complex generalized Hermitian-definite eigenproblem.
+  !>
+  !> The problem solved is either A*x = lambda*B*x, A*B*x = lambda*x, or B*A*x = lambda*x,
+  !> depending on the value of ITYPE. Eigenvectors are computed depending on EVECT.
+  !>
+  !> @details This function computes all the eigenvalues, eigenvalues in a half-open interval (VL, VU],
+  !> or the IL-th through IU-th eigenvalues, depending on ERANGE. If EVECT = rocblas_evect_original,
+  !> the corresponding eigenvectors are also computed.
+  !>
+  !> @param[in] HANDLE  Handle to the rocBLAS library context.
+  !> @param[in] ITYPE   Specifies the form of the generalized eigenproblem.
+  !> @param[in] EVECT   Specifies whether eigenvectors are computed.
+  !> @param[in] ERANGE  Specifies the type of range of eigenvalues to compute.
+  !> @param[in] UPLO    Specifies whether upper or lower parts of A and B are stored.
+  !> @param[in] N       The order of matrices A and B.
+  !> @param[in,out] A   On entry, the matrix A. On exit, contents are destroyed.
+  !> @param[in] LDA     Leading dimension of A.
+  !> @param[in,out] B   On entry, Hermitian positive definite matrix B. On exit, triangular factor from POTRF.
+  !> @param[in] LDB     Leading dimension of B.
+  !> @param[in] VL      Lower bound of search interval. Ignored if ERANGE != rocblas_erange_value.
+  !> @param[in] VU      Upper bound of search interval. Ignored if ERANGE != rocblas_erange_value.
+  !> @param[in] IL      Index of smallest eigenvalue to compute. Ignored if ERANGE != rocblas_erange_index.
+  !> @param[in] IU      Index of largest eigenvalue to compute. Ignored if ERANGE != rocblas_erange_index.
+  !> @param[in] ABSTOL  Absolute tolerance for eigenvalues. If negative, machine epsilon * ||T||_1 used.
+  !> @param[out] NEV    Total number of eigenvalues found.
+  !> @param[out] W      Array to hold computed eigenvalues.
+  !> @param[out] Z      Array to hold computed eigenvectors (if EVECT != none).
+  !> @param[in] LDZ     Leading dimension of Z.
+  !> @param[out] IFAIL  Array indicating indices of eigenvectors that failed to converge.
+  !> @param[out] INFO   Info on success or failure of computation.
+
+  interface rocsolver_chegvx
+    function rocsolver_chegvx_(handle, itype, evect, erange, uplo, n, A, lda, &
+               B, ldb, vl, vu, il, iu, abstol, nev, W, Z, ldz, ifail, info) bind(c, name="rocsolver_chegvx")
+      use iso_c_binding
+      use hipfort_rocsolver_enums
+      use hipfort_rocblas_enums
+      implicit none
+      integer(kind(rocblas_status_success)) :: rocsolver_chegvx_
+      type(c_ptr), value :: handle
+      integer(kind(rocblas_eform_ax)), value :: itype
+      integer(kind(rocblas_evect_original)), value :: evect
+      integer(kind(rocblas_erange_all)), value :: erange
+      integer(kind(rocblas_fill_upper)), value :: uplo
+      integer(c_int), value       :: n
[…]
