### 2025-10-09 ac7d472
Restore intel gpu
