### 2025-10-09 a562537
Print error message if a regression test failed
