+++ test/test_functions/test_class_testresults.py
@@ -118,12 +118,13 @@ def test_attributes_successful(successful_test_dir):
     Test the attributes of class TestResults, for a successful test case.
     """
     gs_files = ['INFO.OUT', 'evalcore.xml', 'geometry.xml', 'eigval.xml', 'atoms.xml']
-    test_results = ExTestResults(successful_test_dir, completed=True, timing=0.0)
+    test_results = ExTestResults(successful_test_dir, completed=True, err_msg="", timing=0.0)
     test_results_dict = get_test_results(successful_test_dir, gs_files)
     test_results.set_results(test_results_dict)
 
     assert os.path.basename(test_results.test_name) == "LDA_VWN-He_passing", "Base test name"
     assert test_results.completed, "exciting run completed (rather than crashed)"
+    assert test_results.err_msg == "", "Expect no error message"
     assert set(test_results.file_names) == set(gs_files), "Expect these ground state files to be regression-tested"
     assert set(test_results.succeeded_files) == set(gs_files), "All tested files should succeed"
     assert len(test_results.files_with_errors) == 0, "All tested files should succeed"
@@ -135,7 +136,7 @@ def test_print_successful(successful_test_dir):
     Test TestResults print() method on a successful test case.
     """
     gs_files = ['INFO.OUT', 'evalcore.xml', 'geometry.xml', 'eigval.xml', 'atoms.xml']
-    test_results = ExTestResults(successful_test_dir, completed=True, timing=0.0)
+    test_results = ExTestResults(successful_test_dir, completed=True, err_msg="", timing=0.0)
     test_results_dict = get_test_results(successful_test_dir, gs_files)
 
     for name, file_results in test_results_dict.items():
@@ -177,7 +178,7 @@ def test_attributes_erroneous_scalars(failing_scalars_test_dir):
         message = "If not ErrorFinder, file_results returns Failure => " + name + " missing"
         assert isinstance(file_results, ErrorFinder), message
 
-    test_results = ExTestResults(failing_scalars_test_dir, completed=True, timing=0.0)
+    test_results = ExTestResults(failing_scalars_test_dir, completed=True, err_msg="", timing=0.0)
     test_results.set_results(test_results_dict)
 
     assert os.path.basename(test_results.test_name) == "LDA_VWN-He_failing_scalars", "Base test name"
@@ -200,7 +201,7 @@ def test_print_erroneous_scalars(failing_scalars_test_dir):
         message = "If not ErrorFinder, file_results returns Failure => " + name + " missing"
         assert isinstance(file_results, ErrorFinder), message
 
-    test_results = ExTestResults(failing_scalars_test_dir, completed=True, timing=0.0)
+    test_results = ExTestResults(failing_scalars_test_dir, completed=True, err_msg="", timing=0.0)
     test_results.set_results(test_results_dict)
 
     @redirect_stdout
@@ -253,7 +254,7 @@ def test_attributes_failing_array(failing_arrays1d2d_test_dir):
         message = "If not ErrorFinder, file_results returns Failure => " + name + " missing"
         assert isinstance(file_results, ErrorFinder), message
 
-    test_results = ExTestResults(failing_arrays1d2d_test_dir, completed=True, timing=0.0)
+    test_results = ExTestResults(failing_arrays1d2d_test_dir, completed=True, err_msg="", timing=0.0)
     test_results.set_results(test_results_dict)
 
     assert os.path.basename(test_results.test_name) == "GW_ZrO2_failing_arrays", "Base test name"
@@ -275,7 +276,7 @@ def test_print_failing_array(failing_arrays1d2d_test_dir):
[…]
