### 2025-10-10 0c94764
Return inverse wings and head from angavsc0
