+++ src/src_xs/angavsc0.F90
@@ -5,434 +5,458 @@
 !BOP
 ! !ROUTINE: angavsc0
 ! !INTERFACE:
-subroutine angavsc0(n, nmax, scrnh, scrnw, scrn, scieff)
-! !USES:
-  use modmpi
-  use constants, only: pi, twopi, fourpi, zzero
-  use mod_qpoint, only: ngridq
-  use mod_lattice, only: omega, binv
-  use modinput, only: input
-  use modxs, only: dielten0, dielten, symt2, tleblaik,&
-                 & lmmaxdielt, sptclg
-  use invert
-! !INPUT/OUTPUT PARAMETERS:
-!   IN:
-!   n, integer               : Number of G+q vectors for current q-point
-!   nmax, integer            : Maximum number of G+q vectors over all q-points
-!   scrnh(3,3), complex(8)   : Head of dielectric matrix
-!   scrnw(n,2,3), complex(8) : Wings of dielectric matrix
-!   scrn(n,n), complex(8)    : Body of dielectric matrix
-!   OUT:
-!   scieff(nmax,nmax), complex(8) : Screened coulomb potential 
-!   
-! !DESCRIPTION:
-! This routine deals with the divergences of 
-! $\frac{\tilde{\varepsilon}_{G,G'}(q,\omega=0)}{|G+q||G'+q|}$
-! for $G=G'=q=0$ by averaging around the Gamma point.
-!
-! !REVISION HISTORY:
-! Added to documentation scheme. (Aurich)
-!
-!EOP
-!BOC
+subroutine angavsc0(n, nmax, scrnh, scrnw, scrn, scieff, eps_inv_head, eps_inv_wings)
+  ! !USES:
+    use modmpi
+    use constants, only: pi, twopi, fourpi, zzero
+    use mod_qpoint, only: ngridq
+    use mod_lattice, only: omega, binv
+    use modinput, only: input
+    use modxs, only: dielten0, dielten, symt2, tleblaik,&
+                   & lmmaxdielt, sptclg
+    use invert
+  ! !INPUT/OUTPUT PARAMETERS:
+  !   IN:
+  !   n, integer               : Number of G+q vectors for current q-point
+  !   nmax, integer            : Maximum number of G+q vectors over all q-points
+  !   scrnh(3,3), complex(8)   : Head of dielectric matrix
+  !   scrnw(n,2,3), complex(8) : Wings of dielectric matrix
+  !   scrn(n,n), complex(8)    : Body of dielectric matrix
+  !   OUT:
+  !   scieff(nmax,nmax), complex(8) : Screened coulomb potential 
+  !   
+  ! !DESCRIPTION:
+  ! This routine deals with the divergences of 
+  ! $\frac{\tilde{\varepsilon}_{G,G'}(q,\omega=0)}{|G+q||G'+q|}$
+  ! for $G=G'=q=0$ by averaging around the Gamma point.
+  !
+  ! !REVISION HISTORY:
+  ! Added to documentation scheme. (Aurich)
+  !
+  !EOP
+  !BOC
+  
+    implicit none
+  
+    ! Arguments
+    integer, intent(in) :: n, nmax
+    complex(8), intent(in) :: scrn(n, n), scrnw(n, 2, 3), scrnh(3, 3)
+    complex(8), intent(out) :: scieff(nmax, nmax)
+  
+    ! Local variables
+    integer, parameter :: nsphcov = 5810, iq0 = 1
+    integer :: iop, jop, j1, j2, itp, lm, ntpsph
+    real(8) :: vomega, t00, r, qsz
+    complex(8) :: dtns(3, 3), w1, w2,scol(3),zt
[…]
