+++ paper/paper.md
@@ -141,13 +141,13 @@ $$
 N_\mu = c_\mu \sqrt{\sqrt{N_{\text{pairs}}}}\:,
 $$
 where $N_{\text{pairs}}$ refers to the number of wave function pairs for which ISDF is computed. Note that $N_\text{pairs}$ depends on $N_\mathbf{k}$. The double square-root dependence ensures that the overall scaling remains below $\mathcal{O}(N_\mathbf{k}^2)$. In Fig.(\autoref{fig:newparams}) we present, for the example of diamond, the difference in exciton binding energies obtained with the new implementation and a reference calculation. The reference, based on the direct implementation, sets up and diagonalizes the full BSH and depends neither on $n_\mathbf{r}$ nor $c_\mu$. The results are shown as functions of $n_\mathbf{r}$ and $c_\mu$ for a small $\mathbf{k}$-grid of $2\times 2\times 2$. Additionally, we show the spectral similarities compared to the reference calculation as functions of $n_\mathbf{r}$ and $c_\mu$. For both parameters, both properties converge as their values increase. To find the optimal interpolation grid for ISDF, we first converge $n_\mathbf{r}$, then $c_\mu$. In our example, $n_\mathbf{r}=138$ [a.u.] and $c_\mu = 40.0$ yield converged results. This corresponds to a real-space sampling of $22 \times 22 \times 22$ and numbers of interpolation points $N_\mu^V = 202$, $N_\mu^{W_o} = 322$, and $N_\mu^{W_u} = 360$.
-  ![Difference of the exciton binding energy and spectral similarity between the new method and the direct method as functions of $n_\mathbf{r}$ (upper panel) and $c_\mu$ (lower panel) for diamond on a $2 \times 2 \times 2$ $\mathbf k$-grid. \label{fig:newparams}](Diamond_ngridrdens_cisdf.png)
+  ![Difference of the exciton binding energy and spectral similarity between the new method and the direct method as functions of $n_\mathbf{r}$ (upper panel) and $c_\mu$ (lower panel) for diamond on a $2 \times 2 \times 2$ $\mathbf k$-grid.](Diamond_ngridrdens_cisdf.png){#fig:newparams}
 
 In Fig.(\autoref{fig:error}) we compare the exciton binding energies and spectra of the new implementation to those of the old implementation for increasing $\mathbf k$-grids, while keeping $n_\mathbf{r}$ as well as $N_\mu^V$,  $N_\mu^{W_o}$, and $N_\mu^{W_u}$ fixed at the values above. We observe that the results converge as $N_\mathbf{k}$ increases. Thus, the number of interpolation points is asymptotically independent of $N_\mathbf{k}$. A similar behavior of ISDF was observed in [@Lu:2016].
-![Difference of the exciton binding energy and spectral similarity between the new method and the direct method as functions of $N_{\mathbf k}$ for diamond. \label{fig:error}](Diamond_ngridk_fastDirectComparison.png)
+![Difference of the exciton binding energy and spectral similarity between the new method and the direct method as functions of $N_{\mathbf k}$ for diamond.](Diamond_ngridk_fastDirectComparison.png){#fig:error}
 
 In Fig.(\autoref{fig:scaling}) we show the wall times for solving the BSE with the new and direct implementations for increasing $N_\mathbf{k}$. The new algorithm massively outperforms the direct one, and the speedup increases more than linearly with $N_\mathbf{k}$. We also show the wall times for computing the RPA screening with increasing $N_\mathbf{k}$, which is now clearly the bottleneck in solving the BSE.
[…]
