### 2025-10-21 83b9e63
fixed figure captions
