### 2025-10-24 1d025f6
RT: fix Hamiltonian initialization, rename polarization output, add polarization parser
