### 2025-10-28 #8866 @jjkas
@maurerben,
Thanks for your patience.

I've got the tutorial running now. Note that the input file for fastBSE has a typo:
<doonly task="fastBSE_human_readable_output">
should be
<doonly task="fastBSE_human_readable_output"/>

at least for me it complained that the xml file was not readable.

Now it has run, and the results look essentially identical to direct BSE, which is great, as the calculation was about 1/3 the time even for the 444 run. Very nice piece of work.

**My review is as follows:**

The present work implements a BSE solver using interpolative separable density fitting (ISDF). The new solver reproduces the results of direct solution of the BSE Hamiltonian for optical spectra, and speeds the calculations up substantially, especially for dense k-grids. This is an important development, especially for investigating low energy excitations in systems requiring a dense k-grid. I suggest publication of this work.

One comment on the process (partly for the authors, partly for the editors):

For reviewers, a single clear document showing how to compile and test the current implementation would be extremely helpful. This would include not just unit tests, but the type of test that a reviewer should perform in order to check functionality, reproducibility, and performance as defined by JOSS. This might be something that goes into the overall documentation of the software.
