+++ src/writers/putevalsv.f90
@@ -26,7 +26,7 @@ Subroutine putevalsv (ik, evalsvp)
          Open (70, File=outfilenamestring(filetag, ik), Action='WRITE', &
         & Form='UNFORMATTED', Access='DIRECT', Recl=Recl)
          If (splittfile) Then
-            koffset = ik - firstk (procofk(ik, nkpt), nkpt) + 1
+            koffset = ik - firstofset (procofindex(ik, nkpt), nkpt) + 1
          Else
             koffset = ik
          End If
