### 2025-11-05 00c198b
Small math update
