### 2025-11-10 #8866 @zhubonan
Hi @maurerben, I can see that all three reviewers are happy with the paper and their comments are address. Before I move this to the next stage, i.e. recommend accept, could you please complete the following:

1. Make a software release, please report the version that should be linked with paper here.
2. Archive the release (e.g. your code) with a repository service, e.g. Zenodo and let me know the DOI. Please make sure the title and authors match with that of the paper. Please also make sure that the license matches with that of the software. You can check existing JOSS paper to see how they link with an archive.
3. Please double check the author's ORCID

PS you can also refer to the *author's check list* below.

Thanks to @ruiyiQM @jjkas @EderGio for their comprehensive review of the paper.
