+++ external/libXC/src/libxc_funcs.f90
@@ -1,0 +1,266 @@
+module libxc_funcs_m
+  implicit none
+
+  public
+
+  integer, parameter :: XC_LDA_X             =   1  !  Exchange                     
+  integer, parameter :: XC_LDA_C_WIGNER      =   2  !  Wigner parametrization       
+  integer, parameter :: XC_LDA_C_RPA         =   3  !  Random Phase Approximation   
+  integer, parameter :: XC_LDA_C_HL          =   4  !  Hedin & Lundqvist            
+  integer, parameter :: XC_LDA_C_GL          =   5  !  Gunnarson & Lundqvist        
+  integer, parameter :: XC_LDA_C_XALPHA      =   6  !  Slater Xalpha                
+  integer, parameter :: XC_LDA_C_VWN         =   7  !  Vosko, Wilk, & Nussair (5)   
+  integer, parameter :: XC_LDA_C_VWN_RPA     =   8  !  Vosko, Wilk, & Nussair (RPA) 
+  integer, parameter :: XC_LDA_C_PZ          =   9  !  Perdew & Zunger              
+  integer, parameter :: XC_LDA_C_PZ_MOD      =  10  !  Perdew & Zunger (Modified)   
+  integer, parameter :: XC_LDA_C_OB_PZ       =  11  !  Ortiz & Ballone (PZ)         
+  integer, parameter :: XC_LDA_C_PW          =  12  !  Perdew & Wang                
+  integer, parameter :: XC_LDA_C_PW_MOD      =  13  !  Perdew & Wang (Modified)     
+  integer, parameter :: XC_LDA_C_OB_PW       =  14  !  Ortiz & Ballone (PW)         
+  integer, parameter :: XC_LDA_C_2D_AMGB     =  15  !  Attacalite et al             
+  integer, parameter :: XC_LDA_C_2D_PRM      =  16  !  Pittalis, Rasanen & Marques correlation in 2D 
+  integer, parameter :: XC_LDA_C_vBH         =  17  !  von Barth & Hedin            
+  integer, parameter :: XC_LDA_C_1D_CSC      =  18  !  Casula, Sorella, and Senatore 1D correlation     
+  integer, parameter :: XC_LDA_X_2D          =  19  !  Exchange in 2D 
+  integer, parameter :: XC_LDA_XC_TETER93    =  20  !  Teter 93 parametrization                
+  integer, parameter :: XC_LDA_X_1D          =  21  !  Exchange in 1D     
+  integer, parameter :: XC_LDA_C_ML1         =  22  !  Modified LSD (version 1) of Proynov and Salahub 
+  integer, parameter :: XC_LDA_C_ML2         =  23  !  Modified LSD (version 2) of Proynov and Salahub 
+  integer, parameter :: XC_LDA_C_GOMBAS      =  24  !  Gombas parametrization       
+  integer, parameter :: XC_LDA_C_PW_RPA      =  25  !  Perdew & Wang fit of the RPA 
+  integer, parameter :: XC_LDA_C_1D_LOOS     =  26  !  P-F Loos correlation LDA     
+  integer, parameter :: XC_LDA_C_RC04        =  27  !  Ragot-Cortona 
+  integer, parameter :: XC_LDA_C_VWN_1       =  28  !  Vosko, Wilk, & Nussair (1)   
+  integer, parameter :: XC_LDA_C_VWN_2       =  29  !  Vosko, Wilk, & Nussair (2)   
+  integer, parameter :: XC_LDA_C_VWN_3       =  30  !  Vosko, Wilk, & Nussair (3)   
+  integer, parameter :: XC_LDA_C_VWN_4       =  31  !  Vosko, Wilk, & Nussair (4)   
[…]
