+++ src/src_xs/m_dhesolver.f90
@@ -1,7 +1,7 @@
 module m_dhesolver
-  use modmpi 
+  use modmpi
+  use herm_eigensolver
   use modscl
-  use m_hesolver
   use mod_hdf5
   implicit none
 
@@ -14,23 +14,23 @@ module m_dhesolver
     ! !INPUT/OUTPUT PARAMETERS:
     ! IN:
     !   type(blacsinfo) :: binfo       ! Info type describing the BLACS grid
-    !   integer(4), optional :: i1, i2 ! Index range of eigen-solutions 
+    !   integer(4), optional :: i1, i2 ! Index range of eigen-solutions
     !   real(8), optional :: v1, v2    ! Range of eigenvalues to search for
     !   integer(4), optional :: eecs   ! Estimate for eigenvalue clustering.
     !                                  ! Apriori not known but needed for propper
     !                                  ! orthogonalization of eigenvectors (default = 3)
     ! IN/OUT:
-    !   type(dzmat) :: ham         ! 2D block cyclic distributed hermitian matrix 
-    !   real(8) :: eval(ham%nrows) ! Real valued eigenvalues in ascending order 
+    !   type(dzmat) :: ham         ! 2D block cyclic distributed hermitian matrix
+    !   real(8) :: eval(ham%nrows) ! Real valued eigenvalues in ascending order
     !   type(dzmat), optional :: evec   ! 2D block cyclic distributed eigenvector matrix
     ! OUT:
     !   integer(4), optional :: found ! How many solutions were found
     !
     ! !DESCRIPTION:
     !   Takes the upper triangular part of an distributed complex matrix matrix,
-    !   assumed to be hermitian and finds eigenvalues and eigenvectors using 
-    !   the scalapack routine {\tt pzheevx}.
-    ! 
+    !   assumed to be hermitian and finds eigenvalues and eigenvectors using
+    !   the appropriate solver (ELPA, ScaLAPACK, or LAPACK).
+    !
     ! !REVISION HISTORY:
     !   Created 2016 (Aurich)
     !EOP
@@ -42,403 +42,15 @@ module m_dhesolver
       type(dzmat), intent(inout) :: ham
       real(8), intent(inout) :: eval(:)
       type(blacsinfo), intent(in) :: binfo
-      ! Optional argumets
+      ! Optional arguments
       type(dzmat), intent(inout), optional :: evec
       integer(4), intent(in), optional :: eecs
       integer(4), intent(in), optional :: i1, i2
       real(8), intent(in), optional :: v1, v2
       integer(4), intent(out), optional :: found
 
-      logical :: evalsonly, distributed, sane
-      type(dzmat) :: evecdummy
-
-      ! Local variables
-#ifdef SCAL
-      character(1) :: rangechar, jobzchar
-      integer(4) :: info
-      integer(4) :: lwork, lrwork, liwork
-      integer(4) :: il, iu
-      real(8) :: abstol, vl, vu
-      complex(8), allocatable :: work(:)
-      real(8), allocatable :: rwork(:)
-      integer(4), allocatable :: iwork(:)
-      real(8) :: orfac
-      integer(4) :: ia, ja, iz, jz
-      integer(4) :: nevalfound, nevecfound
-      real(8), allocatable :: gap(:)
-      integer(4), allocatable :: iclustr(:), ifail(:)
-      integer(4) :: clusterguess
-
-      ! External functions
-      real(8), external :: pdlamch
-#endif
-
-      if(present(evec)) then
-        evalsonly = .false.
-        distributed = ham%isdistributed .and. evec%isdistributed
-        sane = (ham%isdistributed .eqv. evec%isdistributed)
-      else
-        evalsonly = .true.
[…]
