### 2025-11-14 7391153
Fix pipeline timeout with -O3 optimization
