### 2025-11-16 9b34a2d
Release new version of excitingtools
