### 2025-11-20 b3243a9
Fix rttddft tutorial
