### 2025-11-26 68c4b93
fix install instructions for fastBSE
