### 2025-11-28 01956c5
G0W0@DFT-1/2
