### 2025-12-01 33dd7ec
RT: scissor correction in length gauge
