+++ src/src_xs/src_rttddft/rttddft_sanity_checks.f90
@@ -1,13 +1,18 @@
 !> This module handles checks of the user-provided input for the RT module. The calculations are 
 !> terminated, if the input is inconsistent or incompatible with the current version of the code.
 module rttddft_sanity_checks
+  use asserts, only: assert
   use constants, only: real_zero
+  use exciting_mpi, only: xmpi_allgatherv
+  use general_find_vbm_cbm, only: find_vbm_cbm
+  use mod_mpi_env, only: mpiinfo
   use modinput, only: input_type
   use modmpi, only: terminate, terminate_if_false
   use physical_constants, only: c
   use precision, only: dp, i32, sp
   use rttddft_electric_field, only: Electric_Field
   use rttddft_VectorPotential, only: Vector_Potential
+  use rttddft_Wavefunction, only: wavefunction_set
   use to_char_conversion, only: to_char
   use vector_multiplication, only: norm
 
@@ -17,9 +22,14 @@ module rttddft_sanity_checks
   public :: check_rttddft_input, check_rttddft_setup
 
   integer(i32), parameter :: n_cartesian_directions = 3
-  real(dp), parameter :: e_field_squared_au_to_intensity_wcm2 = 3.50941e16_dp, &
-    intensity_extremely_high = 1.e18_dp, eps_kick_width = 1.e-14_dp, t_step_scale = 0.2_dp, &
-    eps_energy_gap = 1.e-5_dp, eps_e_field = 1.e-12_dp, &
+  real(dp), parameter :: &
+    e_field_squared_au_to_intensity_wcm2 = 3.50941e16_dp, &
+    intensity_extremely_high = 1.e18_dp, &
+    eps_kick_width = 1.e-14_dp, &
+    t_step_scale = 0.2_dp, &
+    eps_energy_gap = 1.e-5_dp, &
+    eps_e_field = 1.e-12_dp, &
+    eps_scissor = 1.e-10_dp, &
     e_field_extremely_high = sqrt( intensity_extremely_high / e_field_squared_au_to_intensity_wcm2 )
   
 contains
@@ -65,11 +75,11 @@ contains
       call terminate_if_false( trim( inp%xs%realTimeTDDFT%laser%fieldType ) == "total", &
         "Berry-phase coupling is currently available only with total field given" )
     else
+      call terminate_if_false( inp%xs%scissor < eps_scissor, "Scissor correction is not implemented for velocity gauge field coupling" )
       call terminate_if_false( associated( inp%xs%realTimeTDDFT%pmat ), &
       & 'Element <pmat> in <realTimeTDDFT> not found' )
     end if
 
-    ! No restart currently possible for MD calculations
     if( associated( inp%MD ) ) then
       if( trim( inp%xs%realTimeTDDFT%do ) /= "fromscratch" ) then 
         call terminate_if_false( trim( inp%xs%realTimeTDDFT%propagator ) == "SE" .or. trim( inp%xs%realTimeTDDFT%propagator ) == "EH", &
@@ -77,11 +87,12 @@ contains
         call terminate_if_false( inp%xs%realTimeTDDFT%timeStep == inp%MD%timeStep, &
           "Restart for Ehrenfest MD is currently only implemented when the RT-TDDFT and MD timesteps are the same")
       end if
-
       call terminate_if_false( inp%xs%realTimeTDDFT%numberOfFrozenStates == 0, &
         "No state freezing currently possible for MD calculations" )
       call terminate_if_false( trim( inp%xs%realTimeTDDFT%basis ) == "LAPWlo", &
         "Usage of the KS basis set is currently unavailable for MD calculations" )
[…]
