### 2025-12-02 #8866 @maurerben
Dear @zhubonan,

Please excuse the long delay. I had to discuss the release first with the team. We would like to connect the paper with the recent sodium.alpha release. The fastBSE code has not changed from the reviewed version, I hope this works.

The DOI I created with zenedo for it is https://doi.org/10.5281/zenodo.17791882.

I doubled checked affiliations and ORCIDS, both are correct. Claudia Draxl's ORCID is not shown in the compiled paper, but present in the paper.md.

Thank you very much for guiding me through the review process.

With best Regards
