### 2025-12-03 95c7b1f
Remove n_assertions from unit test initialization
