+++ src/src_hybrids/hse_singularity_test.f90
@@ -22,11 +22,9 @@ contains
     logical, optional :: kill_on_failure
     !> Test report object
     type(unit_test_type) :: test_report
-    !> Number of assertions
-    integer, parameter :: n_assertions = 2
 
     ! Initialize test object
-    call test_report%init(n_assertions, mpiglobal)
+    call test_report%init( mpiglobal)
 
     ! Run and assert tests
 
