+++ src/lapack_wrappers/decomposition/QR_factorization_test.f90
@@ -29,10 +29,8 @@ module qr_factorization_test
 
     !> Test report object
     type(unit_test_type) :: test_report
-    !> Number of assertions
-    integer, parameter :: n_assertions = 12
 
-    call test_report%init(n_assertions, mpiglobal)
+    call test_report%init( mpiglobal)
 
     call test_qr_column_pivot_real(test_report)
    
