+++ src/lapack_wrappers/multiplication/vector_multiplication_test.f90
@@ -28,10 +28,8 @@ contains
 
     !> Test report object
     type(unit_test_type) :: test_report
-    !> Number of assertions
-    integer, parameter :: n_assertions = 14
 
-    call test_report%init(n_assertions, mpiglobal)
+    call test_report%init( mpiglobal)
 
     ! Run unit tests
 
