+++ src/math/grid_utils_test.f90
@@ -26,10 +26,8 @@ contains
 
       !> Test report object
       type(unit_test_type) :: test_report
-      !> Number of assertions
-      integer, parameter :: n_assertions = 51
 
-      call test_report%init(n_assertions, mpiglobal)
+      call test_report%init( mpiglobal)
 
       ! Run unit tests
       call test_mesh_1d(test_report)
