### 2025-12-16 #8866 @maurerben
Hi @zhubonan I changed the title of the publication to the paper title and the put the correct license (GPL). If I understand correctly, the doi is only for linking the paper to the specific version of the code and meant as reference to the fast BSE implementation. Concequently I removed all other names from the authorlist of this specific release. We will create a new DOI for the next minor release of exciting that will be published soon. I hope this works. The DOI is still https://doi.org/10.5281/zenodo.17791882. Thank you!
