### 2026-01-08 #8866 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
