### 2026-01-12 #8866 @rkurchin
👋 Hi, congrats on being nearly done (and apologies for the slight delay here)! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [x] please add contributing guidelines for the project
- [x] please issue a tagged GitHub release corresponding to the version listed here and in the Zenodo archive
- [x] the Zenodo archive lists two different licenses (GPL 2.0 and 2.1), but the repo has only 2.0 (plus the note about dependencies being under different licenses), please make sure this is consistent
- [x] check that `{Capitalization}` is being enforced where necessary in citations (e.g. in Dong et al. for Voronoi, Lu & Ying 2016 for Bloch, and Vorwerk et al., for Salpeter)
- [x] the paper itself is rather long (we typically shoot for 1000 words or less); looking at the present content, there seems to be a fair bit of theoretical background that could be left to a citation and/or to the package documentation. Please see about cutting this down a bit; once you've done so, I can do my editorial pass over the text
