### 2026-01-14 47ad9f5
Simplify interaction kernel equations and ISDF usage
Removed detailed equations for interaction kernels and their application to vectors, focusing on the efficiency of ISDF representation and Lanczos algorithm for solving the BSE.
