+++ paper/paper.md
@@ -62,71 +62,7 @@ u_{i\mathbf{k}}^*(\mathbf{r})u_{j\mathbf{k}'}(\mathbf{r}) \approx \sum_{\mu=1}^{
 $$
 where $N_\mu$ is the number of interpolation points, and $\zeta_\mu(\mathbf r)$ are the expansion coefficients. Due to the tensor product structure of $u_{i\mathbf{k}}^*(\mathbf{r})u_{j\mathbf{k}'}(\mathbf{r})$, $\zeta_\mu(\mathbf r)$ can be computed efficiently, and the scaling is bounded by $\mathcal{O}(N_\mu^3)$ [@Hu:2017]. We observe that we can always choose $N_\mu \ll N_o N_u N_\mathbf{k}$, thus ISDF is never a bottleneck. The interpolation points are computed efficiently with centroidal Voronoi tessellation within $\mathcal{O}(N_\mu N_r)$ [@Dong:2018].
 
-Inserting ISDF in the equations for the interaction kernels yields
-$$
-V_{ou\mathbf{k}, o'u'\mathbf{k}'} \approx \frac{1}{N_k^2} \sum_{\mu=1}^{N_\mu^V} \sum_{\nu=1}^{N_\mu^V}
-u_{o\mathbf{k}}(\mathbf{r}_\mu^V) 
-u^*_{u\mathbf{k}} (\mathbf{r}_\mu^V)
-\tilde V_{\mu \nu}
-u^*_{o'\mathbf{k}'} (\mathbf{r}_\nu^V) 
-u_{u'\mathbf{k}'} (\mathbf{r}_\nu^V),
-$$
-$$
-W_{ou\mathbf{k}, o'u'\mathbf{k}'} \approx \frac{1}{N_k^2} \sum_{\mu=1}^{N_{\mu}^{W_u}} \sum_{\nu=1}^{N_{\mu}^{W_o}}
-u^*_{u\mathbf{k}} (\mathbf{r}_\mu^{W_u}) 
-u_{u'\mathbf{k}'} (\mathbf{r}_\mu^{W_u})
-\tilde W_{\mu\nu, \mathbf{k-k'}}
-u_{o\mathbf{k}}(\mathbf{r}_\nu^{W_o})
-u^*_{o'\mathbf{k}'} (\mathbf{r}_\nu^{W_o}) 
-\:,
-$$
-where we have shifted the integration from the wavefunction pairs to the interpolation coefficients such that
-$$
-\tilde V_{\mu \nu} = \int_{\Omega^l\times \Omega^l}drdr' \zeta_\mu^{*V}(\mathbf{r}) V(\mathbf{r},\mathbf{r'}) \zeta_\nu^{V}(\mathbf{r'}) \:,
-$$
-$$
-\tilde W_{\mu\nu, \mathbf{k-k'}} = \int_{\Omega^l\times \Omega^l}drdr' \zeta_\mu^{*W_u}(\mathbf{r}) W_{\mathbf{k}-\mathbf{k'}}(\mathbf{r},\mathbf{r'}) \zeta_\nu^{W_o}(\mathbf{r'}) \:.
-$$
[…]
