### 2026-01-14 7c84988
Enforce capitalization in paper.bib
