+++ paper/paper.bib
@@ -16,7 +16,7 @@
   author = {Dong, Kun and Hu, Wei and Lin, Lin},
   month = {01},
   pages = {1311-1320},
-  title = {Interpolative Separable Density Fitting through Centroidal Voronoi Tessellation with Applications to Hybrid Functional Electronic Structure Calculations},
+  title = {Interpolative Separable Density Fitting through Centroidal {Voronoi} Tessellation with Applications to Hybrid Functional Electronic Structure Calculations},
   doi = {10.1021/acs.jctc.7b01113},
   urldate = {2022-04-24},
   volume = {14},
@@ -53,7 +53,7 @@
   month = {08},
   pages = {037001-037001},
   publisher = {IOP Publishing},
-  title = {Bethe–Salpeter equation for absorption and scattering spectroscopy: implementation in the exciting code},
+  title = {Bethe–{Salpeter} equation for absorption and scattering spectroscopy: implementation in the exciting code},
   doi = {10.1088/2516-1075/ab3123},
   volume = {1},
   year = {2019},
@@ -73,7 +73,7 @@
 }
 
 @article{Lu:2016,
-	title = {Fast algorithm for periodic density fitting for Bloch waves},
+	title = {Fast algorithm for periodic density fitting for {Bloch} waves},
 	volume = {1},
 	issn = {2380-2898},
 	url = {https://link.intlpress.com/JDetail/1805564664274690050},
