### 2026-01-16 9d37255
neon -> sodium in README and INSTALL add CONTRIBUTING
