### 2026-01-28 #8866 @maurerben
Hi,

Please excuse my late reply, I was sure I went through these points already. Maybe I forgot to hit Comment. Anyway, here is my reply to the points above:

To be sure I created a new, specific release, included the CONTRIBUTIONS file, corrected the Liscense and made sure that it only features the names on the paper.

This is the new DOI: https://zenodo.org/records/18249116
This is the new release: https://github.com/exciting/exciting/tree/sodium.alpha.joss-fast-bse

I corrected the capitalization in the references and cut most of the math.
