### 2026-01-29 #8866 @rkurchin
Copyediting...
- [x] line 16: "framework" should be one word
- [x] 19: hyphenate "Schrödinger-like"
- [x] 20: add space before citation
- [x] 30: remove comma after "see"
- [x] 35: either citation should be parenthetical, or incorporate it into the sentence, as in "... (ISDF), originally introduced by Lu & Ying (2016)."
- [x] 46: remove period after big-O expression and before citation
- [x] 49: Brillouin is misspelled
- [x] 65: remove "Fig.()" so that rendered text just reads "In Figure 1 we present..."
- [x] Figure 1: I presume $S$ is the spectral similarity; please actually define this somewhere (text and/or figure caption)
- [x] 75: similar comment to the one above for Figure 1, but also add comma after "Figure 2"
- [x] 80: same as for Figure 2 reference
- [x] for all figures with double y-axes, I suggest rotating the right-hand axis labels by 180 degrees; they're easier for a reader to parse that way
