### 2026-01-31 fb26dfb
Fix items listed in the copyedit
The new implementation of the BSE solver integrates ISDF for improved efficiency, significantly reducing computational time while maintaining accuracy. This allows for the study of complex materials and systems that were previously infeasible due to scaling issues.
