### 2026-02-04 #8866 @rkurchin
🎉 Congratulations to @maurerben and team, and thanks to editor @zhubonan and reviewers @jjkas, @ruiyiQM, and @EderGio for your work on this – JOSS couldn't run without your efforts! 🦾
